Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...

Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json

A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")
#> - "advanced search" ----------------------------- 6 packages in 0.001 seconds -
#>   #     package     version   by         @ title                               
#>  1  100 bibtex      0.4.2.3   ORPHANED  1y Bibtex Parser                       
#>  2   96 penalized   0.9.51    ORPHANED  4y L1 (Lasso and Fused Lasso) and L2...
#>  3   93 ggmap       3.0.0     ORPHANED  3y Spatial Visualization with ggplot2  
#>  4   93 cairoDevice 2.28.2.1  ORPHANED  4M Embeddable Cairo Graphics Device ...
#>  5   90 distances   0.1.8     ORPHANED  2y Tools for Distance Metrics          
#>  6   90 RGtk2       2.20.36.2 ORPHANED  4M R Bindings for Gtk 2.8.0 and Above  

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")
#> - "advanced search" ---------------------------- 18 packages in 0.001 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi       2M Find Tools Needed to Build...
#>   2  82 memoise     2.0.1   Winston Chang      3M 'Memoisation' of Functions   
#>   3  78 xml2        1.3.3   Hadley Wickham     3M Parse XML                    
#>   4  74 sessioninfo 1.2.2   Gábor Csárdi       2M R Session Information        
#>   5  74 devtools    2.4.3   Jennifer Bryan     3M Tools to Make Developing R...
#>   6  72 odbc        1.3.3   Hadley Wickham     3M Connect to ODBC Compatible...
#>   7  69 remotes     2.4.2   Gábor Csárdi       3M R Package Installation fro...
#>   8  67 meltr       1.0.0   Duncan Garmonsway  6M Read Non-Rectangular Text ...
#>   9  67 pkgload     1.2.4   Lionel Henry       3M Simulate Package Installat...
#>  10  66 withr       2.4.3   Lionel Henry       3M Run Code 'With' Temporaril...
advanced_search("Author: Hester AND Author: Wickham")
#> - "advanced search" ---------------------------- 18 packages in 0.001 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi       2M Find Tools Needed to Build...
#>   2  82 memoise     2.0.1   Winston Chang      3M 'Memoisation' of Functions   
#>   3  78 xml2        1.3.3   Hadley Wickham     3M Parse XML                    
#>   4  74 sessioninfo 1.2.2   Gábor Csárdi       2M R Session Information        
#>   5  74 devtools    2.4.3   Jennifer Bryan     3M Tools to Make Developing R...
#>   6  72 odbc        1.3.3   Hadley Wickham     3M Connect to ODBC Compatible...
#>   7  69 remotes     2.4.2   Gábor Csárdi       3M R Package Installation fro...
#>   8  67 meltr       1.0.0   Duncan Garmonsway  6M Read Non-Rectangular Text ...
#>   9  67 pkgload     1.2.4   Lionel Henry       3M Simulate Package Installat...
#>  10  66 withr       2.4.3   Lionel Henry       3M Run Code 'With' Temporaril...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")
#> - "advanced search" ---------------------------- 28 packages in 0.002 seconds -
#>   #     package        version by                @ title                       
#>   1 100 primerTree     1.0.5   Jim Hester       2y Visually Assessing the Sp...
#>   2  96 knitrBootstrap 1.0.2   Jim Hester       4y 'knitr' Bootstrap Framework 
#>   3  89 types          1.0.0   Jim Hester       5y Type Annotations            
#>   4  77 pak            0.2.1   Gábor Csárdi     2M Another Approach to Packa...
#>   5  75 dplyr.teradata 0.4.1   Koji Makiyama    1y A 'Teradata' Backend for ...
#>   6  74 gmailr         1.0.1   Jennifer Bryan   3M Access the 'Gmail' 'RESTf...
#>   7  70 decor          1.0.1   Romain François  3M Retrieve Code Decorations   
#>   8  70 lintr          2.0.1   Jim Hester       2y A 'Linter' for R Code       
#>   9  69 mockery        0.4.2   Jim Hester       2y Mocking Library for R       
#>  10  67 rex            1.2.1   Kevin Ushey      3M Friendly Regular Expressions

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")
#> - "advanced search" ---------------------------- 18 packages in 0.004 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi       2M Find Tools Needed to Build...
#>   2  96 odbc        1.3.3   Hadley Wickham     3M Connect to ODBC Compatible...
#>   3  95 xml2        1.3.3   Hadley Wickham     3M Parse XML                    
#>   4  82 memoise     2.0.1   Winston Chang      3M 'Memoisation' of Functions   
#>   5  75 devtools    2.4.3   Jennifer Bryan     3M Tools to Make Developing R...
#>   6  74 sessioninfo 1.2.2   Gábor Csárdi       2M R Session Information        
#>   7  74 pkgload     1.2.4   Lionel Henry       3M Simulate Package Installat...
#>   8  73 remotes     2.4.2   Gábor Csárdi       3M R Package Installation fro...
#>   9  70 fs          1.5.2   Gábor Csárdi       2M Cross-Platform File System...
#>  10  68 meltr       1.0.0   Duncan Garmonsway  6M Read Non-Rectangular Text ...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")
#> - "advanced search" ---------------------------- 28 packages in 0.002 seconds -
#>   #     package        version by                @ title                       
#>   1 100 primerTree     1.0.5   Jim Hester       2y Visually Assessing the Sp...
#>   2  96 knitrBootstrap 1.0.2   Jim Hester       4y 'knitr' Bootstrap Framework 
#>   3  89 types          1.0.0   Jim Hester       5y Type Annotations            
#>   4  77 pak            0.2.1   Gábor Csárdi     2M Another Approach to Packa...
#>   5  75 dplyr.teradata 0.4.1   Koji Makiyama    1y A 'Teradata' Backend for ...
#>   6  74 gmailr         1.0.1   Jennifer Bryan   3M Access the 'Gmail' 'RESTf...
#>   7  70 decor          1.0.1   Romain François  3M Retrieve Code Decorations   
#>   8  70 lintr          2.0.1   Jim Hester       2y A 'Linter' for R Code       
#>   9  69 mockery        0.4.2   Jim Hester       2y Mocking Library for R       
#>  10  67 rex            1.2.1   Kevin Ushey      3M Friendly Regular Expressions

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")
#> - "advanced search" --------------------------- 372 packages in 0.014 seconds -
#>   #     package           version by                         @ title           
#>   1 100 phd               0.2     Jesse Hemerik             1M Permutation T...
#>   2  95 confSAM           0.2     Jesse Hemerik             4y Estimates and...
#>   3  84 permutes          2.3.2   Cesko C. Voeten           2M Permutation T...
#>   4  80 BurStMisc         1.1     Pat Burns                 6y Burns Statist...
#>   5  76 fdrci             2.2     Joshua Millstein          8M Permutation-B...
#>   6  75 MultSurvTests     0.2     Lukas Arnroth             8M Permutation T...
#>   7  75 resample          0.4     Tim Hesterberg            7y Resampling Fu...
#>   8  74 packager          1.14.0  Andreas Dominik Cullmann  2M Create, Build...
#>   9  73 USP               0.1.2   Thomas B. Berrett         1y U-Statistic P...
#>  10  73 phyloseqGraphTest 0.1.0   Julia Fukuyama            2y Graph-Based P...

# Packages that have a certain field
advanced_search("_exists_" = "URL")
#> - "advanced search" -------------------------- 9892 packages in 0.001 seconds -
#>   #     package                          version
#>   1 100 ndvtest                          1.0.0  
#>   2 100 factset.protobuf.stachextensions 1.0.1  
#>   3 100 choroplethrAdmin1                1.1.1  
#>   4 100 VSURF                            1.1.0  
#>   5 100 thinkr                           0.15   
#>   6 100 sparsepca                        0.1.2  
#>   7 100 fRegression                      3042.82
#>   8 100 anacor                           1.1.3  
#>   9 100 digitTests                       0.1.1  
#>  10 100 mcmcderive                       0.1.2  
#>  by                                                 @ title                    
#>  Yves Croissant                                    2d Shi's Non Degenerate V...
#>  Sai Giridhar Pasumarthi<spasumarthi@factset.com>  2d 'FactSet' 'STACH' Exte...
#>  Ari Lamstein                                      5y Contains an Administra...
#>  Robin Genuer                                      3y Variable Selection Usi...
#>  Vincent Guyader                                   2y Tools for Cleaning Up ...
#>  N. Benjamin Erichson                              4y Sparse Principal Compo...
#>  Tobias Setz                                       4y Rmetrics - Regression ...
#>  Patrick Mair                                      5y Simple and Canonical C...
#>  Koen Derks                                        3M Tests for Detecting Ir...
#>  Joe Thorley                                       6M Derive MCMC Parameters   

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")
#> - "advanced search" -------------------------- 9172 packages in 0.001 seconds -
#>   #     package        version by                         @ title              
#>   1 100 PKPDsim        1.1.0   Ron Keizer               28d Tools for Perfor...
#>   2 100 coreheat       0.2.2   Vidal Fey                 5M Correlation Heat...
#>   3 100 chillR         0.72.7  Eike Luedeling            1M Statistical Meth...
#>   4 100 trendchange    1.2     Sandeep Kumar Patakamuri  1M Innovative Trend...
#>   5 100 compas         0.1.1   Samuel W.K. Wong          1M Conformational M...
#>   6 100 ora            2.0.1   Arni Magnusson            8y Convenient Tools...
#>   7 100 hySpc.testthat 0.2.1   Claudia Beleites          2y 'testthat' Unit ...
#>   8 100 vandalico      0.0.1   Alberto Jiménez-Valverde  1d Evaluation of Pr...
#>   9 100 maps           3.4.0   Alex Deckmyn              5M Draw Geographica...
#>  10 100 futile.logger  1.4.3   Brian Lee Yung Rowe       6y A Logging Utilit...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)
#> - "advanced search" ------------------------------ 9172 packages in 0 seconds -
#>   #     package        version by                         @ title              
#>   1 100 PKPDsim        1.1.0   Ron Keizer               28d Tools for Perfor...
#>   2 100 coreheat       0.2.2   Vidal Fey                 5M Correlation Heat...
#>   3 100 chillR         0.72.7  Eike Luedeling            1M Statistical Meth...
#>   4 100 trendchange    1.2     Sandeep Kumar Patakamuri  1M Innovative Trend...
#>   5 100 compas         0.1.1   Samuel W.K. Wong          1M Conformational M...
#>   6 100 ora            2.0.1   Arni Magnusson            8y Convenient Tools...
#>   7 100 hySpc.testthat 0.2.1   Claudia Beleites          2y 'testthat' Unit ...
#>   8 100 vandalico      0.0.1   Alberto Jiménez-Valverde  1d Evaluation of Pr...
#>   9 100 maps           3.4.0   Alex Deckmyn              5M Draw Geographica...
#>  10 100 futile.logger  1.4.3   Brian Lee Yung Rowe       6y A Logging Utilit...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")
#> - "advanced search" --------------------------- 235 packages in 0.001 seconds -
#>   #     package         version by                         @ title             
#>   1 100 xtable          1.8.4   David Scott               3y Export Tables t...
#>   2 100 knitr           1.37    Yihui Xie                 2M A General-Purpo...
#>   3 100 raster          3.5.15  Robert J. Hijmans        25d Geographic Data...
#>   4 100 gdalUtils       2.0.3.2 Jonathan Asher Greenberg  2y Wrappers for th...
#>   5 100 fivethirtyeight 0.6.2   Albert Y. Kim             4M Data and Code B...
#>   6 100 tableone        0.13.0  Kazuki Yoshida            8M Create 'Table 1...
#>   7 100 bridgesampling  1.1.2   Quentin F. Gronau        10M Bridge Sampling...
#>   8 100 rmcorr          0.4.5   Laura R. Marusich         3M Repeated Measur...
#>   9 100 equatiomatic    0.3.1   Daniel Anderson          17d Transform Model...
#>  10 100 tscount         1.4.3   Tobias Liboschik          1y Analysis of Cou...

# Fuzzy search
advanced_search(Author = "Johnathan~1")
#> - "advanced search" --------------------------- 236 packages in 0.002 seconds -
#>   #     package    version by                                        @ title   
#>   1 100 mlmi       1.1.1   Jonathan Bartlett                        3M Maxim...
#>   2 100 UScancer   0.1.2   Jonathan Lee                             8y Creat...
#>   3 100 lda        1.4.2   Jonathan Chang                           6y Colla...
#>   4  95 geomapdata 1.0.4   Jonathan M. Lees<jonathan.lees@unc.edu>  9y Data ...
#>   5  95 handyplots 1.1.3   Jonathan Schwartz                        3y Handy...
#>   6  95 UnitCircle 0.1.3   Jonathan Berrisch                        4y Check...
#>   7  94 RSeed      0.1.60  Claus Jonathan Fritzemeier               5y Boren...
#>   8  94 bootImpute 1.2.0   Jonathan Bartlett                       11M Boots...
#>   9  94 chords     0.95.4  Jonathan Rosenblatt                      5y Estim...
#>  10  93 rSHAPE     0.3.2   Jonathan Dench                           3y Simul...