New CRAN packages

List the latest new CRAN packages.

cran_new(from = "last-week", to = "now", last = Inf)

Arguments

from

Start of the time interval to query. Possible values:

"last-week"

"last-month"

A Date object to be used as a start date.

A POSIXt object to be used as the start date.

A difftime object to used as the time interval until now.

An integer scalar, the number of days until today.

A character string that is converted to a start date using as.POSIXct().

to

End of the time interval to query. It accepts the same kinds of values as from, and additionally it can also be the string "now", to specify the current date and time.

last

Integer to limit the number of returned packages.

Value

Data frame of package descriptions.

Examples

# Last week
cran_new("last-week")
#> # A data frame: 21 × 31
#>    Date/…¹ Package Title Version Autho…² Descr…³ License Encod…⁴ Roxyg…⁵ Needs…⁶
#>    <chr>   <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 2022-1… IGCiti… "Sim… 0.2.0   "c(\np… "Devel… MIT + … UTF-8   7.2.1   no     
#>  2 2022-1… volati… "Hig… 1.0.1   "c(per… "Tradi… GPL-3   UTF-8   7.2.1   no     
#>  3 2022-1… worrrd  "Gen… 0.1.0   "c(\np… "Gener… MIT + … UTF-8   7.1.2   no     
#>  4 2022-1… people… "Dat… 0.1.0   "\nper… "Data … MIT + … UTF-8   7.2.1   no     
#>  5 2022-1… picR    "Pre… 1.0.0   "\nper… "Compu… GPL (>… UTF-8   7.2.0   no     
#>  6 2022-1… Homomo… "BFV… 0.1.0   "\nper… "Imple… GPL (>… UTF-8   7.2.1   no     
#>  7 2022-1… bakeoff "Dat… 0.2.0   "\nc(p… "Data … MIT + … UTF-8   7.2.1.… no     
#>  8 2022-1… crossu… "Cro… 1.0.6   "c(per… "Doubl… MIT + … UTF-8   7.2.1   no     
#>  9 2022-1… Shaple… "Mul… 0.1.0   "\nc(\… "Based… GPL-3   UTF-8   7.2.1   no     
#> 10 2022-1… NestMR… "Sin… 1.0     "\nper… "This … CC0     UTF-8   7.2.1   yes    
#> # … with 11 more rows, 21 more variables: Packaged <chr>, Author <chr>,
#> #   Maintainer <chr>, Repository <chr>, crandb_file_date <chr>, MD5sum <chr>,
#> #   date <chr>, dependencies <list>, Type <chr>, URL <chr>, BugReports <chr>,
#> #   VignetteBuilder <chr>, LazyData <chr>, `Config/testthat/edition` <chr>,
#> #   ByteCompile <chr>, RdMacros <chr>, Date <chr>, Language <chr>,
#> #   `Config/testthat/parallel` <chr>, SystemRequirements <chr>,
#> #   `Config/Needs/website` <chr>, and abbreviated variable names …

# Last month
cran_new("last-month")
#> # A data frame: 179 × 41
#>    Date/…¹ Package Title Version Autho…² Descr…³ License Encod…⁴ Roxyg…⁵ Needs…⁶
#>    <chr>   <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 2022-1… IGCiti… "Sim… 0.2.0   "c(\np… "Devel… MIT + … UTF-8   7.2.1   no     
#>  2 2022-1… volati… "Hig… 1.0.1   "c(per… "Tradi… GPL-3   UTF-8   7.2.1   no     
#>  3 2022-1… worrrd  "Gen… 0.1.0   "c(\np… "Gener… MIT + … UTF-8   7.1.2   no     
#>  4 2022-1… people… "Dat… 0.1.0   "\nper… "Data … MIT + … UTF-8   7.2.1   no     
#>  5 2022-1… picR    "Pre… 1.0.0   "\nper… "Compu… GPL (>… UTF-8   7.2.0   no     
#>  6 2022-1… Homomo… "BFV… 0.1.0   "\nper… "Imple… GPL (>… UTF-8   7.2.1   no     
#>  7 2022-1… bakeoff "Dat… 0.2.0   "\nc(p… "Data … MIT + … UTF-8   7.2.1.… no     
#>  8 2022-1… crossu… "Cro… 1.0.6   "c(per… "Doubl… MIT + … UTF-8   7.2.1   no     
#>  9 2022-1… Shaple… "Mul… 0.1.0   "\nc(\… "Based… GPL-3   UTF-8   7.2.1   no     
#> 10 2022-1… NestMR… "Sin… 1.0     "\nper… "This … CC0     UTF-8   7.2.1   yes    
#> # … with 169 more rows, 31 more variables: Packaged <chr>, Author <chr>,
#> #   Maintainer <chr>, Repository <chr>, crandb_file_date <chr>, MD5sum <chr>,
#> #   date <chr>, dependencies <list>, Type <chr>, URL <chr>, BugReports <chr>,
#> #   VignetteBuilder <chr>, LazyData <chr>, `Config/testthat/edition` <chr>,
#> #   ByteCompile <chr>, RdMacros <chr>, Date <chr>, Language <chr>,
#> #   `Config/testthat/parallel` <chr>, SystemRequirements <chr>,
#> #   `Config/Needs/website` <chr>, biocViews <chr>, Collate <chr>, …

# Last 5 days
cran_new(from = 5)
#> # A data frame: 10 × 26
#>    Date/…¹ Package Title Version Autho…² Descr…³ License Encod…⁴ Roxyg…⁵ Needs…⁶
#>    <chr>   <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 2022-1… IGCiti… "Sim… 0.2.0   "c(\np… "Devel… MIT + … UTF-8   7.2.1   no     
#>  2 2022-1… volati… "Hig… 1.0.1   "c(per… "Tradi… GPL-3   UTF-8   7.2.1   no     
#>  3 2022-1… worrrd  "Gen… 0.1.0   "c(\np… "Gener… MIT + … UTF-8   7.1.2   no     
#>  4 2022-1… people… "Dat… 0.1.0   "\nper… "Data … MIT + … UTF-8   7.2.1   no     
#>  5 2022-1… picR    "Pre… 1.0.0   "\nper… "Compu… GPL (>… UTF-8   7.2.0   no     
#>  6 2022-1… Homomo… "BFV… 0.1.0   "\nper… "Imple… GPL (>… UTF-8   7.2.1   no     
#>  7 2022-1… bakeoff "Dat… 0.2.0   "\nc(p… "Data … MIT + … UTF-8   7.2.1.… no     
#>  8 2022-1… crossu… "Cro… 1.0.6   "c(per… "Doubl… MIT + … UTF-8   7.2.1   no     
#>  9 2022-1… Shaple… "Mul… 0.1.0   "\nc(\… "Based… GPL-3   UTF-8   7.2.1   no     
#> 10 2022-1… NestMR… "Sin… 1.0     "\nper… "This … CC0     UTF-8   7.2.1   yes    
#> # … with 16 more variables: Packaged <chr>, Author <chr>, Maintainer <chr>,
#> #   Repository <chr>, crandb_file_date <chr>, MD5sum <chr>, date <chr>,
#> #   dependencies <list>, Type <chr>, URL <chr>, BugReports <chr>,
#> #   VignetteBuilder <chr>, LazyData <chr>, `Config/testthat/edition` <chr>,
#> #   ByteCompile <chr>, RdMacros <chr>, and abbreviated variable names
#> #   ¹​`Date/Publication`, ²​`Authors@R`, ³​Description, ⁴​Encoding, ⁵​RoxygenNote,
#> #   ⁶​NeedsCompilation

# From a given date, but at most 10
cran_new(from = "2021-04-06", last = 10)
#> # A data frame: 10 × 26
#>    Date/…¹ Package Title Version Autho…² Descr…³ License Encod…⁴ Roxyg…⁵ Needs…⁶
#>    <chr>   <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 2022-1… IGCiti… "Sim… 0.2.0   "c(\np… "Devel… MIT + … UTF-8   7.2.1   no     
#>  2 2022-1… volati… "Hig… 1.0.1   "c(per… "Tradi… GPL-3   UTF-8   7.2.1   no     
#>  3 2022-1… worrrd  "Gen… 0.1.0   "c(\np… "Gener… MIT + … UTF-8   7.1.2   no     
#>  4 2022-1… people… "Dat… 0.1.0   "\nper… "Data … MIT + … UTF-8   7.2.1   no     
#>  5 2022-1… picR    "Pre… 1.0.0   "\nper… "Compu… GPL (>… UTF-8   7.2.0   no     
#>  6 2022-1… Homomo… "BFV… 0.1.0   "\nper… "Imple… GPL (>… UTF-8   7.2.1   no     
#>  7 2022-1… bakeoff "Dat… 0.2.0   "\nc(p… "Data … MIT + … UTF-8   7.2.1.… no     
#>  8 2022-1… crossu… "Cro… 1.0.6   "c(per… "Doubl… MIT + … UTF-8   7.2.1   no     
#>  9 2022-1… Shaple… "Mul… 0.1.0   "\nc(\… "Based… GPL-3   UTF-8   7.2.1   no     
#> 10 2022-1… NestMR… "Sin… 1.0     "\nper… "This … CC0     UTF-8   7.2.1   yes    
#> # … with 16 more variables: Packaged <chr>, Author <chr>, Maintainer <chr>,
#> #   Repository <chr>, crandb_file_date <chr>, MD5sum <chr>, date <chr>,
#> #   dependencies <list>, Type <chr>, URL <chr>, BugReports <chr>,
#> #   VignetteBuilder <chr>, LazyData <chr>, `Config/testthat/edition` <chr>,
#> #   ByteCompile <chr>, RdMacros <chr>, and abbreviated variable names
#> #   ¹​`Date/Publication`, ²​`Authors@R`, ³​Description, ⁴​Encoding, ⁵​RoxygenNote,
#> #   ⁶​NeedsCompilation

# March of 2021
cran_new(from = "2021-03-01", to = "2021-04-01")
#> # A data frame: 231 × 48
#>    Date/Pu…¹ Package Type  Title Version Autho…² Maint…³ Descr…⁴ License Encod…⁵
#>    <chr>     <chr>   <chr> <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 2021-03-… ClustA… Pack… "Too… 0.1.0   "c(per… "Arash… "A set… MIT + … UTF-8  
#>  2 2021-03-… comple… Pack… "Com… 0.1.0   "c(per… "Marco… "Provi… GPL-3   UTF-8  
#>  3 2021-03-… BayesR… Pack… "Bay… 1.0     "\nc(p… "Kuo-J… "To pe… GPL-2   UTF-8  
#>  4 2021-03-… ggfx    NA    "Pix… 1.0.0   "\nc(p… "Thoma… "Provi… MIT + … UTF-8  
#>  5 2021-03-… codeCo… Pack… "Col… 0.1.2   "\nper… "Jari … "Inclu… GPL-2   UTF-8  
#>  6 2021-03-… kindis… NA    "Sim… 0.9.2   "\nper… "Moshe… "Funct… MIT + … UTF-8  
#>  7 2021-03-… NonPar… Pack… "a N… 0.4.0   "c(per… "Josue… "Estim… GPL (>… UTF-8  
#>  8 2021-03-… forplo  Pack… "Fle… 0.1.0   "perso… "\"Vin… "Simpl… GPL-3   UTF-8  
#>  9 2021-03-… OpenSp… Pack… "Ana… 0.8.2   "c(per… "Win C… "Raman… CC BY … UTF-8  
#> 10 2021-03-… cmR     Pack… "Ana… 1.0.1   "c(per… "Volke… "Compu… GPL-3   UTF-8  
#> # … with 221 more rows, 38 more variables: RoxygenNote <chr>, URL <chr>,
#> #   VignetteBuilder <chr>, NeedsCompilation <chr>, Packaged <chr>,
#> #   Author <chr>, Repository <chr>, crandb_file_date <chr>, MD5sum <chr>,
#> #   date <chr>, dependencies <list>, LazyData <chr>, BugReports <chr>,
#> #   Date <chr>, RdMacros <chr>, Language <chr>, LazyLoad <chr>,
#> #   `Config/testthat/edition` <chr>, SystemRequirements <chr>, Collate <chr>,
#> #   OS_type <chr>, ByteCompile <chr>, LazyDataCompression <chr>, Note <chr>, …