List active packages

List active packages

cran_packages(from = "", limit = 10, format = c("short", "latest",
  "full"), archived = FALSE)

Arguments

from
The name of the first package to list. By default it is the first one in alphabetical order.

limit
The number of packages to list.

format
What to return. ‘short’ means the title and version number only. ‘latest’ means the complete description of the latest version. ‘full’ means all versions. Note that the output printing look the same for all formats, although the output is actually different.

archived
Whether to include archived packages in the result. If this is TRUE, then format must be ‘full’.

Value

List of packages.

Examples

# Only title and latest version
(l1 <- cran_packages(format = "short"))

#> CRAN packages (short)----------------------------------------------------------
#>  Package     Version Title                                                     
#>  A3          1.0.0   Accurate, Adaptable, and Accessible Error Metrics for P...
#>  abbyyR      0.5.4   Access to Abbyy Optical Character Recognition (OCR) API   
#>  abc         2.1     Tools for Approximate Bayesian Computation (ABC)          
#>  abc.data    1.0     Data Only: Tools for Approximate Bayesian Computation (...
#>  ABC.RAP     0.9.0   Array Based CpG Region Analysis Pipeline                  
#>  ABCanalysis 1.2.1   Computed ABC Analysis                                     
#>  abcdeFBA    0.4     ABCDE_FBA: A-Biologist-Can-Do-Everything of Flux Balanc...
#>  ABCoptim    0.15.0  Implementation of Artificial Bee Colony (ABC) Optimizat...
#>  ABCp2       1.2     Approximate Bayesian Computational Model for Estimating...
#>  abcrf       1.7.1   Approximate Bayesian Computation via Random Forests       

l1[[1]]

#> $version
#> [1] "1.0.0"
#> 
#> $title
#> [1] "Accurate, Adaptable, and Accessible Error Metrics for Predictive\nModels"
#> 

# Metadata for the latest version
(l2 <- cran_packages(format = "latest"))

#> CRAN packages (latest versions)------------------------------------------------
#>  Package     Version Title                                                     
#>  A3          1.0.0   Accurate, Adaptable, and Accessible Error Metrics for P...
#>  abbyyR      0.5.4   Access to Abbyy Optical Character Recognition (OCR) API   
#>  abc         2.1     Tools for Approximate Bayesian Computation (ABC)          
#>  abc.data    1.0     Data Only: Tools for Approximate Bayesian Computation (...
#>  ABC.RAP     0.9.0   Array Based CpG Region Analysis Pipeline                  
#>  ABCanalysis 1.2.1   Computed ABC Analysis                                     
#>  abcdeFBA    0.4     ABCDE_FBA: A-Biologist-Can-Do-Everything of Flux Balanc...
#>  ABCoptim    0.15.0  Implementation of Artificial Bee Colony (ABC) Optimizat...
#>  ABCp2       1.2     Approximate Bayesian Computational Model for Estimating...
#>  abcrf       1.7.1   Approximate Bayesian Computation via Random Forests       

l2[[1]]

#> $Package
#> [1] "A3"
#> 
#> $Type
#> [1] "Package"
#> 
#> $Title
#> [1] "Accurate, Adaptable, and Accessible Error Metrics for Predictive\nModels"
#> 
#> $Version
#> [1] "1.0.0"
#> 
#> $Date
#> [1] "2015-08-15"
#> 
#> $Author
#> [1] "Scott Fortmann-Roe"
#> 
#> $Maintainer
#> [1] "Scott Fortmann-Roe <scottfr@berkeley.edu>"
#> 
#> $Description
#> [1] "Supplies tools for tabulating and analyzing the results of predictive models. The methods employed are applicable to virtually any predictive model and make comparisons between different methodologies straightforward."
#> 
#> $License
#> [1] "GPL (>= 2)"
#> 
#> $Depends
#> $Depends$R
#> [1] ">= 2.15.0"
#> 
#> $Depends$xtable
#> [1] "*"
#> 
#> $Depends$pbapply
#> [1] "*"
#> 
#> 
#> $Suggests
#> $Suggests$randomForest
#> [1] "*"
#> 
#> $Suggests$e1071
#> [1] "*"
#> 
#> 
#> $NeedsCompilation
#> [1] "no"
#> 
#> $Packaged
#> [1] "2015-08-16 14:17:33 UTC; scott"
#> 
#> $Repository
#> [1] "CRAN"
#> 
#> $`Date/Publication`
#> [1] "2015-08-16 23:05:52"
#> 
#> $crandb_file_date
#> [1] "2015-08-16 17:08:22"
#> 
#> $date
#> [1] "2015-08-16T23:05:52+00:00"
#> 
#> $releases
#> list()
#> 

# All available metadata
(l3 <- cran_packages(format = "full"))

#> CRAN packages (full)-----------------------------------------------------------
#>  Package     Version Title                                                     
#>  A3          1.0.0   Accurate, Adaptable, and Accessible Error Metrics for P...
#>  abbyyR      0.5.4   Access to Abbyy Optical Character Recognition (OCR) API   
#>  abc         2.1     Tools for Approximate Bayesian Computation (ABC)          
#>  abc.data    1.0     Data Only: Tools for Approximate Bayesian Computation (...
#>  ABC.RAP     0.9.0   Array Based CpG Region Analysis Pipeline                  
#>  ABCanalysis 1.2.1   Computed ABC Analysis                                     
#>  abcdeFBA    0.4     ABCDE_FBA: A-Biologist-Can-Do-Everything of Flux Balanc...
#>  ABCoptim    0.15.0  Implementation of Artificial Bee Colony (ABC) Optimizat...
#>  ABCp2       1.2     Approximate Bayesian Computational Model for Estimating...
#>  abcrf       1.7.1   Approximate Bayesian Computation via Random Forests       

l3[[1]]

#> $name
#> [1] "A3"
#> 
#> $versions
#> $versions$`0.9.1`
#> $versions$`0.9.1`$Package
#> [1] "A3"
#> 
#> $versions$`0.9.1`$Type
#> [1] "Package"
#> 
#> $versions$`0.9.1`$Title
#> [1] "A3: Accurate, Adaptable, and Accessible Error Metrics for<U+000a>Predictive Models"
#> 
#> $versions$`0.9.1`$Version
#> [1] "0.9.1"
#> 
#> $versions$`0.9.1`$Date
#> [1] "2013-02-06"
#> 
#> $versions$`0.9.1`$Author
#> [1] "Scott Fortmann-Roe"
#> 
#> $versions$`0.9.1`$Maintainer
#> [1] "Scott Fortmann-Roe <scottfr@berkeley.edu>"
#> 
#> $versions$`0.9.1`$Description
#> [1] "This package supplies tools for tabulating and analyzing<U+000a>the results of predictive models. The methods employed are<U+000a>applicable to virtually any predictive model and make<U+000a>comparisons between different methodologies straightforward."
#> 
#> $versions$`0.9.1`$License
#> [1] "GPL (>= 2)"
#> 
#> $versions$`0.9.1`$Depends
#> $versions$`0.9.1`$Depends$xtable
#> [1] "*"
#> 
#> $versions$`0.9.1`$Depends$pbapply
#> [1] "*"
#> 
#> 
#> $versions$`0.9.1`$Suggests
#> $versions$`0.9.1`$Suggests$randomForest
#> [1] "*"
#> 
#> $versions$`0.9.1`$Suggests$e1071
#> [1] "*"
#> 
#> 
#> $versions$`0.9.1`$Packaged
#> [1] "2013-02-06 16:46:12 UTC; scott"
#> 
#> $versions$`0.9.1`$Repository
#> [1] "CRAN"
#> 
#> $versions$`0.9.1`$`Date/Publication`
#> [1] "2013-02-07 10:00:27"
#> 
#> $versions$`0.9.1`$crandb_file_date
#> [1] "2013-02-07 04:00:29"
#> 
#> $versions$`0.9.1`$date
#> [1] "2013-02-07T10:00:27+00:00"
#> 
#> $versions$`0.9.1`$releases
#> [1] "2.15.3"
#> 
#> 
#> $versions$`0.9.2`
#> $versions$`0.9.2`$Package
#> [1] "A3"
#> 
#> $versions$`0.9.2`$Type
#> [1] "Package"
#> 
#> $versions$`0.9.2`$Title
#> [1] "A3: Accurate, Adaptable, and Accessible Error Metrics for<U+000a>Predictive Models"
#> 
#> $versions$`0.9.2`$Version
#> [1] "0.9.2"
#> 
#> $versions$`0.9.2`$Date
#> [1] "2013-03-24"
#> 
#> $versions$`0.9.2`$Author
#> [1] "Scott Fortmann-Roe"
#> 
#> $versions$`0.9.2`$Maintainer
#> [1] "Scott Fortmann-Roe <scottfr@berkeley.edu>"
#> 
#> $versions$`0.9.2`$Description
#> [1] "This package supplies tools for tabulating and analyzing<U+000a>the results of predictive models. The methods employed are<U+000a>applicable to virtually any predictive model and make<U+000a>comparisons between different methodologies straightforward."
#> 
#> $versions$`0.9.2`$License
#> [1] "GPL (>= 2)"
#> 
#> $versions$`0.9.2`$Depends
#> $versions$`0.9.2`$Depends$R
#> [1] ">= 2.15.0"
#> 
#> $versions$`0.9.2`$Depends$xtable
#> [1] "*"
#> 
#> $versions$`0.9.2`$Depends$pbapply
#> [1] "*"
#> 
#> 
#> $versions$`0.9.2`$Suggests
#> $versions$`0.9.2`$Suggests$randomForest
#> [1] "*"
#> 
#> $versions$`0.9.2`$Suggests$e1071
#> [1] "*"
#> 
#> 
#> $versions$`0.9.2`$Packaged
#> [1] "2013-03-26 18:58:12 UTC; scott"
#> 
#> $versions$`0.9.2`$Repository
#> [1] "CRAN"
#> 
#> $versions$`0.9.2`$`Date/Publication`
#> [1] "2013-03-26 19:58:40"
#> 
#> $versions$`0.9.2`$crandb_file_date
#> [1] "2013-03-26 14:58:40"
#> 
#> $versions$`0.9.2`$NeedsCompilation
#> [1] "no"
#> 
#> $versions$`0.9.2`$date
#> [1] "2013-03-26T19:58:40+00:00"
#> 
#> $versions$`0.9.2`$releases
#> [1] "3.0.0" "3.0.1" "3.0.2" "3.0.3" "3.1.0" "3.1.1"
#> 
#> 
#> $versions$`1.0.0`
#> $versions$`1.0.0`$Package
#> [1] "A3"
#> 
#> $versions$`1.0.0`$Type
#> [1] "Package"
#> 
#> $versions$`1.0.0`$Title
#> [1] "Accurate, Adaptable, and Accessible Error Metrics for Predictive\nModels"
#> 
#> $versions$`1.0.0`$Version
#> [1] "1.0.0"
#> 
#> $versions$`1.0.0`$Date
#> [1] "2015-08-15"
#> 
#> $versions$`1.0.0`$Author
#> [1] "Scott Fortmann-Roe"
#> 
#> $versions$`1.0.0`$Maintainer
#> [1] "Scott Fortmann-Roe <scottfr@berkeley.edu>"
#> 
#> $versions$`1.0.0`$Description
#> [1] "Supplies tools for tabulating and analyzing the results of predictive models. The methods employed are applicable to virtually any predictive model and make comparisons between different methodologies straightforward."
#> 
#> $versions$`1.0.0`$License
#> [1] "GPL (>= 2)"
#> 
#> $versions$`1.0.0`$Depends
#> $versions$`1.0.0`$Depends$R
#> [1] ">= 2.15.0"
#> 
#> $versions$`1.0.0`$Depends$xtable
#> [1] "*"
#> 
#> $versions$`1.0.0`$Depends$pbapply
#> [1] "*"
#> 
#> 
#> $versions$`1.0.0`$Suggests
#> $versions$`1.0.0`$Suggests$randomForest
#> [1] "*"
#> 
#> $versions$`1.0.0`$Suggests$e1071
#> [1] "*"
#> 
#> 
#> $versions$`1.0.0`$NeedsCompilation
#> [1] "no"
#> 
#> $versions$`1.0.0`$Packaged
#> [1] "2015-08-16 14:17:33 UTC; scott"
#> 
#> $versions$`1.0.0`$Repository
#> [1] "CRAN"
#> 
#> $versions$`1.0.0`$`Date/Publication`
#> [1] "2015-08-16 23:05:52"
#> 
#> $versions$`1.0.0`$crandb_file_date
#> [1] "2015-08-16 17:08:22"
#> 
#> $versions$`1.0.0`$date
#> [1] "2015-08-16T23:05:52+00:00"
#> 
#> $versions$`1.0.0`$releases
#> list()
#> 
#> 
#> 
#> $timeline
#> $timeline$`0.9.1`
#> [1] "2013-02-07T10:00:27+00:00"
#> 
#> $timeline$`0.9.2`
#> [1] "2013-03-26T19:58:40+00:00"
#> 
#> $timeline$`1.0.0`
#> [1] "2015-08-16T23:05:52+00:00"
#> 
#> 
#> $latest
#> [1] "1.0.0"
#> 
#> $title
#> [1] "Accurate, Adaptable, and Accessible Error Metrics for Predictive\nModels"
#> 
#> $archived
#> [1] FALSE
#>