Trending R packages

Trending packages are the ones that were downloaded at least 1000 times during last week, and that substantially increased their download counts, compared to the average weekly downloads in the previous 24 weeks. The percentage of increase is also shown in the output.

cran_trending()

Value

Tibble of trending packages.

Examples

cran_trending()

#> # A tibble: 100 x 2
#>    package           score                 
#>    <chr>             <chr>                 
#>  1 effectsize        62015.4696132596685100
#>  2 leaflet.providers 6643.3748373908195500 
#>  3 farver            2829.3291388772503400 
#>  4 renv              1054.6458141674333000 
#>  5 joint.Cox         713.9481572914829800  
#>  6 rsq               605.0799623706491100  
#>  7 NNS               604.4613918017159200  
#>  8 fable             575.6197740941763200  
#>  9 lobstr            558.1014729950900200  
#> 10 feasts            523.3285233285233300  
#> # … with 90 more rows