RStudio addin to search CRAN packages

Call this function from RStudio for best results. You can also use it without RStudio, then it will run in the web browser.

pkg_search_addin(query = "", viewer = c("dialog", "browser"))

Arguments

query

Query string to start the addin with.

viewer

Whether to show the addin within RStudio ("dialog"), or in a web browser ("browser").

Details

The app has:

A search tab for free text search, very much like the pkg_search() function.

The list of recently updated packages.

The list of top packages: most downloaded, most depended upon, and trending packages.

Package list by maintainer.

Examples

pkg_search_addin()

# Start with a search query
pkg_search_addin("permutation test")