Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...
Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json
A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")

#> - "advanced search" ---------------------------- 54 packages in 0.001 seconds -
#>   #     package       version   by         @ title                             
#>   1 100 gplots        3.0.1.2   ORPHANED 25d Various R Programming Tools for...
#>   2 100 aws.s3        0.3.12.1  ORPHANED 23d 'AWS S3' Client Package           
#>   3 100 DAAG          1.22.1    ORPHANED  1y Data Analysis and Graphics Data...
#>   4 100 trimcluster   0.1.2.1   ORPHANED  2y Cluster Analysis with Trimming    
#>   5 100 distances     0.1.8     ORPHANED  5M Tools for Distance Metrics        
#>   6 100 mvcwt         1.3.1     ORPHANED  1y Wavelet Analysis of Multiple Ti...
#>   7 100 Funclustering 1.0.2     ORPHANED  1y Functional Data Clustering        
#>   8  96 robust        0.4.18.2  ORPHANED  1M Port of the S+ "Robust Library"   
#>   9  96 SDMTools      1.1.221.2 ORPHANED  2M Species Distribution Modelling ...
#>  10  96 aws.polly     0.1.2.1   ORPHANED 23d Client for AWS Polly              

# Packages that match a certain field
advanced_search(Author = "Hester AND NOT Wickham")

#> - "advanced search" ---------------------------- 21 packages in 0.004 seconds -
#>   #     package        version by              @ title                         
#>   1 100 gmailr         1.0.0   Jim Hester     6M Access the 'Gmail' 'RESTful...
#>   2  98 primerTree     1.0.5   Jim Hester    13d Visually Assessing the Spec...
#>   3  93 knitrBootstrap 1.0.2   Jim Hester     2y 'knitr' Bootstrap Framework   
#>   4  88 glue           1.3.1   Jim Hester    11M Interpreted String Literals   
#>   5  85 types          1.0.0   Jim Hester     3y Type Annotations              
#>   6  73 lintr          2.0.0   Jim Hester     4M A 'Linter' for R Code         
#>   7  71 dplyr.teradata 0.3.2   Koji Makiyama  9M A 'Teradata' Backend for 'd...
#>   8  70 pak            0.1.2   Gábor Csárdi   1y Another Approach to Package...
#>   9  65 mockery        0.4.2   Jim Hester     5M Mocking Library for R         
#>  10  65 bench          1.1.1   Jim Hester    23d High Precision Timing of R ...

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")

#> - "advanced search" --------------------------- 254 packages in 0.017 seconds -
#>   #     package           version by                         @ title           
#>   1 100 phd               0.1     Jesse Hemerik             7M Permutation T...
#>   2  96 confSAM           0.2     Jesse Hemerik             2y Estimates and...
#>   3  85 permutes          1.0     Cesko C. Voeten           7M Permutation T...
#>   4  79 BurStMisc         1.1     Pat Burns                 3y Burns Statist...
#>   5  77 fdrci             2.1     Joshua Millstein          3y Permutation-B...
#>   6  74 packager          1.0.0   Andreas Dominik Cullmann 23d Create, Build...
#>   7  74 phyloseqGraphTest 0.0.2   Julia Fukuyama            2y Graph-Based P...
#>   8  74 resample          0.4     Tim Hesterberg            5y Resampling Fu...
#>   9  62 shiftR            1.5     Andrey A Shabalin        11M Fast Enrichme...
#>  10  61 flipscores        0.2     Jesse Hemerik             1y Robust Testin...

# Packages that have a certain field
advanced_search("_exists_" = "URL")

#> - "advanced search" -------------------------- 7706 packages in 0.001 seconds -
#>   #     package       version by                         @ title               
#>   1 100 packager      1.0.0   Andreas Dominik Cullmann 23d Create, Build and...
#>   2 100 rocTree       1.1.0   Sy Han Chiou              4M Receiver Operatin...
#>   3 100 miceExt       1.1.0   Tobias Schumacher         2y Extension Package...
#>   4 100 yearn         0.1.3   Brian O'Meara             2y Use and if Needed...
#>   5 100 ubeR          0.1.4   Andrew Collier            3y Interface to the ...
#>   6 100 noteMD        0.1.0   Jiena McLellan            1y Print Text from '...
#>   7 100 PoloniexR     0.0.1   Jellen Vermeir            3y Interface to the ...
#>   8 100 softermax     0.3.0   Brian Connelly            2y Read Exported Dat...
#>   9 100 rFTRLProximal 1.0.0   Yachen Yan                3y FTRL-Proximal Alg...
#>  10 100 frite         0.1.0   Christopher Peralta       2y Functions Writing...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")

#> - "advanced search" -------------------------- 8341 packages in 0.001 seconds -
#>   #     package       version by                  @ title                      
#>   1 100 FUNLDA        1.1     Daniel Backenroth  3y Genomic Latent Dirichlet...
#>   2 100 SeqKat        0.0.6   Paul C. Boutros    2y Detection of Kataegis      
#>   3 100 ELMSurv       0.6     Hong Wang          8M Extreme Learning Machine...
#>   4 100 miniparquet   0.1.2   Hannes Mühleisen   4M Read Parquet Files         
#>   5 100 diffMeanVar   0.0.6   Weiliang Qiu       2y Detecting Gene Probes wi...
#>   6 100 NRejections   1.0.0   Maya B. Mathur     1y Metrics for Multiple Tes...
#>   7 100 SSL           0.1     Junxiang Wang      4y Semi-Supervised Learning   
#>   8 100 DevTreatRules 1.0.0   Jeremy Roth        9M Develop Treatment Rules ...
#>   9 100 CpDyna        1.0     Margot Selosse     1y Change Point Detection i...
#>  10 100 xlm           0.1.1   Oliver Frost       1y A 'Stellar' Client         

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)

#> - "advanced search" ------------------------------ 8341 packages in 0 seconds -
#>   #     package       version by                  @ title                      
#>   1 100 FUNLDA        1.1     Daniel Backenroth  3y Genomic Latent Dirichlet...
#>   2 100 SeqKat        0.0.6   Paul C. Boutros    2y Detection of Kataegis      
#>   3 100 ELMSurv       0.6     Hong Wang          8M Extreme Learning Machine...
#>   4 100 miniparquet   0.1.2   Hannes Mühleisen   4M Read Parquet Files         
#>   5 100 diffMeanVar   0.0.6   Weiliang Qiu       2y Detecting Gene Probes wi...
#>   6 100 NRejections   1.0.0   Maya B. Mathur     1y Metrics for Multiple Tes...
#>   7 100 SSL           0.1     Junxiang Wang      4y Semi-Supervised Learning   
#>   8 100 DevTreatRules 1.0.0   Jeremy Roth        9M Develop Treatment Rules ...
#>   9 100 CpDyna        1.0     Margot Selosse     1y Change Point Detection i...
#>  10 100 xlm           0.1.1   Oliver Frost       1y A 'Stellar' Client         

# Regular expressions
advanced_search(Author = "/Joh?nathan/")

#> - "advanced search" --------------------------- 189 packages in 0.002 seconds -
#>   #     package         version by                         @ title             
#>   1 100 ubeR            0.1.4   Andrew Collier            3y Interface to th...
#>   2 100 texPreview      1.4.3   Jonathan Sidi             2M Compile and Pre...
#>   3 100 knitr           1.27    Yihui Xie                20d A General-Purpo...
#>   4 100 raster          3.0.12  Robert J. Hijmans         6d Geographic Data...
#>   5 100 xtable          1.8.4   David Scott              10M Export Tables t...
#>   6 100 tableone        0.10.0  Kazuki Yoshida            1y Create 'Table 1...
#>   7 100 bridgesampling  0.8.1   Quentin F. Gronau        20d Bridge Sampling...
#>   8 100 gdalUtils       2.0.3   Jonathan Asher Greenberg 20d Wrappers for th...
#>   9 100 fivethirtyeight 0.5.0   Albert Y. Kim             6M Data and Code B...
#>  10 100 tscount         1.4.1   Tobias Liboschik          2y Analysis of Cou...

# Fuzzy search
advanced_search(Author = "Johnathan~1")

#> - "advanced search" --------------------------- 190 packages in 0.002 seconds -
#>   #     package       version by                                        @ title
#>   1 100 UScancer      0.1.2   Jonathan Lee                             5y Cr...
#>   2 100 mlmi          1.0.0   Jonathan Bartlett                        6M Ma...
#>   3  99 lda           1.4.2   Jonathan Chang                           4y Co...
#>   4  97 rSHAPE        0.3.2   Jonathan Dench                           7M Si...
#>   5  95 chords        0.95.4  Jonathan Rosenblatt                      3y Es...
#>   6  95 bootImpute    1.0.0   Jonathan Bartlett                        7M Bo...
#>   7  95 handyplots    1.1.3   Jonathan Schwartz                        1y Ha...
#>   8  95 UnitCircle    0.1.3   Jonathan Berrisch                        2y Ch...
#>   9  94 spatial.tools 1.6.0   Jonathan Asher Greenberg                 2y R ...
#>  10  94 geomapdata    1.0.4   Jonathan M. Lees<jonathan.lees@unc.edu>  7y Da...