Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...
Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json
A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

Value

Search hits.

Examples

 if (pingr::is_online()) withAutoprint({
# All orphaned packages
advanced_search(Maintainer = "ORPHANED")

# Packages that match a certain field
advanced_search(Author = "Hester AND NOT Wickham")

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")

# Packages that have a certain field
advanced_search("_exists_" = "URL")

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)

# Regular expressions
advanced_search(Author = "/Joh?nathan/")

# Fuzzy search
advanced_search(Author = "Johnathan~1")
 })

#> > advanced_search(Maintainer = "ORPHANED")
#> - "advanced search" ---------------------------- 60 packages in 0.001 seconds -
#>   #     package      version   by         @ title                              
#>   1 100 SDMTools     1.1.221.1 ORPHANED  7M Species Distribution Modelling T...
#>   2 100 AlgDesign    1.1.7.3   ORPHANED  5y Algorithmic Experimental Design    
#>   3 100 regress      1.3.15    ORPHANED  3y Gaussian Linear Models with Line...
#>   4 100 smoothSurv   2.0.1     ORPHANED  1y Survival Regression with Smoothe...
#>   5 100 DirichletReg 0.6.3.1   ORPHANED 10M Dirichlet Regression in R          
#>   6 100 bfp          0.0.40    ORPHANED 10M Bayesian Fractional Polynomials    
#>   7 100 datadr       0.8.6.1   ORPHANED  1y Divide and Recombine for Large, ...
#>   8 100 nhlscrapr    1.8.1     ORPHANED  3y Compiling the NHL Real Time Scor...
#>   9  98 caTools      1.17.1.2  ORPHANED  8M Tools: moving window statistics,...
#>  10  98 DAAG         1.22.1    ORPHANED  9M Data Analysis and Graphics Data ...
#> > advanced_search(Author = "Hester AND NOT Wickham")
#> - "advanced search" ---------------------------- 21 packages in 0.006 seconds -
#>   #     package        version by              @ title                         
#>   1 100 gmailr         1.0.0   Jim Hester     3M Access the 'Gmail' 'RESTful...
#>   2  98 primerTree     1.0.4   Jim Hester    11M Visually Assessing the Spec...
#>   3  93 knitrBootstrap 1.0.2   Jim Hester     1y 'knitr' Bootstrap Framework   
#>   4  88 glue           1.3.1   Jim Hester     8M Interpreted String Literals   
#>   5  85 types          1.0.0   Jim Hester     3y Type Annotations              
#>   6  82 bench          1.0.4   Jim Hester     2M High Precision Timing of R ...
#>   7  73 lintr          2.0.0   Jim Hester     1M A 'Linter' for R Code         
#>   8  70 dplyr.teradata 0.3.2   Koji Makiyama  6M A 'Teradata' Backend for 'd...
#>   9  70 pak            0.1.2   Gábor Csárdi   9M Another Approach to Package...
#>  10  64 mockery        0.4.2   Jim Hester     2M Mocking Library for R         
#> > advanced_search("permutation test AND NOT License: GPL OR GNU")
#> - "advanced search" --------------------------- 247 packages in 0.022 seconds -
#>   #     package           version by                  @ title                  
#>   1 100 phd               0.1     Jesse Hemerik      4M Permutation Testing ...
#>   2  96 confSAM           0.2     Jesse Hemerik      2y Estimates and Bounds...
#>   3  85 permutes          1.0     Cesko C. Voeten    4M Permutation Tests fo...
#>   4  78 BurStMisc         1.1     Pat Burns          3y Burns Statistics Mis...
#>   5  77 fdrci             2.1     Joshua Millstein   3y Permutation-Based FD...
#>   6  75 phyloseqGraphTest 0.0.2   Julia Fukuyama     1y Graph-Based Permutat...
#>   7  73 resample          0.4     Tim Hesterberg     5y Resampling Functions   
#>   8  63 shiftR            1.5     Andrey A Shabalin  8M Fast Enrichment Anal...
#>   9  61 flipscores        0.2     Jesse Hemerik      1y Robust Testing in GLMs 
#>  10  57 GEEaSPU           1.0.2   Junghi Kim         3y Adaptive Association...
#> > advanced_search(`_exists_` = "URL")
#> - "advanced search" -------------------------- 7432 packages in 0.001 seconds -
#>   #     package          version by                   @ title                  
#>   1 100 yearn            0.1.3   Brian O'Meara       2y Use and if Needed In...
#>   2 100 wrspathrowData   1.0     Alex Zvoleff        6y Data used by the wrs...
#>   3 100 PoloniexR        0.0.1   Jellen Vermeir      2y Interface to the Pol...
#>   4 100 softermax        0.3.0   Brian Connelly      2y Read Exported Data f...
#>   5 100 anomalyDetection 0.2.5   Bradley Boehmke     2y Implementation of Au...
#>   6 100 cliqueMS         0.3.1   Oriol Senan Campos  7M Annotation of Isotop...
#>   7 100 rFTRLProximal    1.0.0   Yachen Yan          3y FTRL-Proximal Algorithm
#>   8 100 sylly.en         0.1.3   Meik Michalke       2y Language Support for...
#>   9 100 netrankr         0.2.1   David Schoch        1y Analyzing Partial Ra...
#>  10 100 mdsr             0.1.7   Ben Baumer          6M Complement to 'Moder...
#> > advanced_search("NOT _exists_: URL")
#> - "advanced search" -------------------------- 8227 packages in 0.001 seconds -
#>   #     package         version by                     @ title                 
#>   1 100 miniparquet     0.1.2   Hannes Mühleisen     22d Read Parquet Files    
#>   2 100 ELMSurv         0.6     Hong Wang             6M Extreme Learning Ma...
#>   3 100 SSL             0.1     Junxiang Wang         3y Semi-Supervised Lea...
#>   4 100 icsurvROC       0.1.0   Yunro Chung           1y Interval Censored S...
#>   5 100 RWeightedKmeans 1.1.0   Alexandre Gondeau     1y Weighted Object k-M...
#>   6 100 CpDyna          1.0     Margot Selosse        1y Change Point Detect...
#>   7 100 GHap            1.2.2   Yuri Tani Utsunomiya  3y Genome-Wide Haploty...
#>   8 100 FUNLDA          1.1     Daniel Backenroth     2y Genomic Latent Diri...
#>   9 100 doSNOW          1.0.18  Hong Ooi              4M Foreach Parallel Ad...
#>  10 100 nor1mix         1.3.0   Martin Maechler       5M Normal aka Gaussian...
#> > query <- "{\n\"query\": {\n  \"bool\": {\n    \"must_not\": {\n      \"exists\": {\n        \"field\": \"URL\"\n      }\n    }\n  }\n}\n}"
#> > advanced_search(json = query)
#> - "advanced search" ------------------------------ 8227 packages in 0 seconds -
#>   #     package         version by                     @ title                 
#>   1 100 miniparquet     0.1.2   Hannes Mühleisen     22d Read Parquet Files    
#>   2 100 ELMSurv         0.6     Hong Wang             6M Extreme Learning Ma...
#>   3 100 SSL             0.1     Junxiang Wang         3y Semi-Supervised Lea...
#>   4 100 icsurvROC       0.1.0   Yunro Chung           1y Interval Censored S...
#>   5 100 RWeightedKmeans 1.1.0   Alexandre Gondeau     1y Weighted Object k-M...
#>   6 100 CpDyna          1.0     Margot Selosse        1y Change Point Detect...
#>   7 100 GHap            1.2.2   Yuri Tani Utsunomiya  3y Genome-Wide Haploty...
#>   8 100 FUNLDA          1.1     Daniel Backenroth     2y Genomic Latent Diri...
#>   9 100 doSNOW          1.0.18  Hong Ooi              4M Foreach Parallel Ad...
#>  10 100 nor1mix         1.3.0   Martin Maechler       5M Normal aka Gaussian...
#> > advanced_search(Author = "/Joh?nathan/")
#> - "advanced search" --------------------------- 184 packages in 0.001 seconds -
#>   #     package         version  by                         @ title            
#>   1 100 texPreview      1.4.1    Jonathan Sidi            14d Compile and Pr...
#>   2 100 mlergm          0.4      Jonathan Stewart          8d Multilevel Exp...
#>   3 100 ubeR            0.1.4    Andrew Collier            3y Interface to t...
#>   4 100 knitr           1.25     Yihui Xie                 2M A General-Purp...
#>   5 100 xtable          1.8.4    David Scott               7M Export Tables ...
#>   6 100 raster          3.0.7    Robert J. Hijmans         2M Geographic Dat...
#>   7 100 gdalUtils       2.0.1.14 Jonathan Asher Greenberg  2y Wrappers for t...
#>   8 100 bridgesampling  0.7.2    Quentin F. Gronau         4M Bridge Samplin...
#>   9 100 tableone        0.10.0   Kazuki Yoshida            9M Create 'Table ...
#>  10 100 fivethirtyeight 0.5.0    Albert Y. Kim             3M Data and Code ...
#> > advanced_search(Author = "Johnathan~1")
#> - "advanced search" --------------------------- 185 packages in 0.002 seconds -
#>   #     package       version by                                        @ title
#>   1 100 UScancer      0.1.2   Jonathan Lee                             5y Cr...
#>   2 100 mlmi          1.0.0   Jonathan Bartlett                        3M Ma...
#>   3  99 lda           1.4.2   Jonathan Chang                           4y Co...
#>   4  96 rSHAPE        0.3.2   Jonathan Dench                           4M Si...
#>   5  95 chords        0.95.4  Jonathan Rosenblatt                      3y Es...
#>   6  95 bootImpute    1.0.0   Jonathan Bartlett                        4M Bo...
#>   7  94 handyplots    1.1.3   Jonathan Schwartz                       10M Ha...
#>   8  94 UnitCircle    0.1.3   Jonathan Berrisch                        1y Ch...
#>   9  94 geomapdata    1.0.4   Jonathan M. Lees<jonathan.lees@unc.edu>  7y Da...
#>  10  94 spatial.tools 1.6.0   Jonathan Asher Greenberg                 2y R ...