Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...

Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json

A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")
#> - "advanced search" ----------------------------- 5 packages in 0.001 seconds -
#>   #     package   version  by         @ title                                  
#>  1  100 ggmap     3.0.0    ORPHANED  3y Spatial Visualization with ggplot2     
#>  2  100 udunits2  0.13.2.1 ORPHANED  4M Udunits-2 Bindings for R               
#>  3   97 distances 0.1.8    ORPHANED  3y Tools for Distance Metrics             
#>  4   94 bibtex    0.4.2.3  ORPHANED  2y Bibtex Parser                          
#>  5   94 nparLD    2.1.1    ORPHANED 19d Nonparametric Analysis of Longitudin...

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")
#> - "advanced search" ---------------------------- 19 packages in 0.002 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi       7M Find Tools Needed to Build...
#>   2  95 mockery     0.4.3   Hadley Wickham     5M Mocking Library for R        
#>   3  82 memoise     2.0.1   Winston Chang      8M 'Memoisation' of Functions   
#>   4  78 xml2        1.3.3   Hadley Wickham     8M Parse XML                    
#>   5  75 devtools    2.4.3   Jennifer Bryan     8M Tools to Make Developing R...
#>   6  75 sessioninfo 1.2.2   Gábor Csárdi       7M R Session Information        
#>   7  73 odbc        1.3.3   Hadley Wickham     8M Connect to ODBC Compatible...
#>   8  69 remotes     2.4.2   Gábor Csárdi       8M R Package Installation fro...
#>   9  68 meltr       1.0.0   Duncan Garmonsway 11M Read Non-Rectangular Text ...
#>  10  66 pkgload     1.3.0   Lionel Henry      19d Simulate Package Installat...
advanced_search("Author: Hester AND Author: Wickham")
#> - "advanced search" ---------------------------- 19 packages in 0.002 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi       7M Find Tools Needed to Build...
#>   2  95 mockery     0.4.3   Hadley Wickham     5M Mocking Library for R        
#>   3  82 memoise     2.0.1   Winston Chang      8M 'Memoisation' of Functions   
#>   4  78 xml2        1.3.3   Hadley Wickham     8M Parse XML                    
#>   5  75 devtools    2.4.3   Jennifer Bryan     8M Tools to Make Developing R...
#>   6  75 sessioninfo 1.2.2   Gábor Csárdi       7M R Session Information        
#>   7  73 odbc        1.3.3   Hadley Wickham     8M Connect to ODBC Compatible...
#>   8  69 remotes     2.4.2   Gábor Csárdi       8M R Package Installation fro...
#>   9  68 meltr       1.0.0   Duncan Garmonsway 11M Read Non-Rectangular Text ...
#>  10  66 pkgload     1.3.0   Lionel Henry      19d Simulate Package Installat...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")
#> - "advanced search" ---------------------------- 26 packages in 0.005 seconds -
#>   #     package        version by                @ title                       
#>   1 100 knitrBootstrap 1.0.2   Jim Hester       4y 'knitr' Bootstrap Framework 
#>   2  93 types          1.0.0   Jim Hester       6y Type Annotations            
#>   3  82 primerTree     1.0.6   Matt Cannon      3M Visually Assessing the Sp...
#>   4  78 pak            0.3.0   Gábor Csárdi     3M Another Approach to Packa...
#>   5  78 gmailr         1.0.1   Jennifer Bryan   8M Access the 'Gmail' 'RESTf...
#>   6  74 decor          1.0.1   Romain François  8M Retrieve Code Decorations   
#>   7  70 rex            1.2.1   Kevin Ushey      8M Friendly Regular Expressions
#>   8  66 bench          1.1.2   Davis Vaughan    8M High Precision Timing of ...
#>   9  64 roger          0.99.2  Vincent Goulet   6M Automated Grading of R Sc...
#>  10  63 spelling       2.2     Jeroen Ooms      2y Tools for Spell Checking ...

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")
#> - "advanced search" ---------------------------- 19 packages in 0.007 seconds -
#>   #     package     version by               @ title                           
#>   1 100 mockery     0.4.3   Hadley Wickham  5M Mocking Library for R           
#>   2  93 pkgbuild    1.3.1   Gábor Csárdi    7M Find Tools Needed to Build R ...
#>   3  89 odbc        1.3.3   Hadley Wickham  8M Connect to ODBC Compatible Da...
#>   4  87 xml2        1.3.3   Hadley Wickham  8M Parse XML                       
#>   5  76 memoise     2.0.1   Winston Chang   8M 'Memoisation' of Functions      
#>   6  71 devtools    2.4.3   Jennifer Bryan  8M Tools to Make Developing R Pa...
#>   7  69 sessioninfo 1.2.2   Gábor Csárdi    7M R Session Information           
#>   8  68 remotes     2.4.2   Gábor Csárdi    8M R Package Installation from R...
#>   9  67 pkgload     1.3.0   Lionel Henry   19d Simulate Package Installation...
#>  10  65 fs          1.5.2   Gábor Csárdi    7M Cross-Platform File System Op...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")
#> - "advanced search" ---------------------------- 26 packages in 0.005 seconds -
#>   #     package        version by                @ title                       
#>   1 100 knitrBootstrap 1.0.2   Jim Hester       4y 'knitr' Bootstrap Framework 
#>   2  93 types          1.0.0   Jim Hester       6y Type Annotations            
#>   3  82 primerTree     1.0.6   Matt Cannon      3M Visually Assessing the Sp...
#>   4  78 pak            0.3.0   Gábor Csárdi     3M Another Approach to Packa...
#>   5  78 gmailr         1.0.1   Jennifer Bryan   8M Access the 'Gmail' 'RESTf...
#>   6  74 decor          1.0.1   Romain François  8M Retrieve Code Decorations   
#>   7  70 rex            1.2.1   Kevin Ushey      8M Friendly Regular Expressions
#>   8  66 bench          1.1.2   Davis Vaughan    8M High Precision Timing of ...
#>   9  64 roger          0.99.2  Vincent Goulet   6M Automated Grading of R Sc...
#>  10  63 spelling       2.2     Jeroen Ooms      2y Tools for Spell Checking ...

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")
#> - "advanced search" --------------------------- 376 packages in 0.032 seconds -
#>   #     package           version by                         @ title           
#>   1 100 phd               0.2     Jesse Hemerik             6M Permutation T...
#>   2  92 confSAM           0.2     Jesse Hemerik             4y Estimates and...
#>   3  83 permutes          2.6     Cesko C. Voeten           3d Permutation T...
#>   4  78 BurStMisc         1.1     Pat Burns                 6y Burns Statist...
#>   5  75 resample          0.6     Tim Hesterberg            1M Resampling Fu...
#>   6  74 packager          1.14.0  Andreas Dominik Cullmann  7M Create, Build...
#>   7  73 USP               0.1.2   Thomas B. Berrett         1y U-Statistic P...
#>   8  72 fdrci             2.3     Joshua Millstein          1M Permutation-B...
#>   9  71 MultSurvTests     0.2     Lukas Arnroth             1y Permutation T...
#>  10  69 phyloseqGraphTest 0.1.0   Julia Fukuyama            2y Graph-Based P...

# Packages that have a certain field
advanced_search("_exists_" = "URL")
#> - "advanced search" -------------------------- 9972 packages in 0.003 seconds -
#>   #     package       version by                     @ title                   
#>   1 100 countrycode   1.4.0   Vincent Arel-Bundock  2M Convert Country Names...
#>   2 100 MultinomialCI 1.2     Pablo J. Villacorta   1y Simultaneous Confiden...
#>   3 100 pixarfilms    0.2.1   Eric Leung            1y Pixar Films and Achie...
#>   4 100 optimization  1.0.9   Kai Husmann           5M Flexible Optimization...
#>   5 100 signs         0.1.2   Benjamin E. Wolfe     2y Insert Proper Minus S...
#>   6 100 wcde          0.0.5   Guy J. Abel           1M Download Data from th...
#>   7 100 crestr        1.1.0   Manuel Chevalier     15d A Probabilistic Appro...
#>   8 100 censable      0.0.3   Christopher T. Kenny  9M Making Census Data Mo...
#>   9 100 blockmatrix   1.0     Emanuele Cordano      8y blockmatrix: Tools to...
#>  10 100 rgabriel      0.9     Miao YU               2M Gabriel Multiple Comp...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")
#> - "advanced search" -------------------------- 8408 packages in 0.003 seconds -
#>   #     package          version    by                 @ title                 
#>   1 100 fontLiberation   0.1.0      Lionel Henry      6y Liberation Fonts      
#>   2 100 mcmcplots        0.4.3      S. McKay Curtis   4y Create Plots from M...
#>   3 100 naturalsort      0.1.3      Kosei Abe         6y Natural Ordering      
#>   4 100 optimr           2019.12.16 John C Nash       3y A Replacement and E...
#>   5 100 modest           0.3.1      Philip Pallmann   5y Model-Based Dose-Es...
#>   6 100 icrf             2.0.0      Hunyong Cho       1y Interval Censored R...
#>   7 100 PoissonPCA       1.0.3      Toby Kenney      11M Poisson-Noise Corre...
#>   8 100 panelaggregation 0.1.1      Matthias Bannert  6y Aggregate Longitudi...
#>   9 100 ProAE            0.2.10     Blake Langlais   16d PRO-CTCAE Scoring, ...
#>  10 100 LTRCtrees        1.1.1      Wenbo Jing        2y Survival Trees to F...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)
#> - "advanced search" -------------------------- 8408 packages in 0.001 seconds -
#>   #     package          version    by                 @ title                 
#>   1 100 fontLiberation   0.1.0      Lionel Henry      6y Liberation Fonts      
#>   2 100 mcmcplots        0.4.3      S. McKay Curtis   4y Create Plots from M...
#>   3 100 naturalsort      0.1.3      Kosei Abe         6y Natural Ordering      
#>   4 100 optimr           2019.12.16 John C Nash       3y A Replacement and E...
#>   5 100 modest           0.3.1      Philip Pallmann   5y Model-Based Dose-Es...
#>   6 100 icrf             2.0.0      Hunyong Cho       1y Interval Censored R...
#>   7 100 PoissonPCA       1.0.3      Toby Kenney      11M Poisson-Noise Corre...
#>   8 100 panelaggregation 0.1.1      Matthias Bannert  6y Aggregate Longitudi...
#>   9 100 ProAE            0.2.10     Blake Langlais   16d PRO-CTCAE Scoring, ...
#>  10 100 LTRCtrees        1.1.1      Wenbo Jing        2y Survival Trees to F...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")
#> - "advanced search" --------------------------- 229 packages in 0.004 seconds -
#>   #     package         version by                           @ title           
#>   1 100 Sample.Size     1.0     Wei Jiang<wjiang@kumc.edu>  9y Sample size c...
#>   2 100 knitr           1.39    Yihui Xie                   3M A General-Pur...
#>   3 100 raster          3.5.21  Robert J. Hijmans          19d Geographic Da...
#>   4 100 xtable          1.8.4   David Scott                 3y Export Tables...
#>   5 100 bridgesampling  1.1.2   Quentin F. Gronau           1y Bridge Sampli...
#>   6 100 tableone        0.13.2  Kazuki Yoshida              3M Create 'Table...
#>   7 100 tscount         1.4.3   Tobias Liboschik            2y Analysis of C...
#>   8 100 fivethirtyeight 0.6.2   Albert Y. Kim               9M Data and Code...
#>   9 100 slickR          0.5.0   Jonathan Sidi               2y Create Intera...
#>  10 100 smcfcs          1.7.0   Jonathan Bartlett          24d Multiple Impu...

# Fuzzy search
advanced_search(Author = "Johnathan~1")
#> - "advanced search" --------------------------- 230 packages in 0.006 seconds -
#>   #     package    version by                           @ title                
#>   1 100 lda        1.4.2   Jonathan Chang              7y Collapsed Gibbs Sa...
#>   2  98 mlmi       1.1.1   Jonathan Bartlett           8M Maximum Likelihood...
#>   3  98 bootImpute 1.2.0   Jonathan Bartlett           1y Bootstrap Inferenc...
#>   4  98 chords     0.95.4  Jonathan Rosenblatt         5y Estimation in Resp...
#>   5  95 RSeed      0.1.60  Claus Jonathan Fritzemeier  6y Borenstein Analysis  
#>   6  94 handyplots 1.1.3   Jonathan Schwartz           3y Handy Plots          
#>   7  94 UnitCircle 0.1.3   Jonathan Berrisch           4y Check if Roots of ...
#>   8  93 rSHAPE     0.3.2   Jonathan Dench              3y Simulated Haploid ...
#>   9  90 tensor     1.5     Jonathan Rougier           10y Tensor product of ...
#>  10  90 tint       0.1.3   Dirk Eddelbuettel           2y 'tint' is not 'Tufte'