Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...

Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json

A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")
#> - "advanced search" ----------------------------- 7 packages in 0.001 seconds -
#>   #     package     version   by         @ title                               
#>  1  100 RGtk2       2.20.36.2 ORPHANED  8M R Bindings for Gtk 2.8.0 and Above  
#>  2  100 distances   0.1.8     ORPHANED  3y Tools for Distance Metrics          
#>  3  100 udunits2    0.13.2.1  ORPHANED  3M Udunits-2 Bindings for R            
#>  4   99 cairoDevice 2.28.2.1  ORPHANED  8M Embeddable Cairo Graphics Device ...
#>  5   99 ggmap       3.0.0     ORPHANED  3y Spatial Visualization with ggplot2  
#>  6   94 bibtex      0.4.2.3   ORPHANED  2y Bibtex Parser                       
#>  7   94 gtools      3.9.2.2   ORPHANED  4d Various R Programming Tools         

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")
#> - "advanced search" ---------------------------- 19 packages in 0.001 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi       6M Find Tools Needed to Build...
#>   2  94 mockery     0.4.3   Hadley Wickham     4M Mocking Library for R        
#>   3  85 memoise     2.0.1   Winston Chang      7M 'Memoisation' of Functions   
#>   4  80 xml2        1.3.3   Hadley Wickham     7M Parse XML                    
#>   5  76 sessioninfo 1.2.2   Gábor Csárdi       6M R Session Information        
#>   6  75 devtools    2.4.3   Jennifer Bryan     7M Tools to Make Developing R...
#>   7  73 odbc        1.3.3   Hadley Wickham     7M Connect to ODBC Compatible...
#>   8  70 pkgload     1.2.4   Lionel Henry       7M Simulate Package Installat...
#>   9  70 remotes     2.4.2   Gábor Csárdi       7M R Package Installation fro...
#>  10  69 meltr       1.0.0   Duncan Garmonsway 10M Read Non-Rectangular Text ...
advanced_search("Author: Hester AND Author: Wickham")
#> - "advanced search" ---------------------------- 19 packages in 0.001 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi       6M Find Tools Needed to Build...
#>   2  94 mockery     0.4.3   Hadley Wickham     4M Mocking Library for R        
#>   3  85 memoise     2.0.1   Winston Chang      7M 'Memoisation' of Functions   
#>   4  80 xml2        1.3.3   Hadley Wickham     7M Parse XML                    
#>   5  76 sessioninfo 1.2.2   Gábor Csárdi       6M R Session Information        
#>   6  75 devtools    2.4.3   Jennifer Bryan     7M Tools to Make Developing R...
#>   7  73 odbc        1.3.3   Hadley Wickham     7M Connect to ODBC Compatible...
#>   8  70 pkgload     1.2.4   Lionel Henry       7M Simulate Package Installat...
#>   9  70 remotes     2.4.2   Gábor Csárdi       7M R Package Installation fro...
#>  10  69 meltr       1.0.0   Duncan Garmonsway 10M Read Non-Rectangular Text ...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")
#> - "advanced search" ---------------------------- 27 packages in 0.003 seconds -
#>   #     package        version by                @ title                       
#>   1 100 knitrBootstrap 1.0.2   Jim Hester       4y 'knitr' Bootstrap Framework 
#>   2  93 types          1.0.0   Jim Hester       6y Type Annotations            
#>   3  86 primerTree     1.0.6   Matt Cannon      2M Visually Assessing the Sp...
#>   4  80 gmailr         1.0.1   Jennifer Bryan   7M Access the 'Gmail' 'RESTf...
#>   5  80 pak            0.3.0   Gábor Csárdi     2M Another Approach to Packa...
#>   6  80 dplyr.teradata 0.4.1   Koji Makiyama    2y A 'Teradata' Backend for ...
#>   7  77 decor          1.0.1   Romain François  7M Retrieve Code Decorations   
#>   8  72 rex            1.2.1   Kevin Ushey      7M Friendly Regular Expressions
#>   9  67 bench          1.1.2   Davis Vaughan    7M High Precision Timing of ...
#>  10  65 spelling       2.2     Jeroen Ooms      2y Tools for Spell Checking ...

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")
#> - "advanced search" ---------------------------- 19 packages in 0.005 seconds -
#>   #     package     version by               @ title                           
#>   1 100 mockery     0.4.3   Hadley Wickham  4M Mocking Library for R           
#>   2  95 pkgbuild    1.3.1   Gábor Csárdi    6M Find Tools Needed to Build R ...
#>   3  90 xml2        1.3.3   Hadley Wickham  7M Parse XML                       
#>   4  90 odbc        1.3.3   Hadley Wickham  7M Connect to ODBC Compatible Da...
#>   5  80 memoise     2.0.1   Winston Chang   7M 'Memoisation' of Functions      
#>   6  72 pkgload     1.2.4   Lionel Henry    7M Simulate Package Installation...
#>   7  72 devtools    2.4.3   Jennifer Bryan  7M Tools to Make Developing R Pa...
#>   8  72 sessioninfo 1.2.2   Gábor Csárdi    6M R Session Information           
#>   9  69 remotes     2.4.2   Gábor Csárdi    7M R Package Installation from R...
#>  10  67 fs          1.5.2   Gábor Csárdi    6M Cross-Platform File System Op...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")
#> - "advanced search" ---------------------------- 27 packages in 0.003 seconds -
#>   #     package        version by                @ title                       
#>   1 100 knitrBootstrap 1.0.2   Jim Hester       4y 'knitr' Bootstrap Framework 
#>   2  93 types          1.0.0   Jim Hester       6y Type Annotations            
#>   3  86 primerTree     1.0.6   Matt Cannon      2M Visually Assessing the Sp...
#>   4  80 gmailr         1.0.1   Jennifer Bryan   7M Access the 'Gmail' 'RESTf...
#>   5  80 pak            0.3.0   Gábor Csárdi     2M Another Approach to Packa...
#>   6  80 dplyr.teradata 0.4.1   Koji Makiyama    2y A 'Teradata' Backend for ...
#>   7  77 decor          1.0.1   Romain François  7M Retrieve Code Decorations   
#>   8  72 rex            1.2.1   Kevin Ushey      7M Friendly Regular Expressions
#>   9  67 bench          1.1.2   Davis Vaughan    7M High Precision Timing of ...
#>  10  65 spelling       2.2     Jeroen Ooms      2y Tools for Spell Checking ...

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")
#> - "advanced search" --------------------------- 389 packages in 0.019 seconds -
#>   #     package           version by                         @ title           
#>   1 100 phd               0.2     Jesse Hemerik             5M Permutation T...
#>   2  92 confSAM           0.2     Jesse Hemerik             4y Estimates and...
#>   3  81 permutes          2.5     Cesko C. Voeten           2d Permutation T...
#>   4  76 BurStMisc         1.1     Pat Burns                 6y Burns Statist...
#>   5  74 fdrci             2.3     Joshua Millstein         10d Permutation-B...
#>   6  74 USP               0.1.2   Thomas B. Berrett         1y U-Statistic P...
#>   7  74 resample          0.6     Tim Hesterberg            4d Resampling Fu...
#>   8  73 packager          1.14.0  Andreas Dominik Cullmann  6M Create, Build...
#>   9  72 MultSurvTests     0.2     Lukas Arnroth             1y Permutation T...
#>  10  70 phyloseqGraphTest 0.1.0   Julia Fukuyama            2y Graph-Based P...

# Packages that have a certain field
advanced_search("_exists_" = "URL")
#> - "advanced search" ------------------------- 10365 packages in 0.002 seconds -
#>   #     package     version by                    @ title                      
#>   1 100 DeducerText 0.1.2   Ian Fellows          8y Deducer GUI for Text Data  
#>   2 100 hillmakeR   0.2     David Gilinson       8y Perform occupancy analysis 
#>   3 100 rDotNet     0.9.1   Jonathan Shore       5y Low-Level Interface to t...
#>   4 100 epiphy      0.3.4   Christophe Gigot     4y Analysis of Plant Diseas...
#>   5 100 ROP         1.0     Jean-Michel Nguyen   4y Regression Optimized: Nu...
#>   6 100 rareGE      0.1     Han Chen             8y Testing Gene-Environment...
#>   7 100 extraTrees  1.0.5   Jaak Simm            7y Extremely Randomized Tre...
#>   8 100 spgrass6    0.8.9   Roger Bivand         6y Interface Between GRASS ...
#>   9 100 gamlss.nl   4.1.0   Mikis Stasinopoulos 10y Fitting non linear param...
#>  10 100 ffbase      0.13.3  Edwin de Jonge       1y Basic Statistical Functi...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")
#> - "advanced search" -------------------------- 9408 packages in 0.001 seconds -
#>   #     package                 version    by                 @ title          
#>   1 100 DnE                     2.1.0      Junyao Chen       8y Distribution...
#>   2 100 opentraj                1.0        Thalles Silva     8y Tools for Cr...
#>   3 100 empiricalFDR.DESeq2     1.0.3      Mikhail V. Matz   7y Simulation-B...
#>   4 100 breakage                1.1.1      Matthew Caldwell  8y SICM pipette...
#>   5 100 CRAC                    1.0        Jiayi Liu         8y Cosmology R ...
#>   6 100 mlica2                  2.1        Thomas Nelson    10y Independent ...
#>   7 100 FastLZeroSpikeInference 2018.12.10 Sean Jewell       3y Fast Nonconv...
#>   8 100 NoiseFiltersR           0.1.0      Julian Luengo     6y Label Noise ...
#>   9 100 SEMrushR                0.1.0      Remi Bacha        4y R Interface ...
#>  10 100 GSAgm                   1.0        Rama Raghavan     8y Gene Set Ana...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)
#> - "advanced search" -------------------------- 9408 packages in 0.001 seconds -
#>   #     package                 version    by                 @ title          
#>   1 100 DnE                     2.1.0      Junyao Chen       8y Distribution...
#>   2 100 opentraj                1.0        Thalles Silva     8y Tools for Cr...
#>   3 100 empiricalFDR.DESeq2     1.0.3      Mikhail V. Matz   7y Simulation-B...
#>   4 100 breakage                1.1.1      Matthew Caldwell  8y SICM pipette...
#>   5 100 CRAC                    1.0        Jiayi Liu         8y Cosmology R ...
#>   6 100 mlica2                  2.1        Thomas Nelson    10y Independent ...
#>   7 100 FastLZeroSpikeInference 2018.12.10 Sean Jewell       3y Fast Nonconv...
#>   8 100 NoiseFiltersR           0.1.0      Julian Luengo     6y Label Noise ...
#>   9 100 SEMrushR                0.1.0      Remi Bacha        4y R Interface ...
#>  10 100 GSAgm                   1.0        Rama Raghavan     8y Gene Set Ana...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")
#> - "advanced search" --------------------------- 240 packages in 0.003 seconds -
#>   #     package     version  by                           @ title              
#>   1 100 rDotNet     0.9.1    Jonathan Shore              5y Low-Level Interf...
#>   2 100 RPANDA      2.0      Hélène Morlon               7M Phylogenetic ANa...
#>   3 100 gdalUtils   2.0.3.2  Jonathan Asher Greenberg    2y Wrappers for the...
#>   4 100 glmdm       2.60     Jeff Gill                   9y R Code for Simul...
#>   5 100 sourceR     1.1.0    Poppy Miller                2y Fits a Non-Param...
#>   6 100 TauP.R      1.5      Jake Anderson               4y Earthquake Trave...
#>   7 100 prinsimp    0.8.8    Davor Cubranic              9y Finding and plot...
#>   8 100 Sample.Size 1.0      Wei Jiang<wjiang@kumc.edu>  9y Sample size calc...
#>   9 100 plink       1.5.1    Jonathan P. Weeks           5y IRT Separate Cal...
#>  10 100 RTriangle   1.6.0.10 David C. Sterratt           4y Triangle - A 2D ...

# Fuzzy search
advanced_search(Author = "Johnathan~1")
#> - "advanced search" --------------------------- 241 packages in 0.005 seconds -
#>   #     package    version by                           @ title                
#>   1 100 UScancer   0.1.2   Jonathan Lee                8y Create US cancer d...
#>   2 100 mlmi       1.1.1   Jonathan Bartlett           7M Maximum Likelihood...
#>   3  99 lda        1.4.2   Jonathan Chang              7y Collapsed Gibbs Sa...
#>   4  98 chords     0.95.4  Jonathan Rosenblatt         5y Estimation in Resp...
#>   5  98 bootImpute 1.2.0   Jonathan Bartlett           1y Bootstrap Inferenc...
#>   6  96 handyplots 1.1.3   Jonathan Schwartz           3y Handy Plots          
#>   7  96 UnitCircle 0.1.3   Jonathan Berrisch           4y Check if Roots of ...
#>   8  95 RSeed      0.1.60  Claus Jonathan Fritzemeier  6y Borenstein Analysis  
#>   9  94 rSHAPE     0.3.2   Jonathan Dench              3y Simulated Haploid ...
#>  10  92 Oarray     1.4.9   Robin K. S. Hankin          4y Arrays with Arbitr...