Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...

Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json

A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10000 milliseconds

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10001 milliseconds
advanced_search("Author: Hester AND Author: Wickham")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10001 milliseconds

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10001 milliseconds

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10001 milliseconds

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10001 milliseconds

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10001 milliseconds

# Packages that have a certain field
advanced_search("_exists_" = "URL")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10000 milliseconds

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10001 milliseconds

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10000 milliseconds

# Regular expressions
advanced_search(Author = "/Joh?nathan/")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10001 milliseconds

# Fuzzy search
advanced_search(Author = "Johnathan~1")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10001 milliseconds