Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...
Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json
A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")

#> - "advanced search" ---------------------------- 39 packages in 0.001 seconds -
#>   #     package        version   by         @ title                            
#>   1 100 distances      0.1.8     ORPHANED  1y Tools for Distance Metrics       
#>   2 100 mvcwt          1.3.1     ORPHANED  2y Wavelet Analysis of Multiple T...
#>   3 100 Funclustering  1.0.2     ORPHANED  2y Functional Data Clustering       
#>   4  96 SDMTools       1.1.221.2 ORPHANED 10M Species Distribution Modelling...
#>   5  96 NanoStringNorm 1.2.1.1   ORPHANED 28d Normalize NanoString miRNA and...
#>   6  96 datadr         0.8.6.1   ORPHANED  2y Divide and Recombine for Large...
#>   7  96 nhlscrapr      1.8.1     ORPHANED  4y Compiling the NHL Real Time Sc...
#>   8  86 ggmap          3.0.0     ORPHANED  2y Spatial Visualization with ggp...
#>   9  86 gWidgetstcltk  0.0.55.1  ORPHANED  5M Toolkit implementation of gWid...
#>  10  86 gWidgets       0.0.54.2  ORPHANED  5M gWidgets API for Building Tool...

# Packages that match a certain field
advanced_search(Author = "Hester AND NOT Wickham")

#> - "advanced search" ---------------------------- 25 packages in 0.004 seconds -
#>   #     package        version by              @ title                         
#>   1 100 gmailr         1.0.0   Jim Hester     1y Access the 'Gmail' 'RESTful...
#>   2  95 primerTree     1.0.5   Jim Hester     9M Visually Assessing the Spec...
#>   3  89 knitrBootstrap 1.0.2   Jim Hester     2y 'knitr' Bootstrap Framework   
#>   4  86 types          1.0.0   Jim Hester     4y Type Annotations              
#>   5  85 glue           1.4.2   Jim Hester     1M Interpreted String Literals   
#>   6  71 pak            0.1.2   Gábor Csárdi   2y Another Approach to Package...
#>   7  69 dplyr.teradata 0.3.2   Koji Makiyama  1y A 'Teradata' Backend for 'd...
#>   8  66 lintr          2.0.1   Jim Hester     8M A 'Linter' for R Code         
#>   9  65 decor          1.0.0   Jim Hester     3M Retrieve Code Decorations     
#>  10  63 bench          1.1.1   Jim Hester     9M High Precision Timing of R ...

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")

#> - "advanced search" --------------------------- 299 packages in 0.013 seconds -
#>   #     package           version by                         @ title           
#>   1 100 phd               0.1     Jesse Hemerik             1y Permutation T...
#>   2  97 confSAM           0.2     Jesse Hemerik             3y Estimates and...
#>   3  85 permutes          1.0     Cesko C. Voeten           1y Permutation T...
#>   4  80 BurStMisc         1.1     Pat Burns                 4y Burns Statist...
#>   5  77 fdrci             2.1     Joshua Millstein          4y Permutation-B...
#>   6  75 packager          1.5.0   Andreas Dominik Cullmann 11d Create, Build...
#>   7  75 phyloseqGraphTest 0.1.0   Julia Fukuyama            8M Graph-Based P...
#>   8  74 resample          0.4     Tim Hesterberg            5y Resampling Fu...
#>   9  63 shiftR            1.5     Andrey A Shabalin         2y Fast Enrichme...
#>  10  60 mindr             1.2.3   Peng Zhao                 7M Convert Files...

# Packages that have a certain field
advanced_search("_exists_" = "URL")

#> - "advanced search" -------------------------- 8738 packages in 0.001 seconds -
#>   #     package     version   by                   @ title                     
#>   1 100 imgw        0.3.0     Bartosz Czernecki   8M Polish Meteorological a...
#>   2 100 pathmox     0.2.0     Gaston Sanchez      7y Pathmox Approach of Seg...
#>   3 100 MFKnockoffs 0.9.1     Matteo Sesia        3y Model-Free Knockoff Fil...
#>   4 100 profilr     0.1.0     Paul Hendricks      5y Quickly Profile Data in R 
#>   5 100 infuser     0.2.8     Bart Smeets         3y A Very Basic Templating...
#>   6 100 theseus     0.1.0     Stephen Woloszynek  3y Analysis and Visualizat...
#>   7 100 TBRDist     1.0.2     Martin R. Smith    22d Rearrangement Distances...
#>   8 100 ldatuning   1.0.2     Nathan Chaney       6M Tuning of the Latent Di...
#>   9 100 SDMTools    1.1.221.2 ORPHANED           10M Species Distribution Mo...
#>  10 100 HiClimR     2.1.6     Hamada S. Badr      8M Hierarchical Climate Re...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")

#> - "advanced search" -------------------------- 9001 packages in 0.001 seconds -
#>   #     package      version by                               @ title          
#>   1 100 refund.shiny 0.3.0   Julia Wrobel                    4y Interactive ...
#>   2 100 TLdating     0.1.3   David Strebler                  4y Tools for Th...
#>   3 100 MRCE         2.1     Adam J. Rothman                 4y Multivariate...
#>   4 100 lift         0.0.2   Steven Hoornaert                5y Compute the ...
#>   5 100 argosfilter  0.63    Carla Freitas                   8y Argos locati...
#>   6 100 equateIRT    2.1.0   Michela Battauz                 2y IRT Equating...
#>   7 100 GLDEX        2.0.0.7 Steve Su                        8M Fitting Sing...
#>   8 100 ifs          0.1.5   S. M. Iacus                     5y Iterated Fun...
#>   9 100 orders       0.1.4   Carlos Alberto Cardozo Delgado 15d Sampling fro...
#>  10 100 extremefit   1.0.2   Kevin Jaunatre                  1y Estimation o...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)

#> - "advanced search" -------------------------- 9001 packages in 0.001 seconds -
#>   #     package      version by                               @ title          
#>   1 100 refund.shiny 0.3.0   Julia Wrobel                    4y Interactive ...
#>   2 100 TLdating     0.1.3   David Strebler                  4y Tools for Th...
#>   3 100 MRCE         2.1     Adam J. Rothman                 4y Multivariate...
#>   4 100 lift         0.0.2   Steven Hoornaert                5y Compute the ...
#>   5 100 argosfilter  0.63    Carla Freitas                   8y Argos locati...
#>   6 100 equateIRT    2.1.0   Michela Battauz                 2y IRT Equating...
#>   7 100 GLDEX        2.0.0.7 Steve Su                        8M Fitting Sing...
#>   8 100 ifs          0.1.5   S. M. Iacus                     5y Iterated Fun...
#>   9 100 orders       0.1.4   Carlos Alberto Cardozo Delgado 15d Sampling fro...
#>  10 100 extremefit   1.0.2   Kevin Jaunatre                  1y Estimation o...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")

#> - "advanced search" --------------------------- 202 packages in 0.001 seconds -
#>   #     package            version by                         @ title          
#>   1 100 selectiveInference 1.2.5   Rob Tibshirani            1y Tools for Po...
#>   2 100 knitr              1.30    Yihui Xie                17d A General-Pu...
#>   3 100 xtable             1.8.4   David Scott               1y Export Table...
#>   4 100 raster             3.3.13  Robert J. Hijmans         3M Geographic D...
#>   5 100 bridgesampling     1.0.0   Quentin F. Gronau         8M Bridge Sampl...
#>   6 100 tableone           0.12.0  Kazuki Yoshida            3M Create 'Tabl...
#>   7 100 fivethirtyeight    0.6.1   Albert Y. Kim             2M Data and Cod...
#>   8 100 gdalUtils          2.0.3.2 Jonathan Asher Greenberg  8M Wrappers for...
#>   9 100 tscount            1.4.3   Tobias Liboschik          1M Analysis of ...
#>  10 100 smcfcs             1.4.1   Jonathan Bartlett         6M Multiple Imp...

# Fuzzy search
advanced_search(Author = "Johnathan~1")

#> - "advanced search" --------------------------- 203 packages in 0.002 seconds -
#>   #     package       version by                                        @ title
#>   1 100 UScancer      0.1.2   Jonathan Lee                             6y Cr...
#>   2 100 mlmi          1.0.0   Jonathan Bartlett                        1y Ma...
#>   3  99 lda           1.4.2   Jonathan Chang                           5y Co...
#>   4  96 chords        0.95.4  Jonathan Rosenblatt                      4y Es...
#>   5  96 bootImpute    1.1.0   Jonathan Bartlett                        3M Bo...
#>   6  96 rSHAPE        0.3.2   Jonathan Dench                           1y Si...
#>   7  96 handyplots    1.1.3   Jonathan Schwartz                        2y Ha...
#>   8  96 UnitCircle    0.1.3   Jonathan Berrisch                        2y Ch...
#>   9  95 geomapdata    1.0.4   Jonathan M. Lees<jonathan.lees@unc.edu>  8y Da...
#>  10  95 spatial.tools 1.6.2   Jonathan Asher Greenberg                 8M R ...