Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...
Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json
A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")

#> - "advanced search" ---------------------------- 39 packages in 0.001 seconds -
#>   #     package        version   by         @ title                            
#>   1 100 mvcwt          1.3.1     ORPHANED  2y Wavelet Analysis of Multiple T...
#>   2 100 distances      0.1.8     ORPHANED  1y Tools for Distance Metrics       
#>   3 100 Funclustering  1.0.2     ORPHANED  2y Functional Data Clustering       
#>   4  96 datadr         0.8.6.1   ORPHANED  2y Divide and Recombine for Large...
#>   5  96 SDMTools       1.1.221.2 ORPHANED 10M Species Distribution Modelling...
#>   6  96 NanoStringNorm 1.2.1.1   ORPHANED 27d Normalize NanoString miRNA and...
#>   7  96 nhlscrapr      1.8.1     ORPHANED  4y Compiling the NHL Real Time Sc...
#>   8  86 gWidgets       0.0.54.2  ORPHANED  5M gWidgets API for Building Tool...
#>   9  86 clues          0.6.2.2   ORPHANED 10M Clustering Method Based on Local 
#>  10  86 rda            1.0.2.2.1 ORPHANED  2y Shrunken Centroids Regularized...

# Packages that match a certain field
advanced_search(Author = "Hester AND NOT Wickham")

#> - "advanced search" ---------------------------- 25 packages in 0.004 seconds -
#>   #     package        version by              @ title                         
#>   1 100 gmailr         1.0.0   Jim Hester     1y Access the 'Gmail' 'RESTful...
#>   2  95 primerTree     1.0.5   Jim Hester     9M Visually Assessing the Spec...
#>   3  89 knitrBootstrap 1.0.2   Jim Hester     2y 'knitr' Bootstrap Framework   
#>   4  86 types          1.0.0   Jim Hester     4y Type Annotations              
#>   5  85 glue           1.4.2   Jim Hester     1M Interpreted String Literals   
#>   6  71 pak            0.1.2   Gábor Csárdi   2y Another Approach to Package...
#>   7  69 dplyr.teradata 0.3.2   Koji Makiyama  1y A 'Teradata' Backend for 'd...
#>   8  66 lintr          2.0.1   Jim Hester     8M A 'Linter' for R Code         
#>   9  65 decor          1.0.0   Jim Hester     3M Retrieve Code Decorations     
#>  10  63 mockery        0.4.2   Jim Hester     1y Mocking Library for R         

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")

#> - "advanced search" --------------------------- 299 packages in 0.014 seconds -
#>   #     package           version by                         @ title           
#>   1 100 phd               0.1     Jesse Hemerik             1y Permutation T...
#>   2  97 confSAM           0.2     Jesse Hemerik             3y Estimates and...
#>   3  85 permutes          1.0     Cesko C. Voeten           1y Permutation T...
#>   4  79 BurStMisc         1.1     Pat Burns                 4y Burns Statist...
#>   5  77 fdrci             2.1     Joshua Millstein          4y Permutation-B...
#>   6  75 packager          1.5.0   Andreas Dominik Cullmann 10d Create, Build...
#>   7  75 phyloseqGraphTest 0.1.0   Julia Fukuyama            8M Graph-Based P...
#>   8  74 resample          0.4     Tim Hesterberg            5y Resampling Fu...
#>   9  63 shiftR            1.5     Andrey A Shabalin         2y Fast Enrichme...
#>  10  60 mindr             1.2.3   Peng Zhao                 7M Convert Files...

# Packages that have a certain field
advanced_search("_exists_" = "URL")

#> - "advanced search" -------------------------- 8738 packages in 0.001 seconds -
#>   #     package              version by                                      @
#>   1 100 tabplotd3            0.3.3   Edwin de Jonge<edwindjonge@gmail.com>  7y
#>   2 100 tsbridge             1.1     Guy J. Abel                            6y
#>   3 100 yearn                0.1.3   Brian O'Meara                          3y
#>   4 100 survAccuracyMeasures 1.2     Marshall Brown                         6y
#>   5 100 softermax            0.3.0   Brian Connelly                         3y
#>   6 100 wrspathrowData       1.0     Alex Zvoleff                           7y
#>   7 100 rFTRLProximal        1.0.0   Yachen Yan                             4y
#>   8 100 ubeR                 0.1.4   Andrew Collier                         4y
#>   9 100 GeneReg              1.1.2   Tao Huang                              8y
#>  10 100 pbdDMAT              0.5.1   Drew Schmidt                           2y

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")

#> - "advanced search" -------------------------- 9001 packages in 0.001 seconds -
#>   #     package          version by                          @ title           
#>   1 100 refund.wave      0.1     Adam Ciarleglio            6y Wavelet-Domai...
#>   2 100 MoonFinder       1.0.3   Lixin Cheng                1y Identify Moon...
#>   3 100 cudaBayesregData 0.3.11  Adelino Ferreira da Silva  8y Data sets for...
#>   4 100 expectreg        0.50    Fabian Otto-Sobotka        1y Expectile and...
#>   5 100 glmmBUGS         2.4.2   Patrick Brown              2y Generalised L...
#>   6 100 FUNLDA           1.1     Daniel Backenroth          3y Genomic Laten...
#>   7 100 RadioGx          0.0.2   Benjamin Haibe-Kains      10M Analysis of L...
#>   8 100 miniparquet      0.1.2   Hannes Mühleisen           1y Read Parquet ...
#>   9 100 ELMSurv          0.6     Hong Wang                  1y Extreme Learn...
#>  10 100 MExPosition      2.0.3   Cherise R. Chin Fatt       7y Multi-table E...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)

#> - "advanced search" ------------------------------ 9001 packages in 0 seconds -
#>   #     package          version by                          @ title           
#>   1 100 refund.wave      0.1     Adam Ciarleglio            6y Wavelet-Domai...
#>   2 100 MoonFinder       1.0.3   Lixin Cheng                1y Identify Moon...
#>   3 100 cudaBayesregData 0.3.11  Adelino Ferreira da Silva  8y Data sets for...
#>   4 100 expectreg        0.50    Fabian Otto-Sobotka        1y Expectile and...
#>   5 100 glmmBUGS         2.4.2   Patrick Brown              2y Generalised L...
#>   6 100 FUNLDA           1.1     Daniel Backenroth          3y Genomic Laten...
#>   7 100 RadioGx          0.0.2   Benjamin Haibe-Kains      10M Analysis of L...
#>   8 100 miniparquet      0.1.2   Hannes Mühleisen           1y Read Parquet ...
#>   9 100 ELMSurv          0.6     Hong Wang                  1y Extreme Learn...
#>  10 100 MExPosition      2.0.3   Cherise R. Chin Fatt       7y Multi-table E...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")

#> - "advanced search" --------------------------- 202 packages in 0.001 seconds -
#>   #     package         version by                         @ title             
#>   1 100 ubeR            0.1.4   Andrew Collier            4y Interface to th...
#>   2 100 crunchy         0.3.2   Greg Freedman Ellis       3M Shiny Apps on C...
#>   3 100 hergm           4.1.6   Michael Schweinberger     8M Hierarchical Ex...
#>   4 100 xtable          1.8.4   David Scott               1y Export Tables t...
#>   5 100 raster          3.3.13  Robert J. Hijmans         3M Geographic Data...
#>   6 100 knitr           1.30    Yihui Xie                16d A General-Purpo...
#>   7 100 bridgesampling  1.0.0   Quentin F. Gronau         8M Bridge Sampling...
#>   8 100 gdalUtils       2.0.3.2 Jonathan Asher Greenberg  8M Wrappers for th...
#>   9 100 fivethirtyeight 0.6.1   Albert Y. Kim             2M Data and Code B...
#>  10 100 tableone        0.12.0  Kazuki Yoshida            2M Create 'Table 1...

# Fuzzy search
advanced_search(Author = "Johnathan~1")

#> - "advanced search" --------------------------- 203 packages in 0.002 seconds -
#>   #     package       version by                                        @ title
#>   1 100 UScancer      0.1.2   Jonathan Lee                             6y Cr...
#>   2 100 mlmi          1.0.0   Jonathan Bartlett                        1y Ma...
#>   3  99 lda           1.4.2   Jonathan Chang                           5y Co...
#>   4  96 bootImpute    1.1.0   Jonathan Bartlett                        3M Bo...
#>   5  96 chords        0.95.4  Jonathan Rosenblatt                      4y Es...
#>   6  96 rSHAPE        0.3.2   Jonathan Dench                           1y Si...
#>   7  95 handyplots    1.1.3   Jonathan Schwartz                        2y Ha...
#>   8  95 UnitCircle    0.1.3   Jonathan Berrisch                        2y Ch...
#>   9  95 spatial.tools 1.6.2   Jonathan Asher Greenberg                 8M R ...
#>  10  95 geomapdata    1.0.4   Jonathan M. Lees<jonathan.lees@unc.edu>  8y Da...