Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...
Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json
A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")

#> - "advanced search" ---------------------------- 37 packages in 0.001 seconds -
#>   #     package        version   by         @ title                            
#>   1 100 Funclustering  1.0.2     ORPHANED  2y Functional Data Clustering       
#>   2 100 distances      0.1.8     ORPHANED  2y Tools for Distance Metrics       
#>   3 100 mvcwt          1.3.1     ORPHANED  2y Wavelet Analysis of Multiple T...
#>   4  96 NanoStringNorm 1.2.1.1   ORPHANED  7M Normalize NanoString miRNA and...
#>   5  96 SDMTools       1.1.221.2 ORPHANED  1y Species Distribution Modelling...
#>   6  96 datadr         0.8.6.1   ORPHANED  3y Divide and Recombine for Large...
#>   7  96 nhlscrapr      1.8.1     ORPHANED  4y Compiling the NHL Real Time Sc...
#>   8  86 ggmap          3.0.0     ORPHANED  2y Spatial Visualization with ggp...
#>   9  86 gWidgetstcltk  0.0.55.1  ORPHANED  1y Toolkit implementation of gWid...
#>  10  86 gWidgets       0.0.54.2  ORPHANED  1y gWidgets API for Building Tool...

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")

#> - "advanced search" ---------------------------- 17 packages in 0.001 seconds -
#>   #     package     version by             @ title                             
#>   1 100 pkgbuild    1.2.0   Jim Hester    4M Find Tools Needed to Build R Pa...
#>   2  86 devtools    2.4.0   Jim Hester    4d Tools to Make Developing R Pack...
#>   3  80 sessioninfo 1.1.1   Gábor Csárdi  2y R Session Information             
#>   4  70 memoise     2.0.0   Jim Hester    2M Memoisation of Functions          
#>   5  68 xml2        1.3.2   Jim Hester    1y Parse XML                         
#>   6  64 odbc        1.3.2   Jim Hester    8d Connect to ODBC Compatible Data...
#>   7  62 withr       2.4.1   Jim Hester    2M Run Code 'With' Temporarily Mod...
#>   8  61 pkgload     1.2.1   Jim Hester    5d Simulate Package Installation a...
#>   9  60 fs          1.5.0   Jim Hester    8M Cross-Platform File System Oper...
#>  10  60 remotes     2.3.0   Jim Hester   10d R Package Installation from Rem...

advanced_search("Author: Hester AND Author: Wickham")

#> - "advanced search" ---------------------------- 17 packages in 0.001 seconds -
#>   #     package     version by             @ title                             
#>   1 100 pkgbuild    1.2.0   Jim Hester    4M Find Tools Needed to Build R Pa...
#>   2  86 devtools    2.4.0   Jim Hester    4d Tools to Make Developing R Pack...
#>   3  80 sessioninfo 1.1.1   Gábor Csárdi  2y R Session Information             
#>   4  70 memoise     2.0.0   Jim Hester    2M Memoisation of Functions          
#>   5  68 xml2        1.3.2   Jim Hester    1y Parse XML                         
#>   6  64 odbc        1.3.2   Jim Hester    8d Connect to ODBC Compatible Data...
#>   7  62 withr       2.4.1   Jim Hester    2M Run Code 'With' Temporarily Mod...
#>   8  61 pkgload     1.2.1   Jim Hester    5d Simulate Package Installation a...
#>   9  60 fs          1.5.0   Jim Hester    8M Cross-Platform File System Oper...
#>  10  60 remotes     2.3.0   Jim Hester   10d R Package Installation from Rem...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")

#> - "advanced search" ---------------------------- 26 packages in 0.004 seconds -
#>   #     package        version by              @ title                         
#>   1 100 gmailr         1.0.0   Jim Hester     2y Access the 'Gmail' 'RESTful...
#>   2  93 primerTree     1.0.5   Jim Hester     1y Visually Assessing the Spec...
#>   3  89 knitrBootstrap 1.0.2   Jim Hester     3y 'knitr' Bootstrap Framework   
#>   4  86 types          1.0.0   Jim Hester     4y Type Annotations              
#>   5  84 glue           1.4.2   Jim Hester     8M Interpreted String Literals   
#>   6  72 pak            0.1.2.1 Gábor Csárdi   5M Another Approach to Package...
#>   7  70 dplyr.teradata 0.4.1   Koji Makiyama  5M A 'Teradata' Backend for 'd...
#>   8  67 lintr          2.0.1   Jim Hester     1y A 'Linter' for R Code         
#>   9  65 decor          1.0.0   Jim Hester    10M Retrieve Code Decorations     
#>  10  64 bench          1.1.1   Jim Hester     1y High Precision Timing of R ...

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")

#> - "advanced search" ---------------------------- 17 packages in 0.005 seconds -
#>   #     package     version by             @ title                             
#>   1 100 pkgbuild    1.2.0   Jim Hester    4M Find Tools Needed to Build R Pa...
#>   2  86 devtools    2.4.0   Jim Hester    4d Tools to Make Developing R Pack...
#>   3  80 sessioninfo 1.1.1   Gábor Csárdi  2y R Session Information             
#>   4  70 memoise     2.0.0   Jim Hester    2M Memoisation of Functions          
#>   5  68 xml2        1.3.2   Jim Hester    1y Parse XML                         
#>   6  67 pkgload     1.2.1   Jim Hester    5d Simulate Package Installation a...
#>   7  65 odbc        1.3.2   Jim Hester    8d Connect to ODBC Compatible Data...
#>   8  64 fs          1.5.0   Jim Hester    8M Cross-Platform File System Oper...
#>   9  62 remotes     2.3.0   Jim Hester   10d R Package Installation from Rem...
#>  10  62 withr       2.4.1   Jim Hester    2M Run Code 'With' Temporarily Mod...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")

#> - "advanced search" ---------------------------- 26 packages in 0.004 seconds -
#>   #     package        version by              @ title                         
#>   1 100 gmailr         1.0.0   Jim Hester     2y Access the 'Gmail' 'RESTful...
#>   2  93 primerTree     1.0.5   Jim Hester     1y Visually Assessing the Spec...
#>   3  89 knitrBootstrap 1.0.2   Jim Hester     3y 'knitr' Bootstrap Framework   
#>   4  86 types          1.0.0   Jim Hester     4y Type Annotations              
#>   5  84 glue           1.4.2   Jim Hester     8M Interpreted String Literals   
#>   6  72 pak            0.1.2.1 Gábor Csárdi   5M Another Approach to Package...
#>   7  70 dplyr.teradata 0.4.1   Koji Makiyama  5M A 'Teradata' Backend for 'd...
#>   8  67 lintr          2.0.1   Jim Hester     1y A 'Linter' for R Code         
#>   9  65 decor          1.0.0   Jim Hester    10M Retrieve Code Decorations     
#>  10  64 bench          1.1.1   Jim Hester     1y High Precision Timing of R ...

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")

#> - "advanced search" ---------------------------- 340 packages in 0.02 seconds -
#>   #     package           version by                         @ title           
#>   1 100 phd               0.1     Jesse Hemerik             2y Permutation T...
#>   2  97 confSAM           0.2     Jesse Hemerik             3y Estimates and...
#>   3  86 permutes          2.0.1   Cesko C. Voeten           2M Permutation T...
#>   4  80 BurStMisc         1.1     Pat Burns                 5y Burns Statist...
#>   5  78 fdrci             2.1     Joshua Millstein          4y Permutation-B...
#>   6  77 USP               0.1.2   Thomas B. Berrett         2M U-Statistic P...
#>   7  77 packager          1.11.1  Andreas Dominik Cullmann 21d Create, Build...
#>   8  76 phyloseqGraphTest 0.1.0   Julia Fukuyama            1y Graph-Based P...
#>   9  75 resample          0.4     Tim Hesterberg            6y Resampling Fu...
#>  10  69 IMEC              0.2.0   Maximilian Maier          5M Ising Model o...

# Packages that have a certain field
advanced_search("_exists_" = "URL")

#> - "advanced search" -------------------------- 9557 packages in 0.001 seconds -
#>   #     package     version by                      @ title                    
#>   1 100 exploratory 0.1.53  Jin Kim               20d Conduct Exploratory An...
#>   2 100 jsonstat    0.0.2   Kim Seonghyun          3y Interface to 'JSON-stat' 
#>   3 100 glow        0.10.1  Travers Ching          7M Make Plots that Glow     
#>   4 100 GIFTr       0.1.0   Omar I. Elashkar       1y GIFT Questions Format ...
#>   5 100 path.chain  0.2.0   Krzysztof Joachimiak   7M Concise Structure for ...
#>   6 100 cheatR      1.2.1   Mattan S. Ben-Shachar  1y Catch Cheaters           
#>   7 100 readtext    0.80    Kenneth Benoit         7M Import and Handling fo...
#>   8 100 stm         1.3.6   Brandon Stewart        7M Estimation of the Stru...
#>   9 100 eventdataR  0.2.0   Gert Janssenswillen    3y Event Data Repository    
#>  10 100 MVA         1.0.7   Torsten Hothorn       30d An Introduction to App...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")

#> - "advanced search" -------------------------- 9504 packages in 0.001 seconds -
#>   #     package    version by                         @ title                  
#>   1 100 ADCT       0.1.0   Yalin Zhu                 4y Adaptive Design in C...
#>   2 100 ARHT       0.1.0   Haoran Li                 3y Adaptable Regularize...
#>   3 100 RDFTensor  1.3     Abdelmoneim Amer Desouki  3M Different Tensor Fac...
#>   4 100 shinypanel 0.1.4   Alex Pickering            3M Shiny Control Panel    
#>   5 100 kstMatrix  0.1.2   Cord Hockemeyer           2y Basic Functions in K...
#>   6 100 HETOP      0.2.6   J.R. Lockwood             2y MLE and Bayesian Est...
#>   7 100 simgof     1.0.2   Wolfgang Rolke            2M Simultaneous Goodnes...
#>   8 100 spsurv     1.0.0   Renato Panaro             1y Bernstein Polynomial...
#>   9 100 miselect   0.9.0   Alexander Rix             1y Variable Selection f...
#>  10 100 PCAmixdata 3.1     Marie Chavent             3y Multivariate Analysi...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)

#> - "advanced search" ------------------------------ 9504 packages in 0 seconds -
#>   #     package    version by                         @ title                  
#>   1 100 ADCT       0.1.0   Yalin Zhu                 4y Adaptive Design in C...
#>   2 100 ARHT       0.1.0   Haoran Li                 3y Adaptable Regularize...
#>   3 100 RDFTensor  1.3     Abdelmoneim Amer Desouki  3M Different Tensor Fac...
#>   4 100 shinypanel 0.1.4   Alex Pickering            3M Shiny Control Panel    
#>   5 100 kstMatrix  0.1.2   Cord Hockemeyer           2y Basic Functions in K...
#>   6 100 HETOP      0.2.6   J.R. Lockwood             2y MLE and Bayesian Est...
#>   7 100 simgof     1.0.2   Wolfgang Rolke            2M Simultaneous Goodnes...
#>   8 100 spsurv     1.0.0   Renato Panaro             1y Bernstein Polynomial...
#>   9 100 miselect   0.9.0   Alexander Rix             1y Variable Selection f...
#>  10 100 PCAmixdata 3.1     Marie Chavent             3y Multivariate Analysi...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")

#> - "advanced search" --------------------------- 220 packages in 0.001 seconds -
#>   #     package        version by                         @ title              
#>   1 100 anthro         0.9.4   Dirk Schumacher           5M Computation of t...
#>   2 100 rDotNet        0.9.1   Jonathan Shore            4y Low-Level Interf...
#>   3 100 ubeR           0.1.4   Andrew Collier            4y Interface to the...
#>   4 100 knitr          1.31    Yihui Xie                 2M A General-Purpos...
#>   5 100 raster         3.4.5   Robert J. Hijmans         5M Geographic Data ...
#>   6 100 xtable         1.8.4   David Scott               2y Export Tables to...
#>   7 100 bridgesampling 1.0.0   Quentin F. Gronau         1y Bridge Sampling ...
#>   8 100 tableone       0.12.0  Kazuki Yoshida            9M Create 'Table 1'...
#>   9 100 gdalUtils      2.0.3.2 Jonathan Asher Greenberg  1y Wrappers for the...
#>  10 100 rmcorr         0.4.3   Laura R. Marusich        10d Repeated Measure...

# Fuzzy search
advanced_search(Author = "Johnathan~1")

#> - "advanced search" --------------------------- 221 packages in 0.001 seconds -
#>   #     package       version by                                        @ title
#>   1 100 lda           1.4.2   Jonathan Chang                           5y Co...
#>   2 100 UScancer      0.1.2   Jonathan Lee                             7y Cr...
#>   3 100 mlmi          1.0.0   Jonathan Bartlett                        2y Ma...
#>   4  96 handyplots    1.1.3   Jonathan Schwartz                        2y Ha...
#>   5  96 UnitCircle    0.1.3   Jonathan Berrisch                        3y Ch...
#>   6  95 RSeed         0.1.60  Claus Jonathan Fritzemeier               5y Bo...
#>   7  95 rSHAPE        0.3.2   Jonathan Dench                           2y Si...
#>   8  95 geomapdata    1.0.4   Jonathan M. Lees<jonathan.lees@unc.edu>  8y Da...
#>   9  95 spatial.tools 1.6.2   Jonathan Asher Greenberg                 1y R ...
#>  10  94 bootImpute    1.2.0   Jonathan Bartlett                       10d Bo...