Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...

Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json

A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")
#> - "advanced search" ----------------------------- 3 packages in 0.001 seconds -
#>   #     package   version  by         @ title                                  
#>  1  100 udunits2  0.13.2.1 ORPHANED  1y Udunits-2 Bindings for R               
#>  2  100 fastclime 1.4.1.1  ORPHANED  3M A Fast Solver for Parameterized LP P...
#>  3   98 shipunov  1.17.1   ORPHANED  6M Miscellaneous Functions from Alexey ...

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")
#> - "advanced search" ---------------------------- 19 packages in 0.001 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 mockery     0.4.3   Hadley Wickham     1y Mocking Library for R        
#>   2  97 pkgbuild    1.4.2   Gábor Csárdi       1M Find Tools Needed to Build...
#>   3  88 memoise     2.0.1   Winston Chang      2y 'Memoisation' of Functions   
#>   4  85 xml2        1.3.5   Hadley Wickham    26d Parse XML                    
#>   5  81 devtools    2.4.5   Jennifer Bryan    10M Tools to Make Developing R...
#>   6  81 odbc        1.3.5   Hadley Wickham     1M Connect to ODBC Compatible...
#>   7  80 sessioninfo 1.2.2   Gábor Csárdi       2y R Session Information        
#>   8  74 remotes     2.4.2.1 Gábor Csárdi      14d R Package Installation fro...
#>   9  73 meltr       1.0.1   Duncan Garmonsway 11M Read Non-Rectangular Text ...
#>  10  72 pkgload     1.3.2.1 Lionel Henry      24d Simulate Package Installat...
advanced_search("Author: Hester AND Author: Wickham")
#> - "advanced search" ---------------------------- 19 packages in 0.001 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 mockery     0.4.3   Hadley Wickham     1y Mocking Library for R        
#>   2  97 pkgbuild    1.4.2   Gábor Csárdi       1M Find Tools Needed to Build...
#>   3  88 memoise     2.0.1   Winston Chang      2y 'Memoisation' of Functions   
#>   4  85 xml2        1.3.5   Hadley Wickham    26d Parse XML                    
#>   5  81 devtools    2.4.5   Jennifer Bryan    10M Tools to Make Developing R...
#>   6  81 odbc        1.3.5   Hadley Wickham     1M Connect to ODBC Compatible...
#>   7  80 sessioninfo 1.2.2   Gábor Csárdi       2y R Session Information        
#>   8  74 remotes     2.4.2.1 Gábor Csárdi      14d R Package Installation fro...
#>   9  73 meltr       1.0.1   Duncan Garmonsway 11M Read Non-Rectangular Text ...
#>  10  72 pkgload     1.3.2.1 Lionel Henry      24d Simulate Package Installat...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")
#> - "advanced search" ---------------------------- 27 packages in 0.006 seconds -
#>   #     package        version by                        @ title               
#>   1 100 knitrBootstrap 1.0.2   Jim Hester               5y 'knitr' Bootstrap...
#>   2  96 types          1.0.0   Jim Hester               7y Type Annotations    
#>   3  84 primerTree     1.0.6   Matt Cannon              1y Visually Assessin...
#>   4  81 pak            0.5.1   Gábor Csárdi             3M Another Approach ...
#>   5  76 gmailr         2.0.0   Jennifer Bryan           1M Access the 'Gmail...
#>   6  72 rex            1.2.1   Kevin Ushey              2y Friendly Regular ...
#>   7  67 roger          1.4.0   Vincent Goulet           6M Automated Grading...
#>   8  65 spelling       2.2.1   Jeroen Ooms              4M Tools for Spell C...
#>   9  63 available      1.1.0   Gábor Csárdi             1y Check if the Titl...
#>  10  62 tidycode       0.1.1   Lucy D'Agostino McGowan  4y Analyze Lines of ...

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")
#> - "advanced search" ----------------------------- 19 packages in 0.01 seconds -
#>   #     package     version by               @ title                           
#>   1 100 mockery     0.4.3   Hadley Wickham  1y Mocking Library for R           
#>   2  92 odbc        1.3.5   Hadley Wickham  1M Connect to ODBC Compatible Da...
#>   3  92 xml2        1.3.5   Hadley Wickham 26d Parse XML                       
#>   4  88 pkgbuild    1.4.2   Gábor Csárdi    1M Find Tools Needed to Build R ...
#>   5  78 memoise     2.0.1   Winston Chang   2y 'Memoisation' of Functions      
#>   6  77 devtools    2.4.5   Jennifer Bryan 10M Tools to Make Developing R Pa...
#>   7  72 sessioninfo 1.2.2   Gábor Csárdi    2y R Session Information           
#>   8  69 remotes     2.4.2.1 Gábor Csárdi   14d R Package Installation from R...
#>   9  68 pkgload     1.3.2.1 Lionel Henry   24d Simulate Package Installation...
#>  10  66 fs          1.6.3   Gábor Csárdi   12d Cross-Platform File System Op...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")
#> - "advanced search" ---------------------------- 27 packages in 0.004 seconds -
#>   #     package        version by                        @ title               
#>   1 100 knitrBootstrap 1.0.2   Jim Hester               5y 'knitr' Bootstrap...
#>   2  96 types          1.0.0   Jim Hester               7y Type Annotations    
#>   3  84 primerTree     1.0.6   Matt Cannon              1y Visually Assessin...
#>   4  81 pak            0.5.1   Gábor Csárdi             3M Another Approach ...
#>   5  76 gmailr         2.0.0   Jennifer Bryan           1M Access the 'Gmail...
#>   6  72 rex            1.2.1   Kevin Ushey              2y Friendly Regular ...
#>   7  67 roger          1.4.0   Vincent Goulet           6M Automated Grading...
#>   8  65 spelling       2.2.1   Jeroen Ooms              4M Tools for Spell C...
#>   9  63 available      1.1.0   Gábor Csárdi             1y Check if the Titl...
#>  10  62 tidycode       0.1.1   Lucy D'Agostino McGowan  4y Analyze Lines of ...

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")
#> - "advanced search" ---------------------------- 447 packages in 0.02 seconds -
#>   #     package       version by                         @ title               
#>   1 100 phd           0.2     Jesse Hemerik             2y Permutation Testi...
#>   2  93 confSAM       0.2     Jesse Hemerik             5y Estimates and Bou...
#>   3  86 permutes      2.6     Cesko C. Voeten           1y Permutation Tests...
#>   4  79 BurStMisc     1.1     Pat Burns                 7y Burns Statistics ...
#>   5  77 USP           0.1.2   Thomas B. Berrett         3y U-Statistic Permu...
#>   6  76 packager      1.15.1  Andreas Dominik Cullmann 25d Create, Build and...
#>   7  76 resample      0.6     Tim Hesterberg            1y Resampling Functions
#>   8  73 fdrci         2.4     Joshua Millstein         10M Permutation-Based...
#>   9  68 MultSurvTests 0.2     Lukas Arnroth             2y Permutation Tests...
#>  10  67 IMEC          0.2.0   Maximilian Maier          3y Ising Model of Ex...

# Packages that have a certain field
advanced_search("_exists_" = "URL")
#> - "advanced search" ------------------------- 11426 packages in 0.002 seconds -
#>   #     package                    version by                     @ title      
#>   1 100 opentimsr                  1.0.13  Michał Piotr Startek  1y An Open-...
#>   2 100 polars                     0.7.0   Tatsuya Shima        15d Lightnin...
#>   3 100 RcmdrPlugin.sutteForecastR 1.0.0   Ansari Saleh Ahmar    6y 'Rcmdr' ...
#>   4 100 seplyr                     1.0.4   John Mount            2y Improved...
#>   5 100 sgat                       0.9     Matias Poullain       2y Extract ...
#>   6 100 neotoma                    1.7.4   Simon J. Goring       5y Access t...
#>   7 100 sgmcmc                     0.2.5   Jack Baker            4y Stochast...
#>   8 100 randNames                  0.2.3   Karthik Ram           7y Package ...
#>   9 100 rIP                        1.2.0   Ryan Kennedy          4y Detects ...
#>  10 100 rangemap                   0.1.18  Marlon E. Cobos       2y Simple T...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")
#> - "advanced search" -------------------------- 9287 packages in 0.003 seconds -
#>   #     package     version by                   @ title                       
#>   1 100 MHMM        1.0.0   Matthieu Marbac     3y Finite Mixture of Hidden ...
#>   2 100 SelvarMix   1.2.1   Mohammed Sedki      6y Regularization for Variab...
#>   3 100 isqg        1.4     Fernando H. Toledo 10M In Silico Quantitative Ge...
#>   4 100 snht        1.0.5   Josh Browning       6y Standard Normal Homogenei...
#>   5 100 TileManager 0.4.1   Andrew Plowright    1y Tile Manager                
#>   6 100 carx        0.7.1   Chao Wang           6y Censored Autoregressive M...
#>   7 100 erratum     2.2.0   John Coene          2y Handle Error and Warning ...
#>   8 100 metaMix     0.3     Sofia Morfopoulou   4y Bayesian Mixture Analysis...
#>   9 100 SpNetPrep   1.2     Alvaro Briz-Redon   2y Linear Network Preprocess...
#>  10 100 GlobalFit   1.2     Daniel Hartleb      7y Bi-Level Optimization of ...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)
#> - "advanced search" -------------------------- 9287 packages in 0.001 seconds -
#>   #     package     version by                   @ title                       
#>   1 100 MHMM        1.0.0   Matthieu Marbac     3y Finite Mixture of Hidden ...
#>   2 100 SelvarMix   1.2.1   Mohammed Sedki      6y Regularization for Variab...
#>   3 100 isqg        1.4     Fernando H. Toledo 10M In Silico Quantitative Ge...
#>   4 100 snht        1.0.5   Josh Browning       6y Standard Normal Homogenei...
#>   5 100 TileManager 0.4.1   Andrew Plowright    1y Tile Manager                
#>   6 100 carx        0.7.1   Chao Wang           6y Censored Autoregressive M...
#>   7 100 erratum     2.2.0   John Coene          2y Handle Error and Warning ...
#>   8 100 metaMix     0.3     Sofia Morfopoulou   4y Bayesian Mixture Analysis...
#>   9 100 SpNetPrep   1.2     Alvaro Briz-Redon   2y Linear Network Preprocess...
#>  10 100 GlobalFit   1.2     Daniel Hartleb      7y Bi-Level Optimization of ...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")
#> - "advanced search" --------------------------- 244 packages in 0.019 seconds -
#>   #     package            version  by                  @ title                
#>   1 100 onepass            0.1.2    Jonathan Lin       2y 1password Credenti...
#>   2 100 RTriangle          1.6.0.12 David C. Sterratt  7M Triangle - A 2D Qu...
#>   3 100 selectiveInference 1.2.5    Rob Tibshirani     4y Tools for Post-Sel...
#>   4 100 LeMaRns            0.1.2    Michael A. Spence  4y Length-Based Multi...
#>   5 100 xtable             1.8.4    David Scott        4y Export Tables to L...
#>   6 100 tableone           0.13.2   Kazuki Yoshida     1y Create 'Table 1' t...
#>   7 100 bridgesampling     1.1.2    Quentin F. Gronau  2y Bridge Sampling fo...
#>   8 100 rmcorr             0.5.4    Laura R. Marusich  8M Repeated Measures ...
#>   9 100 smcfcs             1.7.1    Jonathan Bartlett  9M Multiple Imputatio...
#>  10 100 plink              1.5.1    Jonathan P. Weeks  6y IRT Separate Calib...

# Fuzzy search
advanced_search(Author = "Johnathan~1")
#> - "advanced search" --------------------------- 246 packages in 0.003 seconds -
#>   #     package    version by                           @ title                
#>   1 100 lda        1.4.2   Jonathan Chang              8y Collapsed Gibbs Sa...
#>   2  98 chords     0.95.4  Jonathan Rosenblatt         7y Estimation in Resp...
#>   3  98 bootImpute 1.2.1   Jonathan Bartlett           2M Bootstrap Inferenc...
#>   4  97 mlmi       1.1.2   Jonathan Bartlett           2M Maximum Likelihood...
#>   5  96 handyplots 1.1.3   Jonathan Schwartz           5y Handy Plots          
#>   6  96 UnitCircle 0.1.3   Jonathan Berrisch           5y Check if Roots of ...
#>   7  95 RSeed      0.1.60  Claus Jonathan Fritzemeier  7y Borenstein Analysis  
#>   8  95 rSHAPE     0.3.2   Jonathan Dench              4y Simulated Haploid ...
#>   9  91 tensor     1.5     Jonathan Rougier           11y Tensor product of ...
#>  10  91 tint       0.1.4   Dirk Eddelbuettel           9M 'tint' is not 'Tufte'