Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...

Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json

A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")
#> - "advanced search" ----------------------------- 3 packages in 0.004 seconds -
#>   #     package   version  by         @ title                                  
#>  1  100 udunits2  0.13.2.1 ORPHANED  2y Udunits-2 Bindings for R               
#>  2  100 shipunov  1.17.1   ORPHANED  1y Miscellaneous Functions from Alexey ...
#>  3  100 fastclime 1.4.1.1  ORPHANED  1y A Fast Solver for Parameterized LP P...

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")
#> - "advanced search" ---------------------------- 22 packages in 0.005 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 mockery     0.4.4   Hadley Wickham     7M Mocking Library for R        
#>   2  94 pkgbuild    1.4.4   Gábor Csárdi       1M Find Tools Needed to Build...
#>   3  85 memoise     2.0.1   Winston Chang      2y 'Memoisation' of Functions   
#>   4  78 devtools    2.4.5   Jennifer Bryan     2y Tools to Make Developing R...
#>   5  78 sessioninfo 1.2.2   Gábor Csárdi       2y R Session Information        
#>   6  75 xml2        1.3.6   Hadley Wickham     5M Parse XML                    
#>   7  75 odbc        1.4.2   Hadley Wickham     3M Connect to ODBC Compatible...
#>   8  72 meltr       1.0.2   Duncan Garmonsway  3M Read Non-Rectangular Text ...
#>   9  70 fs          1.6.3   Gábor Csárdi       9M Cross-Platform File System...
#>  10  69 remotes     2.5.0   Gábor Csárdi       1M R Package Installation fro...
advanced_search("Author: Hester AND Author: Wickham")
#> - "advanced search" ---------------------------- 22 packages in 0.004 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 mockery     0.4.4   Hadley Wickham     7M Mocking Library for R        
#>   2  94 pkgbuild    1.4.4   Gábor Csárdi       1M Find Tools Needed to Build...
#>   3  85 memoise     2.0.1   Winston Chang      2y 'Memoisation' of Functions   
#>   4  78 devtools    2.4.5   Jennifer Bryan     2y Tools to Make Developing R...
#>   5  78 sessioninfo 1.2.2   Gábor Csárdi       2y R Session Information        
#>   6  75 xml2        1.3.6   Hadley Wickham     5M Parse XML                    
#>   7  75 odbc        1.4.2   Hadley Wickham     3M Connect to ODBC Compatible...
#>   8  72 meltr       1.0.2   Duncan Garmonsway  3M Read Non-Rectangular Text ...
#>   9  70 fs          1.6.3   Gábor Csárdi       9M Cross-Platform File System...
#>  10  69 remotes     2.5.0   Gábor Csárdi       1M R Package Installation fro...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")
#> - "advanced search" ---------------------------- 25 packages in 0.011 seconds -
#>   #     package        version by                        @ title               
#>   1 100 knitrBootstrap 1.0.3   Jim Hester               2M 'knitr' Bootstrap...
#>   2  92 types          1.0.0   Jim Hester               8y Type Annotations    
#>   3  81 primerTree     1.0.6   Matt Cannon              2y Visually Assessin...
#>   4  73 rex            1.2.1   Kevin Ushey              2y Friendly Regular ...
#>   5  69 gmailr         2.0.0   Jennifer Bryan          10M Access the 'Gmail...
#>   6  66 spelling       2.3.0   Jeroen Ooms              2M Tools for Spell C...
#>   7  64 roger          1.5.1   Vincent Goulet           6M Automated Grading...
#>   8  61 available      1.1.0   Gábor Csárdi             2y Check if the Titl...
#>   9  59 bench          1.1.3   Davis Vaughan            1y High Precision Ti...
#>  10  59 tidycode       0.1.1   Lucy D'Agostino McGowan  4y Analyze Lines of ...

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")
#> - "advanced search" ---------------------------- 22 packages in 0.007 seconds -
#>   #     package     version by               @ title                           
#>   1 100 mockery     0.4.4   Hadley Wickham  7M Mocking Library for R           
#>   2  87 xml2        1.3.6   Hadley Wickham  5M Parse XML                       
#>   3  87 odbc        1.4.2   Hadley Wickham  3M Connect to ODBC Compatible Da...
#>   4  87 pkgbuild    1.4.4   Gábor Csárdi    1M Find Tools Needed to Build R ...
#>   5  76 memoise     2.0.1   Winston Chang   2y 'Memoisation' of Functions      
#>   6  74 devtools    2.4.5   Jennifer Bryan  2y Tools to Make Developing R Pa...
#>   7  74 S7          0.1.1   Hadley Wickham  7M An Object Oriented System Mea...
#>   8  71 sessioninfo 1.2.2   Gábor Csárdi    2y R Session Information           
#>   9  67 fs          1.6.3   Gábor Csárdi    9M Cross-Platform File System Op...
#>  10  66 remotes     2.5.0   Gábor Csárdi    1M R Package Installation from R...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")
#> - "advanced search" ---------------------------- 25 packages in 0.011 seconds -
#>   #     package        version by                        @ title               
#>   1 100 knitrBootstrap 1.0.3   Jim Hester               2M 'knitr' Bootstrap...
#>   2  92 types          1.0.0   Jim Hester               8y Type Annotations    
#>   3  81 primerTree     1.0.6   Matt Cannon              2y Visually Assessin...
#>   4  73 rex            1.2.1   Kevin Ushey              2y Friendly Regular ...
#>   5  69 gmailr         2.0.0   Jennifer Bryan          10M Access the 'Gmail...
#>   6  66 spelling       2.3.0   Jeroen Ooms              2M Tools for Spell C...
#>   7  64 roger          1.5.1   Vincent Goulet           6M Automated Grading...
#>   8  61 available      1.1.0   Gábor Csárdi             2y Check if the Titl...
#>   9  59 bench          1.1.3   Davis Vaughan            1y High Precision Ti...
#>  10  59 tidycode       0.1.1   Lucy D'Agostino McGowan  4y Analyze Lines of ...

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")
#> - "advanced search" --------------------------- 506 packages in 0.036 seconds -
#>   #     package       version by                         @ title               
#>   1 100 phd           0.2     Jesse Hemerik             2y Permutation Testi...
#>   2  94 confSAM       0.2     Jesse Hemerik             6y Estimates and Bou...
#>   3  89 permutes      2.8     Cesko C. Voeten           7M Permutation Tests...
#>   4  78 BurStMisc     1.1     Pat Burns                 8y Burns Statistics ...
#>   5  75 resample      0.6     Tim Hesterberg            2y Resampling Functions
#>   6  74 USP           0.1.2   Thomas B. Berrett         3y U-Statistic Permu...
#>   7  74 IMEC          0.2.0   Maximilian Maier          3y Ising Model of Ex...
#>   8  73 packager      1.15.2  Andreas Dominik Cullmann  8M Create, Build and...
#>   9  69 fdrci         2.4     Joshua Millstein          2y Permutation-Based...
#>  10  68 MultSurvTests 0.2     Lukas Arnroth             3y Permutation Tests...

# Packages that have a certain field
advanced_search("_exists_" = "URL")
#> - "advanced search" ------------------------- 10000 packages in 0.004 seconds -
#>   #     package           version by                      @ title              
#>   1 100 mangoTraining     1.1.1   Andrew Little          3y Mango Solutions ...
#>   2 100 assertive.data.us 0.0.2   Richard Cotton         5y Assertions to Ch...
#>   3 100 deepboost         0.1.6   Daniel Marcous         6y Deep Boosting En...
#>   4 100 SPQR              0.1.0   Steven Xu              2y Semi-Parametric ...
#>   5 100 remoter           0.4.0   Drew Schmidt           6y Remote R: Contro...
#>   6 100 irtplay           1.6.5   Hwanggyu Lim           2y Unidimensional I...
#>   7 100 drfit             0.7.2   Johannes Ranke         6y Dose-Response Da...
#>   8 100 UNF               2.0.8   Thomas J. Leeper       2y Tools for Creati...
#>   9 100 dragon            1.2.1   Stephanie J. Spielman  2y Deep Time Redox ...
#>  10 100 mudata2           1.1.3   Dewey Dunnington       1y Interchange Tool...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")
#> - "advanced search" -------------------------- 9736 packages in 0.005 seconds -
#>   #     package     version by                    @ title                      
#>   1 NaN TileManager 0.4.1   Andrew Plowright     2y Tile Manager               
#>   2 NaN hdtg        0.2.0   Zhenyu Zhang         2y Generate Samples from Mu...
#>   3 NaN twingp      1.0     Akhil Vakayil       11M A Fast Global-Local Gaus...
#>   4 NaN MatTransMix 0.1.16  Xuwen Zhu            1y Clustering with Matrix G...
#>   5 NaN JumpeR      0.3.0   Greg Pilgrim         2y Importing and Working wi...
#>   6 NaN HLSM        0.9.0   Tracy Sweet          2y Hierarchical Latent Spac...
#>   7 NaN LRQMM       1.2.3   Sayyed Reza Alavian  3y Fitting Linear Quantile ...
#>   8 NaN coxmeg      1.1.4   Liang He             1y Cox Mixed-Effects Models...
#>   9 NaN RTOMO       1.1.6   Jonathan M. Lees     6y Visualization for Seismi...
#>  10 NaN bmotif      2.0.2   Benno Simmons        4y Motif Analyses of Bipart...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)
#> - "advanced search" -------------------------- 9736 packages in 0.037 seconds -
#>   #     package     version by                    @ title                      
#>   1 NaN TileManager 0.4.1   Andrew Plowright     2y Tile Manager               
#>   2 NaN hdtg        0.2.0   Zhenyu Zhang         2y Generate Samples from Mu...
#>   3 NaN twingp      1.0     Akhil Vakayil       11M A Fast Global-Local Gaus...
#>   4 NaN MatTransMix 0.1.16  Xuwen Zhu            1y Clustering with Matrix G...
#>   5 NaN JumpeR      0.3.0   Greg Pilgrim         2y Importing and Working wi...
#>   6 NaN HLSM        0.9.0   Tracy Sweet          2y Hierarchical Latent Spac...
#>   7 NaN LRQMM       1.2.3   Sayyed Reza Alavian  3y Fitting Linear Quantile ...
#>   8 NaN coxmeg      1.1.4   Liang He             1y Cox Mixed-Effects Models...
#>   9 NaN RTOMO       1.1.6   Jonathan M. Lees     6y Visualization for Seismi...
#>  10 NaN bmotif      2.0.2   Benno Simmons        4y Motif Analyses of Bipart...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")
#> - "advanced search" ----------------------------- 0 packages in 0.003 seconds -

# Fuzzy search
advanced_search(Author = "Johnathan~1")
#> - "advanced search" ---------------------------- 264 packages in 0.02 seconds -
#>   #     package    version by                                        @ title   
#>   1 100 bootImpute 1.2.1   Jonathan Bartlett                       11M Boots...
#>   2 100 lda        1.4.2   Jonathan Chang                           8y Colla...
#>   3 100 mlmi       1.1.2   Jonathan Bartlett                       11M Maxim...
#>   4 100 handyplots 1.1.3   Jonathan Schwartz                        5y Handy...
#>   5 100 UnitCircle 0.1.3   Jonathan Berrisch                        6y Check...
#>   6 100 chords     0.95.4  Jonathan Rosenblatt                      7y Estim...
#>   7 100 rSHAPE     0.3.2   Jonathan Dench                           5y Simul...
#>   8  96 ClamR      2.1.3   Jonathan M. Lees<jonathan.lees@unc.edu>  8M Time ...
#>   9  96 RSeed      0.1.60  Claus Jonathan Fritzemeier               8y Boren...
#>  10  92 iSDM       1.0     Tarek Hattab                             7y Invas...