Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...

Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json

A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")
#> - "advanced search" ----------------------------- 6 packages in 0.001 seconds -
#>   #     package     version   by         @ title                               
#>  1  100 penalized   0.9.51    ORPHANED  4y L1 (Lasso and Fused Lasso) and L2...
#>  2   99 bibtex      0.4.2.3   ORPHANED  1y Bibtex Parser                       
#>  3   95 ggmap       3.0.0     ORPHANED  3y Spatial Visualization with ggplot2  
#>  4   95 cairoDevice 2.28.2.1  ORPHANED  4M Embeddable Cairo Graphics Device ...
#>  5   94 distances   0.1.8     ORPHANED  2y Tools for Distance Metrics          
#>  6   94 RGtk2       2.20.36.2 ORPHANED  4M R Bindings for Gtk 2.8.0 and Above  

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")
#> - "advanced search" ---------------------------- 18 packages in 0.001 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi       1M Find Tools Needed to Build...
#>   2  81 memoise     2.0.1   Winston Chang      2M 'Memoisation' of Functions   
#>   3  80 xml2        1.3.3   Hadley Wickham     2M Parse XML                    
#>   4  75 devtools    2.4.3   Jennifer Bryan     2M Tools to Make Developing R...
#>   5  74 sessioninfo 1.2.2   Gábor Csárdi       2M R Session Information        
#>   6  73 odbc        1.3.3   Hadley Wickham     2M Connect to ODBC Compatible...
#>   7  67 remotes     2.4.2   Gábor Csárdi       2M R Package Installation fro...
#>   8  67 meltr       1.0.0   Duncan Garmonsway  5M Read Non-Rectangular Text ...
#>   9  67 pkgload     1.2.4   Lionel Henry       2M Simulate Package Installat...
#>  10  66 withr       2.4.3   Lionel Henry       2M Run Code 'With' Temporaril...
advanced_search("Author: Hester AND Author: Wickham")
#> - "advanced search" ---------------------------- 18 packages in 0.001 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi       1M Find Tools Needed to Build...
#>   2  81 memoise     2.0.1   Winston Chang      2M 'Memoisation' of Functions   
#>   3  80 xml2        1.3.3   Hadley Wickham     2M Parse XML                    
#>   4  75 devtools    2.4.3   Jennifer Bryan     2M Tools to Make Developing R...
#>   5  74 sessioninfo 1.2.2   Gábor Csárdi       2M R Session Information        
#>   6  73 odbc        1.3.3   Hadley Wickham     2M Connect to ODBC Compatible...
#>   7  67 remotes     2.4.2   Gábor Csárdi       2M R Package Installation fro...
#>   8  67 meltr       1.0.0   Duncan Garmonsway  5M Read Non-Rectangular Text ...
#>   9  67 pkgload     1.2.4   Lionel Henry       2M Simulate Package Installat...
#>  10  66 withr       2.4.3   Lionel Henry       2M Run Code 'With' Temporaril...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")
#> - "advanced search" ---------------------------- 28 packages in 0.003 seconds -
#>   #     package        version by                @ title                       
#>   1 100 primerTree     1.0.5   Jim Hester       2y Visually Assessing the Sp...
#>   2  92 knitrBootstrap 1.0.2   Jim Hester       4y 'knitr' Bootstrap Framework 
#>   3  92 types          1.0.0   Jim Hester       5y Type Annotations            
#>   4  79 pak            0.2.1   Gábor Csárdi     1M Another Approach to Packa...
#>   5  76 gmailr         1.0.1   Jennifer Bryan   2M Access the 'Gmail' 'RESTf...
#>   6  73 dplyr.teradata 0.4.1   Koji Makiyama    1y A 'Teradata' Backend for ...
#>   7  71 lintr          2.0.1   Jim Hester       2y A 'Linter' for R Code       
#>   8  71 decor          1.0.1   Romain François  2M Retrieve Code Decorations   
#>   9  67 mockery        0.4.2   Jim Hester       2y Mocking Library for R       
#>  10  64 rex            1.2.1   Kevin Ushey      2M Friendly Regular Expressions

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")
#> - "advanced search" ---------------------------- 18 packages in 0.004 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi       1M Find Tools Needed to Build...
#>   2  97 odbc        1.3.3   Hadley Wickham     2M Connect to ODBC Compatible...
#>   3  97 xml2        1.3.3   Hadley Wickham     2M Parse XML                    
#>   4  81 memoise     2.0.1   Winston Chang      2M 'Memoisation' of Functions   
#>   5  76 devtools    2.4.3   Jennifer Bryan     2M Tools to Make Developing R...
#>   6  73 sessioninfo 1.2.2   Gábor Csárdi       2M R Session Information        
#>   7  73 pkgload     1.2.4   Lionel Henry       2M Simulate Package Installat...
#>   8  70 remotes     2.4.2   Gábor Csárdi       2M R Package Installation fro...
#>   9  68 fs          1.5.2   Gábor Csárdi       2M Cross-Platform File System...
#>  10  68 meltr       1.0.0   Duncan Garmonsway  5M Read Non-Rectangular Text ...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")
#> - "advanced search" ---------------------------- 28 packages in 0.002 seconds -
#>   #     package        version by                @ title                       
#>   1 100 primerTree     1.0.5   Jim Hester       2y Visually Assessing the Sp...
#>   2  92 knitrBootstrap 1.0.2   Jim Hester       4y 'knitr' Bootstrap Framework 
#>   3  92 types          1.0.0   Jim Hester       5y Type Annotations            
#>   4  79 pak            0.2.1   Gábor Csárdi     1M Another Approach to Packa...
#>   5  76 gmailr         1.0.1   Jennifer Bryan   2M Access the 'Gmail' 'RESTf...
#>   6  73 dplyr.teradata 0.4.1   Koji Makiyama    1y A 'Teradata' Backend for ...
#>   7  71 lintr          2.0.1   Jim Hester       2y A 'Linter' for R Code       
#>   8  71 decor          1.0.1   Romain François  2M Retrieve Code Decorations   
#>   9  67 mockery        0.4.2   Jim Hester       2y Mocking Library for R       
#>  10  64 rex            1.2.1   Kevin Ushey      2M Friendly Regular Expressions

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")
#> - "advanced search" --------------------------- 367 packages in 0.011 seconds -
#>   #     package           version by                         @ title           
#>   1 100 phd               0.2     Jesse Hemerik            25d Permutation T...
#>   2  94 confSAM           0.2     Jesse Hemerik             4y Estimates and...
#>   3  85 permutes          2.3.2   Cesko C. Voeten           2M Permutation T...
#>   4  79 BurStMisc         1.1     Pat Burns                 5y Burns Statist...
#>   5  76 fdrci             2.2     Joshua Millstein          8M Permutation-B...
#>   6  74 MultSurvTests     0.2     Lukas Arnroth             8M Permutation T...
#>   7  74 USP               0.1.2   Thomas B. Berrett         1y U-Statistic P...
#>   8  74 resample          0.4     Tim Hesterberg            7y Resampling Fu...
#>   9  74 packager          1.14.0  Andreas Dominik Cullmann  2M Create, Build...
#>  10  72 phyloseqGraphTest 0.1.0   Julia Fukuyama            2y Graph-Based P...

# Packages that have a certain field
advanced_search("_exists_" = "URL")
#> - "advanced search" -------------------------- 9793 packages in 0.001 seconds -
#>   #     package      version by                  @ title                       
#>   1 100 sandwichr    1.0.0   Yue Lin            2d Spatial Interpolation Bas...
#>   2 100 ChemoSpec    6.1.0   Bryan A. Hanson   22d Exploratory Chemometrics ...
#>   3 100 bayesTFR     7.1.1   Hana Sevcikova     3M Bayesian Fertility Projec...
#>   4 100 OwenQ        1.0.5   Stéphane Laurent  25d Owen Q-Function             
#>   5 100 GetTDData    1.4.2   Marcelo Perlin     2y Get Data for Brazilian Bo...
#>   6 100 airqualityES 1.0.0   Jose V. Die        2y Air Quality Measurements ...
#>   7 100 gettz        0.0.4   Dirk Eddelbuettel  2y Get the Timezone Information
#>   8 100 vegclust     2.0.0   Miquel De Cáceres  8M Fuzzy Clustering of Veget...
#>   9 100 mlrCPO       0.3.7.3 Martin Binder      3M Composable Preprocessing ...
#>  10 100 DCEM         2.0.5   Sharma Parichit   15d Clustering Big Data using...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")
#> - "advanced search" -------------------------- 9120 packages in 0.001 seconds -
#>   #     package    version by                        @ title                   
#>   1 100 PKPDsim    1.1.0   Ron Keizer              12d Tools for Performing ...
#>   2 100 matlab2r   1.0.0   Waldir Leoncio           2d Translation Layer fro...
#>   3 100 BLR        1.6     Paulino Perez Rodriguez  2y Bayesian Linear Regre...
#>   4 100 funtimes   8.1     Vyacheslav Lyubchich     6M Functions for Time Se...
#>   5 100 afmToolkit 0.0.1   Rafael Benitez           5y Functions for Atomic ...
#>   6 100 rkafkajars 1.2     Shruti Gupta             2M External Jars Require...
#>   7 100 ordinalNet 2.10    Michael Wurm             5M Penalized Ordinal Reg...
#>   8 100 intccr     3.0.3   Jun Park                 7M Semiparametric Compet...
#>   9 100 NHMSAR     1.17    Valerie Monbet           2y Non-Homogeneous Marko...
#>  10 100 AFM        2.0     Mathieu Beauvais         1y Atomic Force Microsco...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)
#> - "advanced search" ------------------------------ 9120 packages in 0 seconds -
#>   #     package    version by                        @ title                   
#>   1 100 PKPDsim    1.1.0   Ron Keizer              12d Tools for Performing ...
#>   2 100 matlab2r   1.0.0   Waldir Leoncio           2d Translation Layer fro...
#>   3 100 BLR        1.6     Paulino Perez Rodriguez  2y Bayesian Linear Regre...
#>   4 100 funtimes   8.1     Vyacheslav Lyubchich     6M Functions for Time Se...
#>   5 100 afmToolkit 0.0.1   Rafael Benitez           5y Functions for Atomic ...
#>   6 100 rkafkajars 1.2     Shruti Gupta             2M External Jars Require...
#>   7 100 ordinalNet 2.10    Michael Wurm             5M Penalized Ordinal Reg...
#>   8 100 intccr     3.0.3   Jun Park                 7M Semiparametric Compet...
#>   9 100 NHMSAR     1.17    Valerie Monbet           2y Non-Homogeneous Marko...
#>  10 100 AFM        2.0     Mathieu Beauvais         1y Atomic Force Microsco...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")
#> - "advanced search" --------------------------- 233 packages in 0.001 seconds -
#>   #     package         version by                         @ title             
#>   1 100 rDotNet         0.9.1   Jonathan Shore            4y Low-Level Inter...
#>   2 100 xtable          1.8.4   David Scott               3y Export Tables t...
#>   3 100 raster          3.5.15  Robert J. Hijmans         8d Geographic Data...
#>   4 100 knitr           1.37    Yihui Xie                 2M A General-Purpo...
#>   5 100 bridgesampling  1.1.2   Quentin F. Gronau        10M Bridge Sampling...
#>   6 100 tableone        0.13.0  Kazuki Yoshida            7M Create 'Table 1...
#>   7 100 gdalUtils       2.0.3.2 Jonathan Asher Greenberg  2y Wrappers for th...
#>   8 100 fivethirtyeight 0.6.2   Albert Y. Kim             4M Data and Code B...
#>   9 100 tscount         1.4.3   Tobias Liboschik          1y Analysis of Cou...
#>  10 100 rmcorr          0.4.5   Laura R. Marusich         2M Repeated Measur...

# Fuzzy search
advanced_search(Author = "Johnathan~1")
#> - "advanced search" --------------------------- 234 packages in 0.002 seconds -
#>   #     package    version by                                        @ title   
#>   1 100 UScancer   0.1.2   Jonathan Lee                             7y Creat...
#>   2 100 mlmi       1.1.1   Jonathan Bartlett                        2M Maxim...
#>   3  99 lda        1.4.2   Jonathan Chang                           6y Colla...
#>   4  97 handyplots 1.1.3   Jonathan Schwartz                        3y Handy...
#>   5  97 UnitCircle 0.1.3   Jonathan Berrisch                        4y Check...
#>   6  96 bootImpute 1.2.0   Jonathan Bartlett                       10M Boots...
#>   7  96 chords     0.95.4  Jonathan Rosenblatt                      5y Estim...
#>   8  95 geomapdata 1.0.4   Jonathan M. Lees<jonathan.lees@unc.edu>  9y Data ...
#>   9  94 RSeed      0.1.60  Claus Jonathan Fritzemeier               5y Boren...
#>  10  93 rSHAPE     0.3.2   Jonathan Dench                           3y Simul...