Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...

Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json

A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")
#> - "advanced search" ----------------------------- 3 packages in 0.001 seconds -
#>   #     package   version  by         @ title                             
#>  1  100 ggmap     3.0.0    ORPHANED  4y Spatial Visualization with ggplot2
#>  2   99 distances 0.1.8    ORPHANED  3y Tools for Distance Metrics        
#>  3   99 udunits2  0.13.2.1 ORPHANED  8M Udunits-2 Bindings for R          

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")
#> - "advanced search" ---------------------------- 19 packages in 0.001 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi      10M Find Tools Needed to Build...
#>   2  96 mockery     0.4.3   Hadley Wickham     8M Mocking Library for R        
#>   3  83 memoise     2.0.1   Winston Chang     11M 'Memoisation' of Functions   
#>   4  82 xml2        1.3.3   Hadley Wickham    11M Parse XML                    
#>   5  76 devtools    2.4.5   Jennifer Bryan    15d Tools to Make Developing R...
#>   6  75 sessioninfo 1.2.2   Gábor Csárdi      11M R Session Information        
#>   7  75 odbc        1.3.3   Hadley Wickham    11M Connect to ODBC Compatible...
#>   8  71 remotes     2.4.2   Gábor Csárdi      11M R Package Installation fro...
#>   9  68 meltr       1.0.1   Duncan Garmonsway  2M Read Non-Rectangular Text ...
#>  10  67 pkgload     1.3.0   Lionel Henry       4M Simulate Package Installat...
advanced_search("Author: Hester AND Author: Wickham")
#> - "advanced search" ---------------------------- 19 packages in 0.001 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 pkgbuild    1.3.1   Gábor Csárdi      10M Find Tools Needed to Build...
#>   2  96 mockery     0.4.3   Hadley Wickham     8M Mocking Library for R        
#>   3  83 memoise     2.0.1   Winston Chang     11M 'Memoisation' of Functions   
#>   4  82 xml2        1.3.3   Hadley Wickham    11M Parse XML                    
#>   5  76 devtools    2.4.5   Jennifer Bryan    15d Tools to Make Developing R...
#>   6  75 sessioninfo 1.2.2   Gábor Csárdi      11M R Session Information        
#>   7  75 odbc        1.3.3   Hadley Wickham    11M Connect to ODBC Compatible...
#>   8  71 remotes     2.4.2   Gábor Csárdi      11M R Package Installation fro...
#>   9  68 meltr       1.0.1   Duncan Garmonsway  2M Read Non-Rectangular Text ...
#>  10  67 pkgload     1.3.0   Lionel Henry       4M Simulate Package Installat...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")
#> - "advanced search" ---------------------------- 26 packages in 0.004 seconds -
#>   #     package        version by                @ title                       
#>   1 100 knitrBootstrap 1.0.2   Jim Hester       4y 'knitr' Bootstrap Framework 
#>   2  96 types          1.0.0   Jim Hester       6y Type Annotations            
#>   3  83 primerTree     1.0.6   Matt Cannon      7M Visually Assessing the Sp...
#>   4  80 pak            0.3.1   Gábor Csárdi     2M Another Approach to Packa...
#>   5  80 gmailr         1.0.1   Jennifer Bryan  11M Access the 'Gmail' 'RESTf...
#>   6  74 decor          1.0.1   Romain François 11M Retrieve Code Decorations   
#>   7  71 rex            1.2.1   Kevin Ushey     11M Friendly Regular Expressions
#>   8  66 bench          1.1.2   Davis Vaughan   11M High Precision Timing of ...
#>   9  66 roger          1.2.0   Vincent Goulet  11d Automated Grading of R Sc...
#>  10  64 spelling       2.2     Jeroen Ooms      2y Tools for Spell Checking ...

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")
#> - "advanced search" ---------------------------- 19 packages in 0.006 seconds -
#>   #     package     version by               @ title                           
#>   1 100 mockery     0.4.3   Hadley Wickham  8M Mocking Library for R           
#>   2  92 pkgbuild    1.3.1   Gábor Csárdi   10M Find Tools Needed to Build R ...
#>   3  90 xml2        1.3.3   Hadley Wickham 11M Parse XML                       
#>   4  90 odbc        1.3.3   Hadley Wickham 11M Connect to ODBC Compatible Da...
#>   5  77 memoise     2.0.1   Winston Chang  11M 'Memoisation' of Functions      
#>   6  75 devtools    2.4.5   Jennifer Bryan 15d Tools to Make Developing R Pa...
#>   7  69 sessioninfo 1.2.2   Gábor Csárdi   11M R Session Information           
#>   8  68 remotes     2.4.2   Gábor Csárdi   11M R Package Installation from R...
#>   9  67 pkgload     1.3.0   Lionel Henry    4M Simulate Package Installation...
#>  10  66 fs          1.5.2   Gábor Csárdi   11M Cross-Platform File System Op...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")
#> - "advanced search" ---------------------------- 26 packages in 0.005 seconds -
#>   #     package        version by                @ title                       
#>   1 100 knitrBootstrap 1.0.2   Jim Hester       4y 'knitr' Bootstrap Framework 
#>   2  96 types          1.0.0   Jim Hester       6y Type Annotations            
#>   3  83 primerTree     1.0.6   Matt Cannon      7M Visually Assessing the Sp...
#>   4  80 pak            0.3.1   Gábor Csárdi     2M Another Approach to Packa...
#>   5  80 gmailr         1.0.1   Jennifer Bryan  11M Access the 'Gmail' 'RESTf...
#>   6  74 decor          1.0.1   Romain François 11M Retrieve Code Decorations   
#>   7  71 rex            1.2.1   Kevin Ushey     11M Friendly Regular Expressions
#>   8  66 bench          1.1.2   Davis Vaughan   11M High Precision Timing of ...
#>   9  66 roger          1.2.0   Vincent Goulet  11d Automated Grading of R Sc...
#>  10  64 spelling       2.2     Jeroen Ooms      2y Tools for Spell Checking ...

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")
#> - "advanced search" --------------------------- 398 packages in 0.015 seconds -
#>   #     package           version by                         @ title           
#>   1 100 phd               0.2     Jesse Hemerik            10M Permutation T...
#>   2  93 confSAM           0.2     Jesse Hemerik             5y Estimates and...
#>   3  83 permutes          2.6     Cesko C. Voeten           3M Permutation T...
#>   4  78 BurStMisc         1.1     Pat Burns                 6y Burns Statist...
#>   5  75 resample          0.6     Tim Hesterberg            4M Resampling Fu...
#>   6  73 USP               0.1.2   Thomas B. Berrett         2y U-Statistic P...
#>   7  73 fdrci             2.4     Joshua Millstein          8d Permutation-B...
#>   8  73 packager          1.14.0  Andreas Dominik Cullmann 11M Create, Build...
#>   9  69 MultSurvTests     0.2     Lukas Arnroth             1y Permutation T...
#>  10  67 phyloseqGraphTest 0.1.0   Julia Fukuyama            3y Graph-Based P...

# Packages that have a certain field
advanced_search("_exists_" = "URL")
#> - "advanced search" ------------------------- 10373 packages in 0.002 seconds -
#>   #     package            version by                           @ title        
#>   1 100 epi2me2r           0.1.0   Quentin D. Read             5M Process Na...
#>   2 100 picR               1.0.0   Javier Flores               2d Predictive...
#>   3 100 calendR            1.1     José Carlos Soage González  2y Ready to P...
#>   4 100 pense              2.1.0   David Kepplinger            1y Penalized ...
#>   5 100 MOSS               0.2.2   Agustin Gonzalez-Reymundez  7M Multi-Omic...
#>   6 100 Ostats             0.2.0   Quentin D. Read             1M O-Stats, o...
#>   7 100 smidm              1.0     Sonja Jäckle                2M Statistica...
#>   8 100 processanimateR    1.0.5   Felix Mannhardt             3M Process Ma...
#>   9 100 PortfolioEffectHFT 1.8     Andrey Kostin               6y High Frequ...
#>  10 100 imputeFin          0.1.2   Daniel P. Palomar           2y Imputation...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")
#> - "advanced search" -------------------------- 8706 packages in 0.002 seconds -
#>   #     package       version  by                      @ title                 
#>   1 100 WLogit        1.0      Wencan Zhu             4M Whitening Logistic ...
#>   2 100 blockcpd      1.0.0    Lucas Prates           3M Change Point Detect...
#>   3 100 boxplotdbl    1.4.0    Shinichiro Tomizono    6M Double Box Plot for...
#>   4 100 filebin       0.0.5    Andrew B. Collier     10M Wrapper for the Fil...
#>   5 100 icrf          2.0.0    Hunyong Cho            1y Interval Censored R...
#>   6 100 gds           0.1.1    Partha Sarathi Bishnu  1y Descriptive Statist...
#>   7 100 rLakeAnalyzer 1.11.4.1 Luke Winslow           3y Lake Physics Tools    
#>   8 100 quantCurves   1.0.0    Sandie Ferrigno        8M Estimate Quantiles ...
#>   9 100 unittest      1.5.3    Jamie Lentin           1M TAP-Compliant Unit ...
#>  10 100 depth         2.1.1.1  Jean-Francois Plante   3y Nonparametric Depth...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)
#> - "advanced search" -------------------------- 8706 packages in 0.001 seconds -
#>   #     package       version  by                      @ title                 
#>   1 100 WLogit        1.0      Wencan Zhu             4M Whitening Logistic ...
#>   2 100 blockcpd      1.0.0    Lucas Prates           3M Change Point Detect...
#>   3 100 boxplotdbl    1.4.0    Shinichiro Tomizono    6M Double Box Plot for...
#>   4 100 filebin       0.0.5    Andrew B. Collier     10M Wrapper for the Fil...
#>   5 100 icrf          2.0.0    Hunyong Cho            1y Interval Censored R...
#>   6 100 gds           0.1.1    Partha Sarathi Bishnu  1y Descriptive Statist...
#>   7 100 rLakeAnalyzer 1.11.4.1 Luke Winslow           3y Lake Physics Tools    
#>   8 100 quantCurves   1.0.0    Sandie Ferrigno        8M Estimate Quantiles ...
#>   9 100 unittest      1.5.3    Jamie Lentin           1M TAP-Compliant Unit ...
#>  10 100 depth         2.1.1.1  Jean-Francois Plante   3y Nonparametric Depth...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")
#> - "advanced search" --------------------------- 238 packages in 0.003 seconds -
#>   #     package        version by                           @ title            
#>   1 100 stcos          0.3.0   Andrew M. Raim              2y Space-Time Cha...
#>   2 100 Sample.Size    1.0     Wei Jiang<wjiang@kumc.edu>  9y Sample size ca...
#>   3 100 Pijavski       1.0.3   Gleb Beliakov               2y Global Univari...
#>   4 100 scITD          1.0.2   Jonathan Mitchel            7M Single-Cell In...
#>   5 100 ghee           0.1.0   Jonathan Trattner           2y Provides a Lig...
#>   6 100 knitr          1.40    Yihui Xie                   2M A General-Purp...
#>   7 100 xtable         1.8.4   David Scott                 4y Export Tables ...
#>   8 100 raster         3.6.3   Robert J. Hijmans           1M Geographic Dat...
#>   9 100 tableone       0.13.2  Kazuki Yoshida              6M Create 'Table ...
#>  10 100 bridgesampling 1.1.2   Quentin F. Gronau           2y Bridge Samplin...

# Fuzzy search
advanced_search(Author = "Johnathan~1")
#> - "advanced search" --------------------------- 239 packages in 0.003 seconds -
#>   #     package               version by                           @ title     
#>   1 100 lda                   1.4.2   Jonathan Chang              7y Collaps...
#>   2  99 mlmi                  1.1.1   Jonathan Bartlett           1y Maximum...
#>   3  97 handyplots            1.1.3   Jonathan Schwartz           4y Handy P...
#>   4  97 UnitCircle            0.1.3   Jonathan Berrisch           4y Check i...
#>   5  97 bootImpute            1.2.0   Jonathan Bartlett           2y Bootstr...
#>   6  97 chords                0.95.4  Jonathan Rosenblatt         6y Estimat...
#>   7  95 RSeed                 0.1.60  Claus Jonathan Fritzemeier  6y Borenst...
#>   8  94 rSHAPE                0.3.2   Jonathan Dench              3y Simulat...
#>   9  91 rasterKernelEstimates 1.0.2   Jonathan Lisic              1y Kernel ...
#>  10  91 tensor                1.5     Jonathan Rougier           10y Tensor ...