Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...
Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json
A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")

#> - "advanced search" ---------------------------- 59 packages in 0.001 seconds -
#>   #     package       version   by         @ title                             
#>   1 100 caTools       1.17.1.3  ORPHANED 19d Tools: moving window statistics...
#>   2 100 DAAG          1.22.1    ORPHANED 10M Data Analysis and Graphics Data...
#>   3 100 trimcluster   0.1.2.1   ORPHANED  1y Cluster Analysis with Trimming    
#>   4 100 climdex.pcic  1.1.9.1   ORPHANED  1y PCIC Implementation of Climdex ...
#>   5 100 distances     0.1.8     ORPHANED  3M Tools for Distance Metrics        
#>   6 100 mvcwt         1.3.1     ORPHANED  1y Wavelet Analysis of Multiple Ti...
#>   7 100 Funclustering 1.0.2     ORPHANED  1y Functional Data Clustering        
#>   8 100 SDMTools      1.1.221.2 ORPHANED 19d Species Distribution Modelling ...
#>   9 100 regress       1.3.15    ORPHANED  3y Gaussian Linear Models with Lin...
#>  10 100 smoothSurv    2.0.1     ORPHANED  1y Survival Regression with Smooth...

# Packages that match a certain field
advanced_search(Author = "Hester AND NOT Wickham")

#> - "advanced search" ---------------------------- 21 packages in 0.003 seconds -
#>   #     package        version by              @ title                         
#>   1 100 gmailr         1.0.0   Jim Hester     4M Access the 'Gmail' 'RESTful...
#>   2  98 primerTree     1.0.4   Jim Hester     1y Visually Assessing the Spec...
#>   3  93 knitrBootstrap 1.0.2   Jim Hester     2y 'knitr' Bootstrap Framework   
#>   4  88 glue           1.3.1   Jim Hester     9M Interpreted String Literals   
#>   5  85 types          1.0.0   Jim Hester     3y Type Annotations              
#>   6  83 bench          1.0.4   Jim Hester     3M High Precision Timing of R ...
#>   7  73 lintr          2.0.0   Jim Hester     3M A 'Linter' for R Code         
#>   8  71 dplyr.teradata 0.3.2   Koji Makiyama  7M A 'Teradata' Backend for 'd...
#>   9  70 pak            0.1.2   Gábor Csárdi  10M Another Approach to Package...
#>  10  65 mockery        0.4.2   Jim Hester     4M Mocking Library for R         

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")

#> - "advanced search" --------------------------- 250 packages in 0.025 seconds -
#>   #     package           version by                  @ title                  
#>   1 100 phd               0.1     Jesse Hemerik      6M Permutation Testing ...
#>   2  97 confSAM           0.2     Jesse Hemerik      2y Estimates and Bounds...
#>   3  85 permutes          1.0     Cesko C. Voeten    5M Permutation Tests fo...
#>   4  79 BurStMisc         1.1     Pat Burns          3y Burns Statistics Mis...
#>   5  77 fdrci             2.1     Joshua Millstein   3y Permutation-Based FD...
#>   6  74 phyloseqGraphTest 0.0.2   Julia Fukuyama     1y Graph-Based Permutat...
#>   7  74 resample          0.4     Tim Hesterberg     5y Resampling Functions   
#>   8  62 shiftR            1.5     Andrey A Shabalin  9M Fast Enrichment Anal...
#>   9  61 flipscores        0.2     Jesse Hemerik      1y Robust Testing in GLMs 
#>  10  57 ectotemp          0.1.2   Wouter Beukema     1y Quantitative Estimat...

# Packages that have a certain field
advanced_search("_exists_" = "URL")

#> - "advanced search" -------------------------- 7567 packages in 0.001 seconds -
#>   #     package       version by                    @ title                    
#>   1 100 fmlogcondens  1.0.2   Fabian Rathke        2y Fast Multivariate Log-...
#>   2 100 miceExt       1.1.0   Tobias Schumacher    2y Extension Package to '...
#>   3 100 rFTRLProximal 1.0.0   Yachen Yan           3y FTRL-Proximal Algorithm  
#>   4 100 softermax     0.3.0   Brian Connelly       2y Read Exported Data fro...
#>   5 100 cliqueMS      0.3.1   Oriol Senan Campos   8M Annotation of Isotopes...
#>   6 100 frite         0.1.0   Christopher Peralta  1y Functions Writing Func...
#>   7 100 PoloniexR     0.0.1   Jellen Vermeir       2y Interface to the Polon...
#>   8 100 projector     0.0.2   Michaël Benesty      2y Project Dense Vectors ...
#>   9 100 refimpact     1.0.0   Perry Stephenson     2y API Wrapper for the UK...
#>  10 100 MVA           1.0.6   Torsten Hothorn      4y An Introduction to App...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")

#> - "advanced search" -------------------------- 8284 packages in 0.001 seconds -
#>   #     package         version by                  @ title                    
#>   1 100 FUNLDA          1.1     Daniel Backenroth  2y Genomic Latent Dirichl...
#>   2 100 NRejections     1.0.0   Maya B. Mathur     1y Metrics for Multiple T...
#>   3 100 miniparquet     0.1.2   Hannes Mühleisen   2M Read Parquet Files       
#>   4 100 CpDyna          1.0     Margot Selosse     1y Change Point Detection...
#>   5 100 icsurvROC       0.1.0   Yunro Chung        1y Interval Censored Surv...
#>   6 100 RWeightedKmeans 1.1.0   Alexandre Gondeau  1y Weighted Object k-Mean...
#>   7 100 SSL             0.1     Junxiang Wang      4y Semi-Supervised Learning 
#>   8 100 SeqKat          0.0.6   Paul C. Boutros    2y Detection of Kataegis    
#>   9 100 AssocTests      0.0.4   Lin Wang           2y Genetic Association St...
#>  10 100 xlm             0.1.1   Oliver Frost       1y A 'Stellar' Client       

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)

#> - "advanced search" ------------------------------ 8284 packages in 0 seconds -
#>   #     package         version by                  @ title                    
#>   1 100 FUNLDA          1.1     Daniel Backenroth  2y Genomic Latent Dirichl...
#>   2 100 NRejections     1.0.0   Maya B. Mathur     1y Metrics for Multiple T...
#>   3 100 miniparquet     0.1.2   Hannes Mühleisen   2M Read Parquet Files       
#>   4 100 CpDyna          1.0     Margot Selosse     1y Change Point Detection...
#>   5 100 icsurvROC       0.1.0   Yunro Chung        1y Interval Censored Surv...
#>   6 100 RWeightedKmeans 1.1.0   Alexandre Gondeau  1y Weighted Object k-Mean...
#>   7 100 SSL             0.1     Junxiang Wang      4y Semi-Supervised Learning 
#>   8 100 SeqKat          0.0.6   Paul C. Boutros    2y Detection of Kataegis    
#>   9 100 AssocTests      0.0.4   Lin Wang           2y Genetic Association St...
#>  10 100 xlm             0.1.1   Oliver Frost       1y A 'Stellar' Client       

# Regular expressions
advanced_search(Author = "/Joh?nathan/")

#> - "advanced search" --------------------------- 188 packages in 0.001 seconds -
#>   #     package         version  by                         @ title            
#>   1 100 raster          3.0.7    Robert J. Hijmans         3M Geographic Dat...
#>   2 100 xtable          1.8.4    David Scott               8M Export Tables ...
#>   3 100 gdalUtils       2.0.1.14 Jonathan Asher Greenberg  2y Wrappers for t...
#>   4 100 knitr           1.26     Yihui Xie                 1M A General-Purp...
#>   5 100 tableone        0.10.0   Kazuki Yoshida           10M Create 'Table ...
#>   6 100 bridgesampling  0.7.2    Quentin F. Gronau         5M Bridge Samplin...
#>   7 100 fivethirtyeight 0.5.0    Albert Y. Kim             5M Data and Code ...
#>   8 100 smcfcs          1.4.0    Jonathan Bartlett         9M Multiple Imput...
#>   9 100 tscount         1.4.1    Tobias Liboschik          2y Analysis of Co...
#>  10 100 hergm           4.1.5    Michael Schweinberger     1M Hierarchical E...

# Fuzzy search
advanced_search(Author = "Johnathan~1")

#> - "advanced search" --------------------------- 189 packages in 0.002 seconds -
#>   #     package       version by                                        @ title
#>   1 100 lda           1.4.2   Jonathan Chang                           4y Co...
#>   2 100 UScancer      0.1.2   Jonathan Lee                             5y Cr...
#>   3 100 mlmi          1.0.0   Jonathan Bartlett                        5M Ma...
#>   4  97 rSHAPE        0.3.2   Jonathan Dench                           5M Si...
#>   5  95 UnitCircle    0.1.3   Jonathan Berrisch                        1y Ch...
#>   6  95 handyplots    1.1.3   Jonathan Schwartz                       11M Ha...
#>   7  95 chords        0.95.4  Jonathan Rosenblatt                      3y Es...
#>   8  95 bootImpute    1.0.0   Jonathan Bartlett                        5M Bo...
#>   9  94 geomapdata    1.0.4   Jonathan M. Lees<jonathan.lees@unc.edu>  7y Da...
#>  10  94 spatial.tools 1.6.0   Jonathan Asher Greenberg                 2y R ...