Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...
Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json
A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")

#> - "advanced search" ---------------------------- 37 packages in 0.001 seconds -
#>   #     package        version   by         @ title                            
#>   1 100 distances      0.1.8     ORPHANED  2y Tools for Distance Metrics       
#>   2 100 mvcwt          1.3.1     ORPHANED  2y Wavelet Analysis of Multiple T...
#>   3 100 Funclustering  1.0.2     ORPHANED  2y Functional Data Clustering       
#>   4  96 SDMTools       1.1.221.2 ORPHANED  1y Species Distribution Modelling...
#>   5  96 NanoStringNorm 1.2.1.1   ORPHANED  7M Normalize NanoString miRNA and...
#>   6  96 datadr         0.8.6.1   ORPHANED  3y Divide and Recombine for Large...
#>   7  96 nhlscrapr      1.8.1     ORPHANED  4y Compiling the NHL Real Time Sc...
#>   8  86 ggmap          3.0.0     ORPHANED  2y Spatial Visualization with ggp...
#>   9  86 gWidgetstcltk  0.0.55.1  ORPHANED  1y Toolkit implementation of gWid...
#>  10  86 gWidgets       0.0.54.2  ORPHANED  1y gWidgets API for Building Tool...

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")

#> - "advanced search" ---------------------------- 17 packages in 0.001 seconds -
#>   #     package     version by             @ title                             
#>   1 100 pkgbuild    1.2.0   Jim Hester    4M Find Tools Needed to Build R Pa...
#>   2  86 devtools    2.4.0   Jim Hester    3d Tools to Make Developing R Pack...
#>   3  81 sessioninfo 1.1.1   Gábor Csárdi  2y R Session Information             
#>   4  70 memoise     2.0.0   Jim Hester    2M Memoisation of Functions          
#>   5  68 xml2        1.3.2   Jim Hester    1y Parse XML                         
#>   6  64 odbc        1.3.2   Jim Hester    7d Connect to ODBC Compatible Data...
#>   7  62 withr       2.4.1   Jim Hester    2M Run Code 'With' Temporarily Mod...
#>   8  61 pkgload     1.2.1   Jim Hester    4d Simulate Package Installation a...
#>   9  60 fs          1.5.0   Jim Hester    8M Cross-Platform File System Oper...
#>  10  60 remotes     2.3.0   Jim Hester    9d R Package Installation from Rem...

advanced_search("Author: Hester AND Author: Wickham")

#> - "advanced search" ---------------------------- 17 packages in 0.001 seconds -
#>   #     package     version by             @ title                             
#>   1 100 pkgbuild    1.2.0   Jim Hester    4M Find Tools Needed to Build R Pa...
#>   2  86 devtools    2.4.0   Jim Hester    3d Tools to Make Developing R Pack...
#>   3  81 sessioninfo 1.1.1   Gábor Csárdi  2y R Session Information             
#>   4  70 memoise     2.0.0   Jim Hester    2M Memoisation of Functions          
#>   5  68 xml2        1.3.2   Jim Hester    1y Parse XML                         
#>   6  64 odbc        1.3.2   Jim Hester    7d Connect to ODBC Compatible Data...
#>   7  62 withr       2.4.1   Jim Hester    2M Run Code 'With' Temporarily Mod...
#>   8  61 pkgload     1.2.1   Jim Hester    4d Simulate Package Installation a...
#>   9  60 fs          1.5.0   Jim Hester    8M Cross-Platform File System Oper...
#>  10  60 remotes     2.3.0   Jim Hester    9d R Package Installation from Rem...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")

#> - "advanced search" ---------------------------- 26 packages in 0.004 seconds -
#>   #     package        version by              @ title                         
#>   1 100 gmailr         1.0.0   Jim Hester     2y Access the 'Gmail' 'RESTful...
#>   2  93 primerTree     1.0.5   Jim Hester     1y Visually Assessing the Spec...
#>   3  89 knitrBootstrap 1.0.2   Jim Hester     3y 'knitr' Bootstrap Framework   
#>   4  86 types          1.0.0   Jim Hester     4y Type Annotations              
#>   5  84 glue           1.4.2   Jim Hester     8M Interpreted String Literals   
#>   6  72 pak            0.1.2.1 Gábor Csárdi   5M Another Approach to Package...
#>   7  70 dplyr.teradata 0.4.1   Koji Makiyama  5M A 'Teradata' Backend for 'd...
#>   8  66 lintr          2.0.1   Jim Hester     1y A 'Linter' for R Code         
#>   9  65 decor          1.0.0   Jim Hester     9M Retrieve Code Decorations     
#>  10  64 mockery        0.4.2   Jim Hester     2y Mocking Library for R         

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")

#> - "advanced search" ---------------------------- 17 packages in 0.007 seconds -
#>   #     package     version by             @ title                             
#>   1 100 pkgbuild    1.2.0   Jim Hester    4M Find Tools Needed to Build R Pa...
#>   2  86 devtools    2.4.0   Jim Hester    3d Tools to Make Developing R Pack...
#>   3  81 sessioninfo 1.1.1   Gábor Csárdi  2y R Session Information             
#>   4  70 memoise     2.0.0   Jim Hester    2M Memoisation of Functions          
#>   5  68 xml2        1.3.2   Jim Hester    1y Parse XML                         
#>   6  67 pkgload     1.2.1   Jim Hester    4d Simulate Package Installation a...
#>   7  64 odbc        1.3.2   Jim Hester    7d Connect to ODBC Compatible Data...
#>   8  64 fs          1.5.0   Jim Hester    8M Cross-Platform File System Oper...
#>   9  62 withr       2.4.1   Jim Hester    2M Run Code 'With' Temporarily Mod...
#>  10  62 remotes     2.3.0   Jim Hester    9d R Package Installation from Rem...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")

#> - "advanced search" ---------------------------- 26 packages in 0.004 seconds -
#>   #     package        version by              @ title                         
#>   1 100 gmailr         1.0.0   Jim Hester     2y Access the 'Gmail' 'RESTful...
#>   2  93 primerTree     1.0.5   Jim Hester     1y Visually Assessing the Spec...
#>   3  89 knitrBootstrap 1.0.2   Jim Hester     3y 'knitr' Bootstrap Framework   
#>   4  86 types          1.0.0   Jim Hester     4y Type Annotations              
#>   5  84 glue           1.4.2   Jim Hester     8M Interpreted String Literals   
#>   6  72 pak            0.1.2.1 Gábor Csárdi   5M Another Approach to Package...
#>   7  70 dplyr.teradata 0.4.1   Koji Makiyama  5M A 'Teradata' Backend for 'd...
#>   8  66 lintr          2.0.1   Jim Hester     1y A 'Linter' for R Code         
#>   9  65 decor          1.0.0   Jim Hester     9M Retrieve Code Decorations     
#>  10  64 mockery        0.4.2   Jim Hester     2y Mocking Library for R         

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")

#> - "advanced search" --------------------------- 340 packages in 0.018 seconds -
#>   #     package           version by                         @ title           
#>   1 100 phd               0.1     Jesse Hemerik             2y Permutation T...
#>   2  97 confSAM           0.2     Jesse Hemerik             3y Estimates and...
#>   3  86 permutes          2.0.1   Cesko C. Voeten           1M Permutation T...
#>   4  80 BurStMisc         1.1     Pat Burns                 5y Burns Statist...
#>   5  78 fdrci             2.1     Joshua Millstein          4y Permutation-B...
#>   6  76 packager          1.11.1  Andreas Dominik Cullmann 20d Create, Build...
#>   7  76 USP               0.1.2   Thomas B. Berrett         2M U-Statistic P...
#>   8  76 phyloseqGraphTest 0.1.0   Julia Fukuyama            1y Graph-Based P...
#>   9  75 resample          0.4     Tim Hesterberg            6y Resampling Fu...
#>  10  69 IMEC              0.2.0   Maximilian Maier          4M Ising Model o...

# Packages that have a certain field
advanced_search("_exists_" = "URL")

#> - "advanced search" -------------------------- 9554 packages in 0.005 seconds -
#>   #     package       version  by                     @ title                  
#>   1 100 mangoTraining 1.1      Karina Marks         10M Mango Solutions Trai...
#>   2 100 swatches      0.5.0    Bob Rudis             3y Read, Inspect, and M...
#>   3 100 meditate      0.1.3    Jason C. Fisher       1y Meditation Timer       
#>   4 100 exploratory   0.1.53   Jin Kim              19d Conduct Exploratory ...
#>   5 100 gettz         0.0.4    Dirk Eddelbuettel     1y Get the Timezone Inf...
#>   6 100 rtf           0.4.14.1 Michael E. Schaffer   1y Rich Text Format (RT...
#>   7 100 LFDR.MLE      1.0.1    M. Padilla            2y Estimation of the Lo...
#>   8 100 Cascade       2.0      Frederic Bertrand    22d Selection, Reverse-E...
#>   9 100 SoupX         1.5.0    Matthew Daniel Young  3M Single Cell mRNA Sou...
#>  10 100 vaultr        1.0.2    Rich FitzJohn         2y Vault Client for Sec...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")

#> - "advanced search" -------------------------- 9498 packages in 0.001 seconds -
#>   #     package    version    by                             @ title           
#>   1 100 ScoreGGUM  1.0        David R. King                 6y Score Persons...
#>   2 100 ph2mult    0.1.1      Yalin Zhu                     4y Phase II Clin...
#>   3 100 KRMM       1.0        Laval Jacquin                 4y Kernel Ridge ...
#>   4 100 MullerPlot 0.1.2      Farnoush Farahpour            4y Generates Mul...
#>   5 100 UNCLES     2.0        Basel Abu-Jamous              5y Unification o...
#>   6 100 NBShiny3   0.1.0      Kartikeya Bolar               8M Interactive D...
#>   7 100 PASWR2     1.0.2      Alan T. Arnholt               5y Probability a...
#>   8 100 optimr     2019.12.16 John C Nash                   1y A Replacement...
#>   9 100 CommEcol   1.7.1      Adriano Sanches Melo         25d Community Eco...
#>  10 100 AROC       1.0.3      Maria Xose Rodriguez-Alvarez 24d Covariate-Adj...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)

#> - "advanced search" ------------------------------ 9498 packages in 0 seconds -
#>   #     package    version    by                             @ title           
#>   1 100 ScoreGGUM  1.0        David R. King                 6y Score Persons...
#>   2 100 ph2mult    0.1.1      Yalin Zhu                     4y Phase II Clin...
#>   3 100 KRMM       1.0        Laval Jacquin                 4y Kernel Ridge ...
#>   4 100 MullerPlot 0.1.2      Farnoush Farahpour            4y Generates Mul...
#>   5 100 UNCLES     2.0        Basel Abu-Jamous              5y Unification o...
#>   6 100 NBShiny3   0.1.0      Kartikeya Bolar               8M Interactive D...
#>   7 100 PASWR2     1.0.2      Alan T. Arnholt               5y Probability a...
#>   8 100 optimr     2019.12.16 John C Nash                   1y A Replacement...
#>   9 100 CommEcol   1.7.1      Adriano Sanches Melo         25d Community Eco...
#>  10 100 AROC       1.0.3      Maria Xose Rodriguez-Alvarez 24d Covariate-Adj...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")

#> - "advanced search" --------------------------- 220 packages in 0.001 seconds -
#>   #     package         version by                         @ title             
#>   1 100 xtable          1.8.4   David Scott               2y Export Tables t...
#>   2 100 raster          3.4.5   Robert J. Hijmans         5M Geographic Data...
#>   3 100 knitr           1.31    Yihui Xie                 2M A General-Purpo...
#>   4 100 tableone        0.12.0  Kazuki Yoshida            9M Create 'Table 1...
#>   5 100 bridgesampling  1.0.0   Quentin F. Gronau         1y Bridge Sampling...
#>   6 100 gdalUtils       2.0.3.2 Jonathan Asher Greenberg  1y Wrappers for th...
#>   7 100 tscount         1.4.3   Tobias Liboschik          7M Analysis of Cou...
#>   8 100 rmcorr          0.4.3   Laura R. Marusich         9d Repeated Measur...
#>   9 100 fivethirtyeight 0.6.1   Albert Y. Kim             8M Data and Code B...
#>  10 100 smcfcs          1.5.0   Jonathan Bartlett         9d Multiple Imputa...

# Fuzzy search
advanced_search(Author = "Johnathan~1")

#> - "advanced search" --------------------------- 221 packages in 0.002 seconds -
#>   #     package       version by                                        @ title
#>   1 100 lda           1.4.2   Jonathan Chang                           5y Co...
#>   2 100 UScancer      0.1.2   Jonathan Lee                             7y Cr...
#>   3 100 mlmi          1.0.0   Jonathan Bartlett                        2y Ma...
#>   4  96 handyplots    1.1.3   Jonathan Schwartz                        2y Ha...
#>   5  96 UnitCircle    0.1.3   Jonathan Berrisch                        3y Ch...
#>   6  95 RSeed         0.1.60  Claus Jonathan Fritzemeier               5y Bo...
#>   7  95 rSHAPE        0.3.2   Jonathan Dench                           2y Si...
#>   8  95 geomapdata    1.0.4   Jonathan M. Lees<jonathan.lees@unc.edu>  8y Da...
#>   9  95 spatial.tools 1.6.2   Jonathan Asher Greenberg                 1y R ...
#>  10  94 bootImpute    1.2.0   Jonathan Bartlett                        9d Bo...