Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...
Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json
A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

Value

Search hits.

Examples

 if (pingr::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({
# All orphaned packages
advanced_search(Maintainer = "ORPHANED")

# Packages that match a certain field
advanced_search(Author = "Hester AND NOT Wickham")

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")

# Packages that have a certain field
advanced_search("_exists_" = "URL")

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)

# Regular expressions
advanced_search(Author = "/Joh?nathan/")

# Fuzzy search
advanced_search(Author = "Johnathan~1")
 })

#> > advanced_search(Maintainer = "ORPHANED")
#> - "advanced search" ---------------------------- 58 packages in 0.001 seconds -
#>   #     package      version   by         @ title                              
#>   1 100 SDMTools     1.1.221.1 ORPHANED  7M Species Distribution Modelling T...
#>   2 100 regress      1.3.15    ORPHANED  3y Gaussian Linear Models with Line...
#>   3 100 DirichletReg 0.6.3.1   ORPHANED 10M Dirichlet Regression in R          
#>   4 100 smoothSurv   2.0.1     ORPHANED  1y Survival Regression with Smoothe...
#>   5 100 bfp          0.0.40    ORPHANED 10M Bayesian Fractional Polynomials    
#>   6 100 datadr       0.8.6.1   ORPHANED  1y Divide and Recombine for Large, ...
#>   7 100 nhlscrapr    1.8.1     ORPHANED  3y Compiling the NHL Real Time Scor...
#>   8  96 caTools      1.17.1.2  ORPHANED  9M Tools: moving window statistics,...
#>   9  96 DAAG         1.22.1    ORPHANED  9M Data Analysis and Graphics Data ...
#>  10  96 highlight    0.4.7.2   ORPHANED  2y Syntax Highlighter                 
#> > advanced_search(Author = "Hester AND NOT Wickham")
#> - "advanced search" ---------------------------- 21 packages in 0.003 seconds -
#>   #     package        version by              @ title                         
#>   1 100 gmailr         1.0.0   Jim Hester     3M Access the 'Gmail' 'RESTful...
#>   2  98 primerTree     1.0.4   Jim Hester     1y Visually Assessing the Spec...
#>   3  93 knitrBootstrap 1.0.2   Jim Hester     1y 'knitr' Bootstrap Framework   
#>   4  88 glue           1.3.1   Jim Hester     8M Interpreted String Literals   
#>   5  85 types          1.0.0   Jim Hester     3y Type Annotations              
#>   6  82 bench          1.0.4   Jim Hester     3M High Precision Timing of R ...
#>   7  73 lintr          2.0.0   Jim Hester     2M A 'Linter' for R Code         
#>   8  71 dplyr.teradata 0.3.2   Koji Makiyama  6M A 'Teradata' Backend for 'd...
#>   9  70 pak            0.1.2   Gábor Csárdi   9M Another Approach to Package...
#>  10  64 mockery        0.4.2   Jim Hester     3M Mocking Library for R         
#> > advanced_search("permutation test AND NOT License: GPL OR GNU")
#> - "advanced search" --------------------------- 247 packages in 0.022 seconds -
#>   #     package           version by                  @ title                  
#>   1 100 phd               0.1     Jesse Hemerik      5M Permutation Testing ...
#>   2  97 confSAM           0.2     Jesse Hemerik      2y Estimates and Bounds...
#>   3  85 permutes          1.0     Cesko C. Voeten    4M Permutation Tests fo...
#>   4  78 BurStMisc         1.1     Pat Burns          3y Burns Statistics Mis...
#>   5  77 fdrci             2.1     Joshua Millstein   3y Permutation-Based FD...
#>   6  75 phyloseqGraphTest 0.0.2   Julia Fukuyama     1y Graph-Based Permutat...
#>   7  73 resample          0.4     Tim Hesterberg     5y Resampling Functions   
#>   8  62 shiftR            1.5     Andrey A Shabalin  8M Fast Enrichment Anal...
#>   9  61 flipscores        0.2     Jesse Hemerik      1y Robust Testing in GLMs 
#>  10  57 ectotemp          0.1.2   Wouter Beukema     1y Quantitative Estimat...
#> > advanced_search(`_exists_` = "URL")
#> - "advanced search" -------------------------- 7451 packages in 0.001 seconds -
#>   #     package      version  by                     @ title                   
#>   1 100 PBSmapping   2.72.1   Rowan Haigh           8M Mapping Fisheries Dat...
#>   2 100 reactlog     1.0.0    Barret Schloerke      8M Reactivity Visualizer...
#>   3 100 snowFT       1.6.0    Hana Sevcikova        2y Fault Tolerant Simple...
#>   4 100 Rsymphony    0.1.28   Kurt Hornik           2y SYMPHONY in R           
#>   5 100 trtf         0.3.6    Torsten Hothorn       2M Transformation Trees ...
#>   6 100 captioner    2.2.3    Letaw Alathea         4y Numbers Figures and C...
#>   7 100 eva          0.2.5    Brian Bader           1y Extreme Value Analysi...
#>   8 100 groupdata2   1.1.2    Ludvig Renbo Olsen    4M Creating Groups from ...
#>   9 100 AWR          1.11.189 Gergely Daroczi       2y 'AWS' Java 'SDK' for R  
#>  10 100 dynamicGraph 0.2.2.6  Jens Henrik Badsberg 10y dynamicGraph            
#> > advanced_search("NOT _exists_: URL")
#> - "advanced search" -------------------------- 8243 packages in 0.001 seconds -
#>   #     package                version by                           @ title    
#>   1 100 NRejections            1.0.0   Maya B. Mathur              1y Metric...
#>   2 100 lmom                   2.8     J. R. M. Hosking            8M L-Moments
#>   3 100 mvShapiroTest          1.0     Elizabeth Gonzalez Estrada  6y Genera...
#>   4 100 Rfit                   0.23.0  John Kloke                  3y Rank E...
#>   5 100 RcmdrPlugin.EACSPIR    0.2.2   David Leiva                 5y Plugin...
#>   6 100 landsat                1.0.8   Sarah Goslee                7y Radiom...
#>   7 100 RcmdrPlugin.FuzzyClust 1.1     Achmad Fauzi Bagus F        3y R Comm...
#>   8 100 geofacet               0.1.10  Ryan Hafen                  8M 'ggplo...
#>   9 100 bimets                 1.4.2   Andrea Luciani              1M Time S...
#>  10 100 funtimes               6.1     Vyacheslav Lyubchich        7M Functi...
#> > query <- "{\n\"query\": {\n  \"bool\": {\n    \"must_not\": {\n      \"exists\": {\n        \"field\": \"URL\"\n      }\n    }\n  }\n}\n}"
#> > advanced_search(json = query)
#> - "advanced search" ------------------------------ 8243 packages in 0 seconds -
#>   #     package                version by                           @ title    
#>   1 100 NRejections            1.0.0   Maya B. Mathur              1y Metric...
#>   2 100 lmom                   2.8     J. R. M. Hosking            8M L-Moments
#>   3 100 mvShapiroTest          1.0     Elizabeth Gonzalez Estrada  6y Genera...
#>   4 100 Rfit                   0.23.0  John Kloke                  3y Rank E...
#>   5 100 RcmdrPlugin.EACSPIR    0.2.2   David Leiva                 5y Plugin...
#>   6 100 landsat                1.0.8   Sarah Goslee                7y Radiom...
#>   7 100 RcmdrPlugin.FuzzyClust 1.1     Achmad Fauzi Bagus F        3y R Comm...
#>   8 100 geofacet               0.1.10  Ryan Hafen                  8M 'ggplo...
#>   9 100 bimets                 1.4.2   Andrea Luciani              1M Time S...
#>  10 100 funtimes               6.1     Vyacheslav Lyubchich        7M Functi...
#> > advanced_search(Author = "/Joh?nathan/")
#> - "advanced search" --------------------------- 185 packages in 0.001 seconds -
#>   #     package         version  by                         @ title            
#>   1 100 RTriangle       1.6.0.10 David C. Sterratt         2y Triangle - A 2...
#>   2 100 Pijavski        1.0      Gleb Beliakov             4y Global Univari...
#>   3 100 raster          3.0.7    Robert J. Hijmans         2M Geographic Dat...
#>   4 100 knitr           1.26     Yihui Xie                 9d A General-Purp...
#>   5 100 xtable          1.8.4    David Scott               7M Export Tables ...
#>   6 100 gdalUtils       2.0.1.14 Jonathan Asher Greenberg  2y Wrappers for t...
#>   7 100 bridgesampling  0.7.2    Quentin F. Gronau         4M Bridge Samplin...
#>   8 100 tableone        0.10.0   Kazuki Yoshida            9M Create 'Table ...
#>   9 100 fivethirtyeight 0.5.0    Albert Y. Kim             4M Data and Code ...
#>  10 100 Rwave           2.4.8    Jonathan M. Lees          2y Time-Frequency...
#> > advanced_search(Author = "Johnathan~1")
#> - "advanced search" --------------------------- 186 packages in 0.002 seconds -
#>   #     package       version by                                        @ title
#>   1 100 lda           1.4.2   Jonathan Chang                           4y Co...
#>   2 100 UScancer      0.1.2   Jonathan Lee                             5y Cr...
#>   3 100 mlmi          1.0.0   Jonathan Bartlett                        4M Ma...
#>   4  97 rSHAPE        0.3.2   Jonathan Dench                           4M Si...
#>   5  95 handyplots    1.1.3   Jonathan Schwartz                       10M Ha...
#>   6  95 UnitCircle    0.1.3   Jonathan Berrisch                        1y Ch...
#>   7  95 chords        0.95.4  Jonathan Rosenblatt                      3y Es...
#>   8  95 bootImpute    1.0.0   Jonathan Bartlett                        4M Bo...
#>   9  94 geomapdata    1.0.4   Jonathan M. Lees<jonathan.lees@unc.edu>  7y Da...
#>  10  94 spatial.tools 1.6.0   Jonathan Asher Greenberg                 2y R ...