Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...
Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json
A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")

#> - "advanced search" ---------------------------- 39 packages in 0.001 seconds -
#>   #     package        version   by         @ title                            
#>   1 100 distances      0.1.8     ORPHANED  1y Tools for Distance Metrics       
#>   2 100 mvcwt          1.3.1     ORPHANED  2y Wavelet Analysis of Multiple T...
#>   3 100 Funclustering  1.0.2     ORPHANED  2y Functional Data Clustering       
#>   4  96 SDMTools       1.1.221.2 ORPHANED 11M Species Distribution Modelling...
#>   5  96 NanoStringNorm 1.2.1.1   ORPHANED 29d Normalize NanoString miRNA and...
#>   6  96 datadr         0.8.6.1   ORPHANED  2y Divide and Recombine for Large...
#>   7  96 nhlscrapr      1.8.1     ORPHANED  4y Compiling the NHL Real Time Sc...
#>   8  86 ggmap          3.0.0     ORPHANED  2y Spatial Visualization with ggp...
#>   9  86 gWidgetstcltk  0.0.55.1  ORPHANED  5M Toolkit implementation of gWid...
#>  10  86 gWidgets       0.0.54.2  ORPHANED  5M gWidgets API for Building Tool...

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")

#> - "advanced search" ---------------------------- 17 packages in 0.001 seconds -
#>   #     package     version by             @ title                             
#>   1 100 pkgbuild    1.1.0   Jim Hester    3M Find Tools Needed to Build R Pa...
#>   2  85 memoise     1.1.0   Jim Hester    3y Memoisation of Functions          
#>   3  80 sessioninfo 1.1.1   Gábor Csárdi  2y R Session Information             
#>   4  71 odbc        1.2.3   Jim Hester    4M Connect to ODBC Compatible Data...
#>   5  67 xml2        1.3.2   Jim Hester    6M Parse XML                         
#>   6  62 devtools    2.3.2   Jim Hester   22d Tools to Make Developing R Pack...
#>   7  62 withr       2.3.0   Jim Hester   18d Run Code 'With' Temporarily Mod...
#>   8  61 pkgload     1.1.0   Jim Hester    4M Simulate Package Installation a...
#>   9  59 fs          1.5.0   Jim Hester    2M Cross-Platform File System Oper...
#>  10  59 remotes     2.2.0   Jim Hester    3M R Package Installation from Rem...

advanced_search("Author: Hester AND Author: Wickham")

#> - "advanced search" ---------------------------- 17 packages in 0.001 seconds -
#>   #     package     version by             @ title                             
#>   1 100 pkgbuild    1.1.0   Jim Hester    3M Find Tools Needed to Build R Pa...
#>   2  85 memoise     1.1.0   Jim Hester    3y Memoisation of Functions          
#>   3  80 sessioninfo 1.1.1   Gábor Csárdi  2y R Session Information             
#>   4  71 odbc        1.2.3   Jim Hester    4M Connect to ODBC Compatible Data...
#>   5  67 xml2        1.3.2   Jim Hester    6M Parse XML                         
#>   6  62 devtools    2.3.2   Jim Hester   22d Tools to Make Developing R Pack...
#>   7  62 withr       2.3.0   Jim Hester   18d Run Code 'With' Temporarily Mod...
#>   8  61 pkgload     1.1.0   Jim Hester    4M Simulate Package Installation a...
#>   9  59 fs          1.5.0   Jim Hester    2M Cross-Platform File System Oper...
#>  10  59 remotes     2.2.0   Jim Hester    3M R Package Installation from Rem...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester", Author = "NOT Wickham")

#> Error in do_query(qstr, server, port, from, size): search server failure

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")

#> - "advanced search" ---------------------------- 17 packages in 0.007 seconds -
#>   #     package     version by             @ title                             
#>   1 100 pkgbuild    1.1.0   Jim Hester    3M Find Tools Needed to Build R Pa...
#>   2  85 memoise     1.1.0   Jim Hester    3y Memoisation of Functions          
#>   3  80 sessioninfo 1.1.1   Gábor Csárdi  2y R Session Information             
#>   4  71 odbc        1.2.3   Jim Hester    4M Connect to ODBC Compatible Data...
#>   5  67 xml2        1.3.2   Jim Hester    6M Parse XML                         
#>   6  67 pkgload     1.1.0   Jim Hester    4M Simulate Package Installation a...
#>   7  65 devtools    2.3.2   Jim Hester   22d Tools to Make Developing R Pack...
#>   8  64 fs          1.5.0   Jim Hester    2M Cross-Platform File System Oper...
#>   9  62 withr       2.3.0   Jim Hester   18d Run Code 'With' Temporarily Mod...
#>  10  61 remotes     2.2.0   Jim Hester    3M R Package Installation from Rem...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")

#> - "advanced search" ---------------------------- 25 packages in 0.004 seconds -
#>   #     package        version by              @ title                         
#>   1 100 gmailr         1.0.0   Jim Hester     1y Access the 'Gmail' 'RESTful...
#>   2  95 primerTree     1.0.5   Jim Hester     9M Visually Assessing the Spec...
#>   3  89 knitrBootstrap 1.0.2   Jim Hester     2y 'knitr' Bootstrap Framework   
#>   4  86 types          1.0.0   Jim Hester     4y Type Annotations              
#>   5  85 glue           1.4.2   Jim Hester     1M Interpreted String Literals   
#>   6  71 pak            0.1.2   Gábor Csárdi   2y Another Approach to Package...
#>   7  69 dplyr.teradata 0.3.2   Koji Makiyama  1y A 'Teradata' Backend for 'd...
#>   8  66 lintr          2.0.1   Jim Hester     8M A 'Linter' for R Code         
#>   9  65 decor          1.0.0   Jim Hester     3M Retrieve Code Decorations     
#>  10  63 mockery        0.4.2   Jim Hester     1y Mocking Library for R         

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")

#> - "advanced search" ---------------------------- 299 packages in 0.02 seconds -
#>   #     package           version by                         @ title           
#>   1 100 phd               0.1     Jesse Hemerik             1y Permutation T...
#>   2  97 confSAM           0.2     Jesse Hemerik             3y Estimates and...
#>   3  85 permutes          1.0     Cesko C. Voeten           1y Permutation T...
#>   4  80 BurStMisc         1.1     Pat Burns                 4y Burns Statist...
#>   5  77 fdrci             2.1     Joshua Millstein          4y Permutation-B...
#>   6  75 packager          1.5.0   Andreas Dominik Cullmann 12d Create, Build...
#>   7  75 phyloseqGraphTest 0.1.0   Julia Fukuyama            8M Graph-Based P...
#>   8  74 resample          0.4     Tim Hesterberg            5y Resampling Fu...
#>   9  63 shiftR            1.5     Andrey A Shabalin         2y Fast Enrichme...
#>  10  60 mindr             1.2.3   Peng Zhao                 7M Convert Files...

# Packages that have a certain field
advanced_search("_exists_" = "URL")

#> - "advanced search" -------------------------- 8740 packages in 0.001 seconds -
#>   #     package       version by               @ title                         
#>   1 100 arrayhelpers  1.1.0   C. Beleites     8M Convenience Functions for A...
#>   2 100 pdftables     0.1     Eric Persson    5y Programmatic Conversion of ...
#>   3 100 wpp2015       1.1.2   Hana Sevcikova  2y World Population Prospects ...
#>   4 100 outreg        0.2.2   Kota Mori       4y Regression Table for Public...
#>   5 100 treeplyr      0.1.10  Josef Uyeda    23d 'dplyr' Functionality for M...
#>   6 100 zeroEQpart    0.1.0   Dan Richard     2y Zero Order vs (Semi) Partia...
#>   7 100 RSurveillance 0.2.1   Rohan Sadler    3M Design and Analysis of Dise...
#>   8 100 cellranger    1.1.0   Jennifer Bryan  4y Translate Spreadsheet Cell ...
#>   9 100 sp            1.4.4   Edzer Pebesma   3d Classes and Methods for Spa...
#>  10 100 commonmark    1.7     Jeroen Ooms     2y High Performance CommonMark...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")

#> - "advanced search" -------------------------- 9003 packages in 0.001 seconds -
#>   #     package       version by                     @ title                   
#>   1 100 circular      0.4.93  Claudio Agostinelli   3y Circular Statistics     
#>   2 100 modest        0.3.1   Philip Pallmann       3y Model-Based Dose-Esca...
#>   3 100 tmvmixnorm    1.1.1   Ting Fung (Ralph) Ma 22d Sampling from Truncat...
#>   4 100 SimCop        0.7.0   Berwin A. Turlach     3y Simulate from Arbitra...
#>   5 100 velociraptr   1.1.0   Andrew A Zaffos       1y Fossil Analysis         
#>   6 100 tsqn          1.0.0   Higor Cotta           4y Applications of the Q...
#>   7 100 smoothROCtime 0.1.0   Susana Diaz-Coto      2y Smooth Time-Dependent...
#>   8 100 rrat          1.0.0   Yi He                 1y Robust Regression wit...
#>   9 100 blkbox        1.0     Boris Guennewig       4y Data Exploration with...
#>  10 100 gridExtra     2.3     Baptiste Auguie       3y Miscellaneous Functio...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)

#> - "advanced search" ------------------------------ 9003 packages in 0 seconds -
#>   #     package       version by                     @ title                   
#>   1 100 circular      0.4.93  Claudio Agostinelli   3y Circular Statistics     
#>   2 100 modest        0.3.1   Philip Pallmann       3y Model-Based Dose-Esca...
#>   3 100 tmvmixnorm    1.1.1   Ting Fung (Ralph) Ma 22d Sampling from Truncat...
#>   4 100 SimCop        0.7.0   Berwin A. Turlach     3y Simulate from Arbitra...
#>   5 100 velociraptr   1.1.0   Andrew A Zaffos       1y Fossil Analysis         
#>   6 100 tsqn          1.0.0   Higor Cotta           4y Applications of the Q...
#>   7 100 smoothROCtime 0.1.0   Susana Diaz-Coto      2y Smooth Time-Dependent...
#>   8 100 rrat          1.0.0   Yi He                 1y Robust Regression wit...
#>   9 100 blkbox        1.0     Boris Guennewig       4y Data Exploration with...
#>  10 100 gridExtra     2.3     Baptiste Auguie       3y Miscellaneous Functio...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")

#> - "advanced search" --------------------------- 202 packages in 0.002 seconds -
#>   #     package         version by                         @ title             
#>   1 100 xtable          1.8.4   David Scott               1y Export Tables t...
#>   2 100 knitr           1.30    Yihui Xie                18d A General-Purpo...
#>   3 100 raster          3.3.13  Robert J. Hijmans         3M Geographic Data...
#>   4 100 bridgesampling  1.0.0   Quentin F. Gronau         8M Bridge Sampling...
#>   5 100 tableone        0.12.0  Kazuki Yoshida            3M Create 'Table 1...
#>   6 100 fivethirtyeight 0.6.1   Albert Y. Kim             2M Data and Code B...
#>   7 100 gdalUtils       2.0.3.2 Jonathan Asher Greenberg  8M Wrappers for th...
#>   8 100 tscount         1.4.3   Tobias Liboschik          1M Analysis of Cou...
#>   9 100 smcfcs          1.4.1   Jonathan Bartlett         6M Multiple Imputa...
#>  10 100 rmcorr          0.4.1   Laura R. Marusich         1M Repeated Measur...

# Fuzzy search
advanced_search(Author = "Johnathan~1")

#> - "advanced search" --------------------------- 203 packages in 0.002 seconds -
#>   #     package    version by                           @ title                
#>   1 100 lda        1.4.2   Jonathan Chang              5y Collapsed Gibbs Sa...
#>   2 100 UScancer   0.1.2   Jonathan Lee                6y Create US cancer d...
#>   3 100 mlmi       1.0.0   Jonathan Bartlett           1y Maximum Likelihood...
#>   4  97 chords     0.95.4  Jonathan Rosenblatt         4y Estimation in Resp...
#>   5  97 bootImpute 1.1.0   Jonathan Bartlett           3M Bootstrap Inferenc...
#>   6  96 handyplots 1.1.3   Jonathan Schwartz           2y Handy Plots          
#>   7  96 UnitCircle 0.1.3   Jonathan Berrisch           2y Check if Roots of ...
#>   8  96 rSHAPE     0.3.2   Jonathan Dench              1y Simulated Haploid ...
#>   9  95 Rquake     2.4.0   Jonathan M. Lees            4y Seismic Hypocenter...
#>  10  95 RSeed      0.1.60  Claus Jonathan Fritzemeier  4y Borenstein Analysis