Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...

Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json

A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")
#> - "advanced search" ----------------------------- 4 packages in 0.022 seconds -
#>   #     package   version   by         @ title                                 
#>  1  100 fastclime 1.4.1.1   ORPHANED  2y A Fast Solver for Parameterized LP ...
#>  2  100 udunits2  0.13.2.1  ORPHANED  3y Udunits-2 Bindings for R              
#>  3  100 taxize    0.9.100.1 ORPHANED  3M Taxonomic Information from Around t...
#>  4  100 shipunov  1.17.1    ORPHANED  2y Miscellaneous Functions from Alexey...

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")
#> - "advanced search" ---------------------------- 23 packages in 0.049 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 mockery     0.4.4   Hadley Wickham     1y Mocking Library for R        
#>   2  95 pkgbuild    1.4.5   Gábor Csárdi      21d Find Tools Needed to Build...
#>   3  85 memoise     2.0.1   Winston Chang      3y 'Memoisation' of Functions   
#>   4  79 devtools    2.4.5   Jennifer Bryan     2y Tools to Make Developing R...
#>   5  79 sessioninfo 1.2.2   Gábor Csárdi       3y R Session Information        
#>   6  76 xml2        1.3.6   Hadley Wickham     1y Parse XML                    
#>   7  76 odbc        1.5.0   Hadley Wickham     6M Connect to ODBC Compatible...
#>   8  73 meltr       1.0.2   Duncan Garmonsway 10M Read Non-Rectangular Text ...
#>   9  70 remotes     2.5.0   Gábor Csárdi       8M R Package Installation fro...
#>  10  68 fs          1.6.5   Gábor Csárdi      19d Cross-Platform File System...
advanced_search("Author: Hester AND Author: Wickham")
#> - "advanced search" ---------------------------- 23 packages in 0.003 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 mockery     0.4.4   Hadley Wickham     1y Mocking Library for R        
#>   2  95 pkgbuild    1.4.5   Gábor Csárdi      21d Find Tools Needed to Build...
#>   3  85 memoise     2.0.1   Winston Chang      3y 'Memoisation' of Functions   
#>   4  79 devtools    2.4.5   Jennifer Bryan     2y Tools to Make Developing R...
#>   5  79 sessioninfo 1.2.2   Gábor Csárdi       3y R Session Information        
#>   6  76 xml2        1.3.6   Hadley Wickham     1y Parse XML                    
#>   7  76 odbc        1.5.0   Hadley Wickham     6M Connect to ODBC Compatible...
#>   8  73 meltr       1.0.2   Duncan Garmonsway 10M Read Non-Rectangular Text ...
#>   9  70 remotes     2.5.0   Gábor Csárdi       8M R Package Installation fro...
#>  10  68 fs          1.6.5   Gábor Csárdi      19d Cross-Platform File System...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")
#> - "advanced search" ---------------------------- 26 packages in 0.119 seconds -
#>   #     package        version by                        @ title               
#>   1 100 knitrBootstrap 1.0.3   Jim Hester              10M 'knitr' Bootstrap...
#>   2  92 types          1.0.0   Jim Hester               8y Type Annotations    
#>   3  83 primerTree     1.0.6   Matt Cannon              3y Visually Assessin...
#>   4  75 rex            1.2.1   Kevin Ushey              3y Friendly Regular ...
#>   5  70 gmailr         2.0.0   Jennifer Bryan           1y Access the 'Gmail...
#>   6  68 spelling       2.3.1   Jeroen Ooms              1M Tools for Spell C...
#>   7  66 roger          1.5.1   Vincent Goulet           1y Automated Grading...
#>   8  63 available      1.1.0   Gábor Csárdi             2y Check if the Titl...
#>   9  61 tidycode       0.1.1   Lucy D'Agostino McGowan  5y Analyze Lines of ...
#>  10  61 bench          1.1.3   Davis Vaughan            2y High Precision Ti...

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")
#> - "advanced search" ---------------------------- 23 packages in 0.014 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 mockery     0.4.4   Hadley Wickham     1y Mocking Library for R        
#>   2  88 xml2        1.3.6   Hadley Wickham     1y Parse XML                    
#>   3  88 odbc        1.5.0   Hadley Wickham     6M Connect to ODBC Compatible...
#>   4  88 pkgbuild    1.4.5   Gábor Csárdi      21d Find Tools Needed to Build...
#>   5  77 memoise     2.0.1   Winston Chang      3y 'Memoisation' of Functions   
#>   6  75 devtools    2.4.5   Jennifer Bryan     2y Tools to Make Developing R...
#>   7  75 S7          0.2.0   Hadley Wickham    11d An Object Oriented System ...
#>   8  72 sessioninfo 1.2.2   Gábor Csárdi       3y R Session Information        
#>   9  67 remotes     2.5.0   Gábor Csárdi       8M R Package Installation fro...
#>  10  66 meltr       1.0.2   Duncan Garmonsway 10M Read Non-Rectangular Text ...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")
#> - "advanced search" ---------------------------- 26 packages in 0.022 seconds -
#>   #     package        version by                        @ title               
#>   1 100 knitrBootstrap 1.0.3   Jim Hester              10M 'knitr' Bootstrap...
#>   2  92 types          1.0.0   Jim Hester               8y Type Annotations    
#>   3  83 primerTree     1.0.6   Matt Cannon              3y Visually Assessin...
#>   4  75 rex            1.2.1   Kevin Ushey              3y Friendly Regular ...
#>   5  70 gmailr         2.0.0   Jennifer Bryan           1y Access the 'Gmail...
#>   6  68 spelling       2.3.1   Jeroen Ooms              1M Tools for Spell C...
#>   7  66 roger          1.5.1   Vincent Goulet           1y Automated Grading...
#>   8  63 available      1.1.0   Gábor Csárdi             2y Check if the Titl...
#>   9  61 tidycode       0.1.1   Lucy D'Agostino McGowan  5y Analyze Lines of ...
#>  10  61 bench          1.1.3   Davis Vaughan            2y High Precision Ti...

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")
#> - "advanced search" --------------------------- 551 packages in 0.129 seconds -
#>   #     package       version by                         @ title               
#>   1 100 phd           0.2     Jesse Hemerik             3y Permutation Testi...
#>   2  94 confSAM       0.2     Jesse Hemerik             7y Estimates and Bou...
#>   3  87 permutes      2.8     Cesko C. Voeten           1y Permutation Tests...
#>   4  78 BurStMisc     1.1     Pat Burns                 8y Burns Statistics ...
#>   5  76 IMEC          0.2.0   Maximilian Maier          4y Ising Model of Ex...
#>   6  75 resample      0.6     Tim Hesterberg            2y Resampling Functions
#>   7  74 USP           0.1.2   Thomas B. Berrett         4y U-Statistic Permu...
#>   8  72 packager      1.15.2  Andreas Dominik Cullmann  1y Create, Build and...
#>   9  69 fdrci         2.4     Joshua Millstein          2y Permutation-Based...
#>  10  68 MultSurvTests 0.2     Lukas Arnroth             3y Permutation Tests...

# Packages that have a certain field
advanced_search("_exists_" = "URL")
#> - "advanced search" ------------------------- 10000 packages in 0.047 seconds -
#>   #     package        version by                    @ title                   
#>   1 100 pmparser       1.0.20  Jake Hughey         10M Create and Maintain a...
#>   2 100 prettifyAddins 2.6.1   Stéphane Laurent     1y 'RStudio' Addins to P...
#>   3 100 future.tests   0.7.0   Henrik Bengtsson     1y Test Suite for 'Futur...
#>   4 100 UNF            2.0.8   Thomas J. Leeper     3y Tools for Creating Un...
#>   5 100 rapbase        1.24.2  Are Edvardsen        1y Base Functions and Re...
#>   6 100 dccvalidator   0.3.0   Nicole Kauer         4y Metadata Validation f...
#>   7 100 RGISTools      1.0.2   U Perez - Goya       4y Handling Multiplatfor...
#>   8 100 cort           0.3.2   Oskar Laverny        4y Some Empiric and Nonp...
#>   9 100 Rpvt           0.1.1   Farshad Tabasinejad  5y Estimate the PVT Prop...
#>  10 100 sknifedatar    0.1.2   Rafael Zambrano      3y Swiss Knife of Data     

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")
#> - "advanced search" -------------------------- 10000 packages in 0.05 seconds -
#>   #     package         version by                           @ title           
#>   1 NaN forams          2.0.6   Rodrigo Aluizio             1y Foraminifera ...
#>   2 NaN mldr.resampling 0.2.3   Miguel Ángel Dávila         1y Resampling Al...
#>   3 NaN MonteCarloSEM   0.0.8   Fatih Orcan                 8M Monte Carlo D...
#>   4 NaN hkevp           1.1.5   Leo Belzile                 2y Spatial Extre...
#>   5 NaN perfectphyloR   0.2.1   Charith Karunarathna        4y Reconstruct P...
#>   6 NaN xpose.nlmixr    0.3.0   Justin Wilkins              2y Graphical Dia...
#>   7 NaN WLinfer         1.2.0   Sungbum Kim                 2y Statistical I...
#>   8 NaN intePareto      0.1.2   Yingying Cao                3y Integrative A...
#>   9 NaN ROCS            1.3     Tianwei Yu<tyu8@emory.edu>  8y Receiver Oper...
#>  10 NaN MobileTrigger   0.0.31  Kenith Grey                 6y Run Reports, ...

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)
#> - "advanced search" ------------------------- 10000 packages in 0.041 seconds -
#>   #     package         version by                           @ title           
#>   1 NaN forams          2.0.6   Rodrigo Aluizio             1y Foraminifera ...
#>   2 NaN mldr.resampling 0.2.3   Miguel Ángel Dávila         1y Resampling Al...
#>   3 NaN MonteCarloSEM   0.0.8   Fatih Orcan                 8M Monte Carlo D...
#>   4 NaN hkevp           1.1.5   Leo Belzile                 2y Spatial Extre...
#>   5 NaN perfectphyloR   0.2.1   Charith Karunarathna        4y Reconstruct P...
#>   6 NaN xpose.nlmixr    0.3.0   Justin Wilkins              2y Graphical Dia...
#>   7 NaN WLinfer         1.2.0   Sungbum Kim                 2y Statistical I...
#>   8 NaN intePareto      0.1.2   Yingying Cao                3y Integrative A...
#>   9 NaN ROCS            1.3     Tianwei Yu<tyu8@emory.edu>  8y Receiver Oper...
#>  10 NaN MobileTrigger   0.0.31  Kenith Grey                 6y Run Reports, ...

# Regular expressions
advanced_search(Author = "/Joh?nathan/")
#> - "advanced search" ----------------------------- 0 packages in 0.162 seconds -

# Fuzzy search
advanced_search(Author = "Johnathan~1")
#> - "advanced search" --------------------------- 274 packages in 0.187 seconds -
#>   #     package    version by                                        @ title   
#>   1 100 bootImpute 1.2.1   Jonathan Bartlett                        1y Boots...
#>   2 100 rSHAPE     0.3.2   Jonathan Dench                           5y Simul...
#>   3 100 handyplots 1.1.3   Jonathan Schwartz                        6y Handy...
#>   4 100 UnitCircle 0.1.3   Jonathan Berrisch                        6y Check...
#>   5 100 lda        1.5.2   Santiago Olivella                        7M Colla...
#>   6 100 mlmi       1.1.2   Jonathan Bartlett                        1y Maxim...
#>   7 100 chords     0.95.4  Jonathan Rosenblatt                      8y Estim...
#>   8  96 RSeed      0.1.60  Claus Jonathan Fritzemeier               8y Boren...
#>   9  96 ClamR      2.1.3   Jonathan M. Lees<jonathan.lees@unc.edu>  1y Time ...
#>  10  92 comradesOO 0.1.1   Jonathan Price                           1y Analy...