Advanced CRAN package search

See the Elastic documentation for the syntax and features: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

advanced_search(
  ...,
  json = NULL,
  format = c("short", "long"),
  from = 1,
  size = 10
)

Arguments

...

Search terms. For named terms, the name specifies the field to search for. For unnamed ones, the term is taken as is. The individual terms are combined with the AND operator.

json

A character string that contains the query to send to Elastic. If this is not NULL, then you cannot specify any search terms in ....

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

Value

Search hits.

Examples

# All orphaned packages
advanced_search(Maintainer = "ORPHANED")
#> - "advanced search" ----------------------------- 4 packages in 0.002 seconds -
#>   #     package   version  by         @ title                                  
#>  1  100 udunits2  0.13.2.1 ORPHANED  2y Udunits-2 Bindings for R               
#>  2   98 shipunov  1.17.1   ORPHANED 10M Miscellaneous Functions from Alexey ...
#>  3   97 fastclime 1.4.1.1  ORPHANED  7M A Fast Solver for Parameterized LP P...
#>  4   93 overlap   0.3.4.1  ORPHANED  5d Estimates of Coefficient of Overlapp...

# Packages with both Hester and Wickham as authors
advanced_search(Author = "Hester", Author = "Wickham")
#> - "advanced search" ---------------------------- 21 packages in 0.002 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 mockery     0.4.4   Hadley Wickham     2M Mocking Library for R        
#>   2  98 pkgbuild    1.4.2   Gábor Csárdi       5M Find Tools Needed to Build...
#>   3  89 memoise     2.0.1   Winston Chang      2y 'Memoisation' of Functions   
#>   4  83 xml2        1.3.5   Hadley Wickham     5M Parse XML                    
#>   5  82 devtools    2.4.5   Jennifer Bryan     1y Tools to Make Developing R...
#>   6  82 odbc        1.3.5   Hadley Wickham     5M Connect to ODBC Compatible...
#>   7  81 sessioninfo 1.2.2   Gábor Csárdi       2y R Session Information        
#>   8  75 remotes     2.4.2.1 Gábor Csárdi       4M R Package Installation fro...
#>   9  74 meltr       1.0.1   Duncan Garmonsway  1y Read Non-Rectangular Text ...
#>  10  70 fs          1.6.3   Gábor Csárdi       4M Cross-Platform File System...
advanced_search("Author: Hester AND Author: Wickham")
#> - "advanced search" ---------------------------- 21 packages in 0.002 seconds -
#>   #     package     version by                  @ title                        
#>   1 100 mockery     0.4.4   Hadley Wickham     2M Mocking Library for R        
#>   2  98 pkgbuild    1.4.2   Gábor Csárdi       5M Find Tools Needed to Build...
#>   3  89 memoise     2.0.1   Winston Chang      2y 'Memoisation' of Functions   
#>   4  83 xml2        1.3.5   Hadley Wickham     5M Parse XML                    
#>   5  82 devtools    2.4.5   Jennifer Bryan     1y Tools to Make Developing R...
#>   6  82 odbc        1.3.5   Hadley Wickham     5M Connect to ODBC Compatible...
#>   7  81 sessioninfo 1.2.2   Gábor Csárdi       2y R Session Information        
#>   8  75 remotes     2.4.2.1 Gábor Csárdi       4M R Package Installation fro...
#>   9  74 meltr       1.0.1   Duncan Garmonsway  1y Read Non-Rectangular Text ...
#>  10  70 fs          1.6.3   Gábor Csárdi       4M Cross-Platform File System...

# Packages with Hester but not Wickham as author
advanced_search(Author = "Hester AND NOT Wickham")
#> - "advanced search" ---------------------------- 26 packages in 0.005 seconds -
#>   #     package        version by                        @ title               
#>   1 100 types          1.0.0   Jim Hester               7y Type Annotations    
#>   2  99 knitrBootstrap 1.0.2   Jim Hester               5y 'knitr' Bootstrap...
#>   3  88 primerTree     1.0.6   Matt Cannon              2y Visually Assessin...
#>   4  77 gmailr         2.0.0   Jennifer Bryan           5M Access the 'Gmail...
#>   5  73 rex            1.2.1   Kevin Ushey              2y Friendly Regular ...
#>   6  70 roger          1.5.1   Vincent Goulet           1M Automated Grading...
#>   7  66 spelling       2.2.1   Jeroen Ooms              8M Tools for Spell C...
#>   8  66 available      1.1.0   Gábor Csárdi             1y Check if the Titl...
#>   9  65 tidycode       0.1.1   Lucy D'Agostino McGowan  4y Analyze Lines of ...
#>  10  61 brio           1.1.3   Gábor Csárdi             2y Basic R Input Output

# Packages with Hester as an Author, and Wickham in any field
advanced_search(Author = "Hester", "Wickham")
#> - "advanced search" ---------------------------- 21 packages in 0.011 seconds -
#>   #     package     version by               @ title                           
#>   1 100 mockery     0.4.4   Hadley Wickham  2M Mocking Library for R           
#>   2  96 odbc        1.3.5   Hadley Wickham  5M Connect to ODBC Compatible Da...
#>   3  92 xml2        1.3.5   Hadley Wickham  5M Parse XML                       
#>   4  91 pkgbuild    1.4.2   Gábor Csárdi    5M Find Tools Needed to Build R ...
#>   5  81 memoise     2.0.1   Winston Chang   2y 'Memoisation' of Functions      
#>   6  81 S7          0.1.1   Hadley Wickham  2M An Object Oriented System Mea...
#>   7  80 devtools    2.4.5   Jennifer Bryan  1y Tools to Make Developing R Pa...
#>   8  74 sessioninfo 1.2.2   Gábor Csárdi    2y R Session Information           
#>   9  71 remotes     2.4.2.1 Gábor Csárdi    4M R Package Installation from R...
#>  10  69 pkgload     1.3.3   Lionel Henry    2M Simulate Package Installation...

# Packages with Hester as an Author and Wickham nowhere in the metadata
advanced_search(Author = "Hester", "NOT Wickham")
#> - "advanced search" ---------------------------- 26 packages in 0.005 seconds -
#>   #     package        version by                        @ title               
#>   1 100 types          1.0.0   Jim Hester               7y Type Annotations    
#>   2  99 knitrBootstrap 1.0.2   Jim Hester               5y 'knitr' Bootstrap...
#>   3  88 primerTree     1.0.6   Matt Cannon              2y Visually Assessin...
#>   4  77 gmailr         2.0.0   Jennifer Bryan           5M Access the 'Gmail...
#>   5  73 rex            1.2.1   Kevin Ushey              2y Friendly Regular ...
#>   6  70 roger          1.5.1   Vincent Goulet           1M Automated Grading...
#>   7  66 spelling       2.2.1   Jeroen Ooms              8M Tools for Spell C...
#>   8  66 available      1.1.0   Gábor Csárdi             1y Check if the Titl...
#>   9  65 tidycode       0.1.1   Lucy D'Agostino McGowan  4y Analyze Lines of ...
#>  10  61 brio           1.1.3   Gábor Csárdi             2y Basic R Input Output

# Packages for permutation tests and permissive licenses
advanced_search("permutation test AND NOT License: GPL OR GNU")
#> - "advanced search" --------------------------- 473 packages in 0.035 seconds -
#>   #     package       version by                         @ title               
#>   1 100 phd           0.2     Jesse Hemerik             2y Permutation Testi...
#>   2  92 confSAM       0.2     Jesse Hemerik             6y Estimates and Bou...
#>   3  85 permutes      2.8     Cesko C. Voeten           2M Permutation Tests...
#>   4  77 packager      1.15.2  Andreas Dominik Cullmann  3M Create, Build and...
#>   5  77 BurStMisc     1.1     Pat Burns                 7y Burns Statistics ...
#>   6  77 USP           0.1.2   Thomas B. Berrett         3y U-Statistic Permu...
#>   7  74 resample      0.6     Tim Hesterberg            1y Resampling Functions
#>   8  70 fdrci         2.4     Joshua Millstein          1y Permutation-Based...
#>   9  69 MultSurvTests 0.2     Lukas Arnroth             2y Permutation Tests...
#>  10  68 IMEC          0.2.0   Maximilian Maier          3y Ising Model of Ex...

# Packages that have a certain field
advanced_search("_exists_" = "URL")
#> - "advanced search" ------------------------- 11890 packages in 0.001 seconds -
#>   #     package        version by                     @ title                  
#>   1 100 opentimsr      1.0.13  Michał Piotr Startek  2y An Open-Source Loade...
#>   2 100 rsofun         4.4     Stocker Benjamin     24d The P-Model and Biom...
#>   3 100 gatoRs         1.0.0   Natalie N. Patten     5M Geographic and Taxon...
#>   4 100 oaii           0.1.8   Cezary Kuran          4M 'OpenAI' API R Inter...
#>   5 100 polars         0.7.0   Tatsuya Shima         4M Lightning-Fast 'Data...
#>   6 100 epi2me2r       0.1.0   Quentin D. Read       1y Process Nanopore EPI...
#>   7 100 chromatographR 0.4.4   Ethan Bass            1y Chromatographic Data...
#>   8 100 ggbraid        0.2.2   Neal Grantham         2y Braid Ribbons in 'gg...
#>   9 100 fm.index       0.1.1   Clemens Hug           2y Fast String Searching  
#>  10 100 jjAnno         0.0.3   Junjun Lao            1y An Annotation Packag...

# Packages that do not have a certain field:
advanced_search("NOT _exists_: URL")
#> - "advanced search" -------------------------- 9520 packages in 0.002 seconds -
#>   #     package              version by               @ title                  
#>   1 100 cuml4r               0.1.0   Yitao Li        2y R Interface for the ...
#>   2 100 flashlighttext       0.1.0   Athos Damiani   5M Flashlight Text R In...
#>   3 100 scTEP                0.1.0   Yifan Zhang     1y Single-Cell Trajecto...
#>   4 100 epitopR              0.1.2   Joan Zhang      1y Predict Peptide-MHC ...
#>   5 100 maple                0.99.5  Carter Allen    2y Bayesian Analysis of...
#>   6 100 BIDistances          0.0.2   Quirin Stier    6M Bioinformatic Distances
#>   7 100 gremes               0.1.1   Stefka Asenova  9M Estimation of Tail D...
#>   8 100 changepointsHD       0.3.3   Leland Bybee    4y Change-Point Estimat...
#>   9 100 blockcpd             1.0.0   Lucas Prates    1y Change Point Detecti...
#>  10 100 ForecastCombinations 1.1     Eran Raviv      8y Forecast Combinations  

# The same but as JSON query
query <- '{
"query": {
  "bool": {
    "must_not": {
      "exists": {
        "field": "URL"
      }
    }
  }
}
}'
advanced_search(json = query)
#> - "advanced search" -------------------------- 9520 packages in 0.001 seconds -
#>   #     package              version by               @ title                  
#>   1 100 cuml4r               0.1.0   Yitao Li        2y R Interface for the ...
#>   2 100 flashlighttext       0.1.0   Athos Damiani   5M Flashlight Text R In...
#>   3 100 scTEP                0.1.0   Yifan Zhang     1y Single-Cell Trajecto...
#>   4 100 epitopR              0.1.2   Joan Zhang      1y Predict Peptide-MHC ...
#>   5 100 maple                0.99.5  Carter Allen    2y Bayesian Analysis of...
#>   6 100 BIDistances          0.0.2   Quirin Stier    6M Bioinformatic Distances
#>   7 100 gremes               0.1.1   Stefka Asenova  9M Estimation of Tail D...
#>   8 100 changepointsHD       0.3.3   Leland Bybee    4y Change-Point Estimat...
#>   9 100 blockcpd             1.0.0   Lucas Prates    1y Change Point Detecti...
#>  10 100 ForecastCombinations 1.1     Eran Raviv      8y Forecast Combinations  

# Regular expressions
advanced_search(Author = "/Joh?nathan/")
#> - "advanced search" --------------------------- 255 packages in 0.002 seconds -
#>   #     package         version by                  @ title                    
#>   1 100 sybil           2.2.0   Mayo Roettger      2y Efficient Constrained ...
#>   2 100 xtable          1.8.4   David Scott        5y Export Tables to LaTeX...
#>   3 100 tableone        0.13.2  Kazuki Yoshida     2y Create 'Table 1' to De...
#>   4 100 bridgesampling  1.1.2   Quentin F. Gronau  3y Bridge Sampling for Ma...
#>   5 100 rmcorr          0.6.0   Laura R. Marusich  4M Repeated Measures Corr...
#>   6 100 fivethirtyeight 0.6.2   Albert Y. Kim      2y Data and Code Behind t...
#>   7 100 plink           1.5.1   Jonathan P. Weeks  7y IRT Separate Calibrati...
#>   8 100 smcfcs          1.7.1   Jonathan Bartlett  1y Multiple Imputation of...
#>   9 100 REDCapR         1.1.0   Will Beasley       1y Interaction Between R ...
#>  10 100 Rwave           2.6.5   Jonathan M. Lees   1y Time-Frequency Analysi...

# Fuzzy search
advanced_search(Author = "Johnathan~1")
#> - "advanced search" --------------------------- 257 packages in 0.005 seconds -
#>   #     package    version by                                        @ title   
#>   1 100 bootImpute 1.2.1   Jonathan Bartlett                        6M Boots...
#>   2 100 chords     0.95.4  Jonathan Rosenblatt                      7y Estim...
#>   3  99 mlmi       1.1.2   Jonathan Bartlett                        6M Maxim...
#>   4  98 lda        1.4.2   Jonathan Chang                           8y Colla...
#>   5  97 UnitCircle 0.1.3   Jonathan Berrisch                        5y Check...
#>   6  97 handyplots 1.1.3   Jonathan Schwartz                        5y Handy...
#>   7  96 rSHAPE     0.3.2   Jonathan Dench                           4y Simul...
#>   8  93 RSeed      0.1.60  Claus Jonathan Fritzemeier               7y Boren...
#>   9  92 ClamR      2.1.3   Jonathan M. Lees<jonathan.lees@unc.edu>  3M Time ...
#>  10  91 Oarray     1.4.9   Robin K. S. Hankin                       6y Array...