New CRAN packages

List the latest new CRAN packages.

cran_new(from = "last-week", to = "now", last = Inf)

Arguments

from

Start of the time interval to query. Possible values:

"last-week"

"last-month"

A Date object to be used as a start date.

A POSIXt object to be used as the start date.

A difftime object to used as the time interval until now.

An integer scalar, the number of days until today.

A character string that is converted to a start date using as.POSIXct().

to

End of the time interval to query. It accepts the same kinds of values as from, and additionally it can also be the string "now", to specify the current date and time.

last

Integer to limit the number of returned packages.

Value

Data frame of package descriptions.

Examples

if (FALSE) {
# Last week
cran_new("last-week")

# Last month
cran_new("last-month")

# Last 5 days
cran_new(from = 5)

# From a given date, but at most 10
cran_new(from = "2021-04-06", last = 10)

# March of 2021
cran_new(from = "2021-03-01", to = "2021-04-01")
}