New CRAN packages

List the latest new CRAN packages.

cran_new(from = "last-week", to = "now", last = Inf)

Arguments

from

Start of the time interval to query. Possible values:

"last-week"

"last-month"

A Date object to be used as a start date.

A POSIXt object to be used as the start date.

A difftime object to used as the time interval until now.

An integer scalar, the number of days until today.

A character string that is converted to a start date using as.POSIXct().

to

End of the time interval to query. It accepts the same kinds of values as from, and additionally it can also be the string "now", to specify the current date and time.

last

Integer to limit the number of returned packages.

Value

Data frame of package descriptions.

Examples

# Last week
cran_new("last-week")
#> # A data frame: 38 × 34
#>    `Date/Publication`      Package Title Version `Authors@R` Description License
#>    <chr>                   <chr>   <chr> <chr>   <chr>       <chr>       <chr>  
#>  1 2024-04-19 10:53:02 UTC RbyExa… "Dat… 0.0.100 "c(\nperso… "Data for … GPL (>…
#>  2 2024-04-19 10:22:51 UTC citmre  "Col… 0.1.0   "c(\nperso… "Downloads… MIT + …
#>  3 2024-04-19 10:22:46 UTC puremoe "Pub… 1.0.0   "c(\nperso… "Access a … MIT + …
#>  4 2024-04-18 19:42:49 UTC CirceR  "Con… 1.3.3   "c(\nperso… "Wraps the… Apache…
#>  5 2024-04-18 19:42:45 UTC WordLi… "Mul… 0.2.0   "c(\nperso… "Applicati… GPL (>…
#>  6 2024-04-18 19:32:51 UTC ARUtoo… "Man… 0.6.2   "c(\nperso… "Parse Aut… MIT + …
#>  7 2024-04-18 19:32:48 UTC login   "'sh… 0.9.3   "person(\"… "Framework… GPL (>…
#>  8 2024-04-18 19:23:01 UTC ADLP    "Acc… 0.1.0   "c(\nperso… "Loss rese… GPL-3  
#>  9 2024-04-18 19:22:58 UTC geonod… "Int… 0.1     "c(person(… "Provides … MIT + …
#> 10 2024-04-18 19:13:03 UTC InterNL "Tim… 0.1.0   "c(person(… "Intervent… GPL-3  
#> # ℹ 28 more rows
#> # ℹ 27 more variables: Maintainer <chr>, Encoding <chr>, RoxygenNote <chr>,
#> #   LazyData <chr>, URL <chr>, NeedsCompilation <chr>, Packaged <chr>,
#> #   Author <chr>, Repository <chr>, crandb_file_date <chr>, MD5sum <chr>,
#> #   date <chr>, dependencies <list>, Type <chr>,
#> #   `Config/testthat/edition` <chr>, VignetteBuilder <chr>, BugReports <chr>,
#> #   Date <chr>, SystemRequirements <chr>, Language <chr>, Collate <chr>, …

# Last month
cran_new("last-month")
#> # A data frame: 159 × 43
#>    `Date/Publication`      Package Title Version `Authors@R` Description License
#>    <chr>                   <chr>   <chr> <chr>   <chr>       <chr>       <chr>  
#>  1 2024-04-19 10:53:02 UTC RbyExa… "Dat… 0.0.100 "c(\nperso… "Data for … GPL (>…
#>  2 2024-04-19 10:22:51 UTC citmre  "Col… 0.1.0   "c(\nperso… "Downloads… MIT + …
#>  3 2024-04-19 10:22:46 UTC puremoe "Pub… 1.0.0   "c(\nperso… "Access a … MIT + …
#>  4 2024-04-18 19:42:49 UTC CirceR  "Con… 1.3.3   "c(\nperso… "Wraps the… Apache…
#>  5 2024-04-18 19:42:45 UTC WordLi… "Mul… 0.2.0   "c(\nperso… "Applicati… GPL (>…
#>  6 2024-04-18 19:32:51 UTC ARUtoo… "Man… 0.6.2   "c(\nperso… "Parse Aut… MIT + …
#>  7 2024-04-18 19:32:48 UTC login   "'sh… 0.9.3   "person(\"… "Framework… GPL (>…
#>  8 2024-04-18 19:23:01 UTC ADLP    "Acc… 0.1.0   "c(\nperso… "Loss rese… GPL-3  
#>  9 2024-04-18 19:22:58 UTC geonod… "Int… 0.1     "c(person(… "Provides … MIT + …
#> 10 2024-04-18 19:13:03 UTC InterNL "Tim… 0.1.0   "c(person(… "Intervent… GPL-3  
#> # ℹ 149 more rows
#> # ℹ 36 more variables: Maintainer <chr>, Encoding <chr>, RoxygenNote <chr>,
#> #   LazyData <chr>, URL <chr>, NeedsCompilation <chr>, Packaged <chr>,
#> #   Author <chr>, Repository <chr>, crandb_file_date <chr>, MD5sum <chr>,
#> #   date <chr>, dependencies <list>, Type <chr>,
#> #   `Config/testthat/edition` <chr>, VignetteBuilder <chr>, BugReports <chr>,
#> #   Date <chr>, SystemRequirements <chr>, Language <chr>, Collate <chr>, …

# Last 5 days
cran_new(from = 5)
#> # A data frame: 34 × 32
#>    `Date/Publication`      Package Title Version `Authors@R` Description License
#>    <chr>                   <chr>   <chr> <chr>   <chr>       <chr>       <chr>  
#>  1 2024-04-19 10:53:02 UTC RbyExa… "Dat… 0.0.100 "c(\nperso… "Data for … GPL (>…
#>  2 2024-04-19 10:22:51 UTC citmre  "Col… 0.1.0   "c(\nperso… "Downloads… MIT + …
#>  3 2024-04-19 10:22:46 UTC puremoe "Pub… 1.0.0   "c(\nperso… "Access a … MIT + …
#>  4 2024-04-18 19:42:49 UTC CirceR  "Con… 1.3.3   "c(\nperso… "Wraps the… Apache…
#>  5 2024-04-18 19:42:45 UTC WordLi… "Mul… 0.2.0   "c(\nperso… "Applicati… GPL (>…
#>  6 2024-04-18 19:32:51 UTC ARUtoo… "Man… 0.6.2   "c(\nperso… "Parse Aut… MIT + …
#>  7 2024-04-18 19:32:48 UTC login   "'sh… 0.9.3   "person(\"… "Framework… GPL (>…
#>  8 2024-04-18 19:23:01 UTC ADLP    "Acc… 0.1.0   "c(\nperso… "Loss rese… GPL-3  
#>  9 2024-04-18 19:22:58 UTC geonod… "Int… 0.1     "c(person(… "Provides … MIT + …
#> 10 2024-04-18 19:13:03 UTC InterNL "Tim… 0.1.0   "c(person(… "Intervent… GPL-3  
#> # ℹ 24 more rows
#> # ℹ 25 more variables: Maintainer <chr>, Encoding <chr>, RoxygenNote <chr>,
#> #   LazyData <chr>, URL <chr>, NeedsCompilation <chr>, Packaged <chr>,
#> #   Author <chr>, Repository <chr>, crandb_file_date <chr>, MD5sum <chr>,
#> #   date <chr>, dependencies <list>, Type <chr>,
#> #   `Config/testthat/edition` <chr>, VignetteBuilder <chr>, BugReports <chr>,
#> #   Date <chr>, SystemRequirements <chr>, Language <chr>, Collate <chr>, …

# From a given date, but at most 10
cran_new(from = "2021-04-06", last = 10)
#> # A data frame: 10 × 29
#>    `Date/Publication`      Package Title Version `Authors@R` Description License
#>    <chr>                   <chr>   <chr> <chr>   <chr>       <chr>       <chr>  
#>  1 2024-04-19 10:53:02 UTC RbyExa… "Dat… 0.0.100 "c(\nperso… "Data for … GPL (>…
#>  2 2024-04-19 10:22:51 UTC citmre  "Col… 0.1.0   "c(\nperso… "Downloads… MIT + …
#>  3 2024-04-19 10:22:46 UTC puremoe "Pub… 1.0.0   "c(\nperso… "Access a … MIT + …
#>  4 2024-04-18 19:42:49 UTC CirceR  "Con… 1.3.3   "c(\nperso… "Wraps the… Apache…
#>  5 2024-04-18 19:42:45 UTC WordLi… "Mul… 0.2.0   "c(\nperso… "Applicati… GPL (>…
#>  6 2024-04-18 19:32:51 UTC ARUtoo… "Man… 0.6.2   "c(\nperso… "Parse Aut… MIT + …
#>  7 2024-04-18 19:32:48 UTC login   "'sh… 0.9.3   "person(\"… "Framework… GPL (>…
#>  8 2024-04-18 19:23:01 UTC ADLP    "Acc… 0.1.0   "c(\nperso… "Loss rese… GPL-3  
#>  9 2024-04-18 19:22:58 UTC geonod… "Int… 0.1     "c(person(… "Provides … MIT + …
#> 10 2024-04-18 19:13:03 UTC InterNL "Tim… 0.1.0   "c(person(… "Intervent… GPL-3  
#> # ℹ 22 more variables: Maintainer <chr>, Encoding <chr>, RoxygenNote <chr>,
#> #   LazyData <chr>, URL <chr>, NeedsCompilation <chr>, Packaged <chr>,
#> #   Author <chr>, Repository <chr>, crandb_file_date <chr>, MD5sum <chr>,
#> #   date <chr>, dependencies <list>, Type <chr>,
#> #   `Config/testthat/edition` <chr>, VignetteBuilder <chr>, BugReports <chr>,
#> #   Date <chr>, SystemRequirements <chr>, Language <chr>, Collate <chr>,
#> #   LazyLoad <chr>

# March of 2021
cran_new(from = "2021-03-01", to = "2021-04-01")
#> # A data frame: 231 × 48
#>    `Date/Publication`      Package    Type  Title Version `Authors@R` Maintainer
#>    <chr>                   <chr>      <chr> <chr> <chr>   <chr>       <chr>     
#>  1 2021-03-31 12:00:05 UTC ClustAsse… Pack… "Too… 0.1.0   "c(person(… "Arash Sh…
#>  2 2021-03-31 11:00:02 UTC complexNet Pack… "Com… 0.1.0   "c(person(… "Marco Sm…
#>  3 2021-03-31 09:40:02 UTC BayesRGMM  Pack… "Bay… 1.0     "\nc(perso… "Kuo-Jung…
#>  4 2021-03-31 09:20:02 UTC ggfx       NA    "Pix… 1.0.0   "\nc(perso… "Thomas L…
#>  5 2021-03-31 09:10:09 UTC codeColle… Pack… "Col… 0.1.2   "\nperson(… "Jari Hau…
#>  6 2021-03-31 09:10:02 UTC kindisper… NA    "Sim… 0.9.2   "\nperson(… "Moshe-El…
#>  7 2021-03-31 09:00:05 UTC NonParRol… Pack… "a N… 0.4.0   "c(person(… "Josue M.…
#>  8 2021-03-31 09:00:02 UTC forplo     Pack… "Fle… 0.1.0   "person(\"… "\"Vincen…
#>  9 2021-03-31 08:50:16 UTC OpenSpecy  Pack… "Ana… 0.8.2   "c(person(… "Win Cowg…
#> 10 2021-03-31 08:50:13 UTC cmR        Pack… "Ana… 1.0.1   "c(person(… "Volker S…
#> # ℹ 221 more rows
#> # ℹ 41 more variables: Description <chr>, License <chr>, Encoding <chr>,
#> #   RoxygenNote <chr>, URL <chr>, VignetteBuilder <chr>,
#> #   NeedsCompilation <chr>, Packaged <chr>, Author <chr>, Repository <chr>,
#> #   crandb_file_date <chr>, MD5sum <chr>, date <chr>, dependencies <list>,
#> #   LazyData <chr>, BugReports <chr>, Date <chr>, RdMacros <chr>,
#> #   Language <chr>, LazyLoad <chr>, `Config/testthat/edition` <chr>, …