Metadata about multiple CRAN packages

Metadata about multiple CRAN packages

cran_packages(names)

Arguments

names

Package names. May also contain versions, separated by a @ character.

Value

A data frame of package metadata, one package per row.

Examples

# Get metadata about one package
cran_packages("rhub")
#> # A data frame: 1 × 22
#>   Package Title           Version `Authors@R` Description License LazyData URL  
#> * <chr>   <chr>           <chr>   <chr>       <chr>       <chr>   <chr>    <chr>
#> 1 rhub    Connect to 'R-… 1.1.1   "c(\nperso… "Run 'R CM… MIT + … true     http…
#> # … with 14 more variables: BugReports <chr>, RoxygenNote <chr>,
#> #   Encoding <chr>, VignetteBuilder <chr>, NeedsCompilation <chr>,
#> #   Packaged <chr>, Author <chr>, Maintainer <chr>, Repository <chr>,
#> #   `Date/Publication` <chr>, crandb_file_date <chr>, MD5sum <chr>, date <chr>,
#> #   dependencies <list>
# Get metadata about two packages
cran_packages(c("rhub", "testthat"))
#> # A data frame: 2 × 25
#>   Package  Title          Version `Authors@R` Description License LazyData URL  
#> * <chr>    <chr>          <chr>   <chr>       <chr>       <chr>   <chr>    <chr>
#> 1 rhub     Connect to 'R… 1.1.1   "c(\nperso… "Run 'R CM… MIT + … true     http…
#> 2 testthat Unit Testing … 3.1.4   "\nc(perso… "Software … MIT + … NA       http…
#> # … with 17 more variables: BugReports <chr>, RoxygenNote <chr>,
#> #   Encoding <chr>, VignetteBuilder <chr>, NeedsCompilation <chr>,
#> #   Packaged <chr>, Author <chr>, Maintainer <chr>, Repository <chr>,
#> #   `Date/Publication` <chr>, crandb_file_date <chr>, MD5sum <chr>, date <chr>,
#> #   dependencies <list>, `Config/testthat/edition` <chr>,
#> #   `Config/testthat/start-first` <chr>, `Config/Needs/website` <chr>
# Get metadata about two packages at given versions
cran_packages(c("rhub@1.1.1", "testthat@2.2.1", "testthat@2.2.0"))
#> # A data frame: 3 × 23
#>   Package  Title          Version `Authors@R` Description License LazyData URL  
#> * <chr>    <chr>          <chr>   <chr>       <chr>       <chr>   <chr>    <chr>
#> 1 rhub     Connect to 'R… 1.1.1   "c(\nperso… "Run 'R CM… MIT + … true     http…
#> 2 testthat Unit Testing … 2.2.1   "c(\nperso… "Software … MIT + … NA       http…
#> 3 testthat Unit Testing … 2.2.0   "c(\nperso… "Software … MIT + … NA       http…
#> # … with 15 more variables: BugReports <chr>, RoxygenNote <chr>,
#> #   Encoding <chr>, VignetteBuilder <chr>, NeedsCompilation <chr>,
#> #   Packaged <chr>, Author <chr>, Maintainer <chr>, Repository <chr>,
#> #   `Date/Publication` <chr>, crandb_file_date <chr>, MD5sum <chr>, date <chr>,
#> #   dependencies <list>, Collate <chr>
# If a version does not exist nothing is returned
cran_packages("rhub@notaversion")
#> NULL