Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1,
  size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

Arguments

query
Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

Value

A tibble with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")

#> - "survival" ---------------------------------- 730 packages in 0.008 seconds - 
#>   #     package         version  by                      @ title               
#>   1 100 survival        2.44.1.1 Terry M Therneau      16d Survival Analysis   
#>   2   8 flexsurv        1.1.1    Christopher Jackson   30d Flexible Parametr...
#>   3   7 randomForestSRC 2.8.0    Udaya B. Kogalur       3M Random Forests fo...
#>   4   6 rpart           4.1.15   Beth Atkinson          5d Recursive Partiti...
#>   5   6 timereg         1.9.3    Thomas Scheike         3M Flexible Regressi...
#>   6   6 relsurv         2.2.3    Maja Pohar Perme       5M Relative Survival   
#>   7   5 muhaz           1.2.6.1  David Winsemius        3M Hazard Function E...
#>   8   5 titanic         0.1.0    Paul Hendricks         4y Titanic Passenger...
#>   9   5 survminer       0.4.3    Alboukadel Kassambara  9M Drawing Survival ...
#>  10   5 rstpm2          1.4.5    Mark Clements          3M Generalized Survi...

# Pagination
ps("networks")

#> - "networks" ---------------------------------- 568 packages in 0.013 seconds - 
#>   #     package    version   by                   @ title                      
#>   1 100 igraph     1.2.4     Gábor Csárdi        2M Network Analysis and Vis...
#>   2  50 network    1.15      Carter T. Butts    15d Classes for Relational Data
#>   3  49 RCurl      1.95.4.12 Duncan Temple Lang  1M General Network (HTTP/FT...
#>   4  41 nnet       7.3.12    Brian Ripley        3y Feed-Forward Neural Netw...
#>   5  29 sna        2.4       Carter T. Butts     3y Tools for Social Network...
#>   6  26 DiagrammeR 1.0.0     Richard Iannone     1y Graph/Network Visualization
#>   7  22 visNetwork 2.0.6     Benoit Thieurmel   22d Network Visualization us...
#>   8  22 snow       0.4.3     Luke Tierney        7M Simple Network of Workst...
#>   9  18 ergm       3.9.4     Pavel N. Krivitsky  8M Fit, Simulate and Diagno...
#>  10  17 neuralnet  1.44.2    Marvin N. Wright    2M Training of Neural Networks

more()

#> - "networks" ---------------------------------- 568 packages in 0.009 seconds - 
#>   #    package        version by                    @ title                    
#>  11 17 ggraph         1.0.2   Thomas Lin Pedersen  9M An Implementation of G...
#>  12 16 networkD3      0.4     Christopher Gandrud  2y D3 JavaScript Network ...
#>  13 16 bnlearn        4.4.1   Marco Scutari        1M Bayesian Network Struc...
#>  14 14 diagram        1.6.4   Karline Soetaert     2y Functions for Visualis...
#>  15 13 networkDynamic 0.10.0  Skye Bender-deMoll  12d Dynamic Extensions for...
#>  16 12 intergraph     2.0.2   Michal Bojanowski    4y Coercion Routines for ...
#>  17 11 RSNNS          0.4.11  Christoph Bergmeir   8M Neural Networks using ...
#>  18 11 WGCNA          1.67    Peter Langfelder     6d Weighted Correlation N...
#>  19 10 gRain          1.3.0   Søren Højsgaard      2y Graphical Independence...
#>  20  9 ggnetwork      0.5.1   Francois Briatte     3y Geometries to Plot Net...

# Details
ps("visualization")

#> - "visualization" ----------------------------- 929 packages in 0.009 seconds - 
#>   #     package    version  by                @ title                          
#>   1 100 ggplot2    3.1.1    Hadley Wickham  10d Create Elegant Data Visualis...
#>   2  70 igraph     1.2.4    Gábor Csárdi     2M Network Analysis and Visuali...
#>   3  57 scales     1.0.0    Hadley Wickham   8M Scale Functions for Visualiz...
#>   4  49 rgl        0.100.19 Duncan Murdoch   1M 3D Visualization Using OpenGL  
#>   5  23 ROCR       1.0.7    Tobias Sing      4y Visualizing the Performance ...
#>   6  21 vcd        1.4.4    David Meyer      1y Visualizing Categorical Data   
#>   7  20 ggmap      3.0.0    ORPHANED         2M Spatial Visualization with g...
#>   8  18 DiagrammeR 1.0.0    Richard Iannone  1y Graph/Network Visualization    
#>   9  18 corrplot   0.84     Taiyun Wei       1y Visualization of a Correlati...
#>  10  17 vdiffr     0.3.0    Lionel Henry     3M Visual Regression Testing an...

ps()

#> - "visualization" ----------------------------- 929 packages in 0.009 seconds - 
#> 
#> 1 ggplot2 @ 3.1.1                                   Hadley Wickham, 10 days ago 
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar of
#>   Graphics". You provide the data, tell 'ggplot2' how to map variables to
#>   aesthetics, what graphical primitives to use, and it takes care of the
#>   details.
#>   http://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 2 igraph @ 1.2.4                                     Gábor Csárdi, 2 months ago 
#> ----------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large graphs
#>   very well and provides functions for generating random and regular graphs,
#>   graph visualization, centrality methods and much more.
#>   http://igraph.org
#> 
#> 3 scales @ 1.0.0                                   Hadley Wickham, 8 months ago 
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 4 rgl @ 0.100.19                              Duncan Murdoch, about a month ago 
#> ----------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as well as
#>   functions for constructing representations of geometric objects (cube3d(),
#>   etc.).  Output may be on screen using OpenGL, or to various standard 3D file
#>   formats including WebGL, PLY, OBJ, STL as well as 2D image formats,
#>   including PNG, Postscript, SVG, PGF.
#>   https://r-forge.r-project.org/projects/rgl/
#> 
#> 5 ROCR @ 1.0.7                                         Tobias Sing, 4 years ago 
#> --------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts,<U+000a>and
#>   precision/recall plots are popular examples of
#>   trade-off<U+000a>visualizations for specific pairs of performance measures.
#>   ROCR is a<U+000a>flexible tool for creating cutoff-parameterized 2D
#>   performance curves<U+000a>by freely combining two from over 25 performance
#>   measures (new<U+000a>performance measures can be added using a standard
#>   interface).<U+000a>Curves from different cross-validation or bootstrapping
#>   runs can be<U+000a>averaged by different methods, and standard deviations,
#>   standard<U+000a>errors or box plots can be used to visualize the variability
#>   across<U+000a>the runs. The parameterization can be visualized by printing
#>   cutoff<U+000a>values at the corresponding curve positions, or by coloring
#>   the<U+000a>curve according to cutoff. All components of a performance plot
#>   can<U+000a>be quickly adjusted using a flexible parameter
#>   dispatching<U+000a>mechanism. Despite its flexibility, ROCR is easy to use,
#>   with only<U+000a>three commands and reasonable default values for all
#>   optional<U+000a>parameters.
#>   http://rocr.bioinf.mpi-sb.mpg.de/
#> 
#> 6 vcd @ 1.4.4                                     David Meyer, about a year ago 
#> -------------
#>   # Visualizing Categorical Data
#>   Visualization techniques, data sets, summary and inference procedures aimed
#>   particularly at categorical data. Special emphasis is given to highly
#>   extensible grid graphics. The package was package was originally inspired by
#>   the book "Visualizing Categorical Data" by Michael Friendly and is now the
#>   main support package for a new book, "Discrete Data Analysis with R" by
#>   Michael Friendly and David Meyer (2015).
#> 
#> 7 ggmap @ 3.0.0                                          ORPHANED, 2 months ago 
#> ---------------
#>   # Spatial Visualization with ggplot2
#>   A collection of functions to visualize spatial data and models on top of
#>   static maps from various online sources (e.g Google Maps and Stamen Maps).
#>   It includes tools common to those tasks, including functions for geolocation
#>   and routing.
#>   https://github.com/dkahle/ggmap
#> 
#> 8 DiagrammeR @ 1.0.0                          Richard Iannone, about a year ago 
#> --------------------
#>   # Graph/Network Visualization
#>   Build graph/network structures using functions for stepwise addition and
#>   deletion of nodes and edges. Work with data available in tables for bulk
#>   addition of nodes, edges, and associated metadata. Use graph selections and
#>   traversals to apply changes to specific nodes or edges. A wide selection of
#>   graph algorithms allow for the analysis of graphs. Visualize the graphs and
#>   take advantage of any aesthetic properties assigned to nodes and edges.
#>   https://github.com/rich-iannone/DiagrammeR
#> 
#> 9 corrplot @ 0.84                                  Taiyun Wei, about a year ago 
#> -----------------
#>   # Visualization of a Correlation Matrix
#>   A graphical display of a correlation matrix or general matrix. It also
#>   contains some algorithms to do matrix reordering. In addition, corrplot is
#>   good at details, including choosing color, text labels, color labels,
#>   layout, etc.
#>   https://github.com/taiyun/corrplot
#> 
#> 10 vdiffr @ 0.3.0                                    Lionel Henry, 3 months ago 
#> -----------------
#>   # Visual Regression Testing and Graphical Diffing
#>   An extension to the 'testthat' package that makes it easy to add graphical
#>   unit tests. It provides a Shiny application to manage the test cases.
#>   https://github.com/r-lib/vdiffr

# See the underlying tibble
ps("ropensci")

#> - "ropensci" ---------------------------------- 198 packages in 0.009 seconds - 
#>   #     package    version by                  @ title                         
#>   1 100 webmockr   0.3.4   Scott Chamberlain  2M Stubbing and Setting Expect...
#>   2  87 chromer    0.1     Matthew Pennell    4y Interface to Chromosome Cou...
#>   3  86 rfisheries 0.2     Karthik Ram        3y 'Programmatic Interface to ...
#>   4  86 wdman      0.2.4   Ju Yeong Kim       9M 'Webdriver'/'Selenium' Bina...
#>   5  86 ropenaq    0.2.7   Maëlle Salmon      3M Accesses Air Quality Data f...
#>   6  85 tabulizer  0.2.2   Tom Paskhalis     10M Bindings for 'Tabula' PDF T...
#>   7  77 jsonld     2.1     Jeroen Ooms        2M JSON for Linking Data         
#>   8  72 RSelenium  1.7.5   Ju Yeong Kim       3M R Bindings for 'Selenium We...
#>   9  70 taxize     0.9.6   Scott Chamberlain  2M Taxonomic Information from ...
#>  10  64 spocc      0.9.0   Scott Chamberlain  5M Interface to Species Occurr...

ps()[]

#> - "ropensci" ---------------------------------- 198 packages in 0.009 seconds - 
#> 
#> 1 webmockr @ 0.3.4                              Scott Chamberlain, 2 months ago 
#> ------------------
#>   # Stubbing and Setting Expectations on 'HTTP' Requests
#>   Stubbing and setting expectations on 'HTTP' requests. Includes tools for
#>   stubbing 'HTTP' requests, including expected request conditions and response
#>   conditions. Match on 'HTTP' method, query parameters, request body, headers
#>   and more. Can be used for unit tests or outside of a testing context.
#>   https://github.com/ropensci/webmockr
#>   (devel)
#>   https://ropensci.github.io/http-testing-book/
#>   (user
#>   manual)
#> 
#> 2 chromer @ 0.1                                    Matthew Pennell, 4 years ago 
#> ---------------
#>   # Interface to Chromosome Counts Database API
#>   A programmatic interface to the Chromosome Counts Database
#>   (http://ccdb.tau.ac.il/). This package is part of the rOpenSci suite
#>   (http://ropensci.org)
#>   http://www.github.com/ropensci/chromer
#> 
#> 3 rfisheries @ 0.2                                     Karthik Ram, 3 years ago 
#> ------------------
#>   # 'Programmatic Interface to the 'openfisheries.org' API'
#>   A programmatic interface to 'openfisheries.org'. This package is part of the
#>   'rOpenSci' suite (http://ropensci.org).
#>   http://www.github.com/ropensci/rfisheries
#> 
#> 4 wdman @ 0.2.4                                      Ju Yeong Kim, 9 months ago 
#> ---------------
#>   # 'Webdriver'/'Selenium' Binary Manager
#>   There are a number of binary files associated with the
#>   'Webdriver'/'Selenium' project (see <http://www.seleniumhq.org/download/>,
#>   <https://sites.google.com/a/chromium.org/chromedriver/>,
#>   <https://github.com/mozilla/geckodriver>,
#>   <http://phantomjs.org/download.html> and
#>   <https://github.com/SeleniumHQ/selenium/wiki/InternetExplorerDriver> for
#>   more information). This package provides functions to download these
#>   binaries and to manage processes involving them.
#>   https://github.com/ropensci/wdman
#> 
#> 5 ropenaq @ 0.2.7                                   Maëlle Salmon, 3 months ago 
#> -----------------
#>   # Accesses Air Quality Data from the Open Data Platform OpenAQ
#>   Allows access to air quality data from the API of the OpenAQ platform
#>   <https://docs.openaq.org/>, with the different services the API offers
#>   (getting measurements for a given query, getting latest measurements,
#>   getting lists of available countries/cities/locations). This package has
#>   been peer-reviewed by rOpenSci (v. 0.1.0).
#>   http://github.com/ropensci/ropenaq
#> 
#> 6 tabulizer @ 0.2.2                                Tom Paskhalis, 10 months ago 
#> -------------------
#>   # Bindings for 'Tabula' PDF Table Extractor Library
#>   Bindings for the 'Tabula' <http://tabula.technology/> 'Java' library, which
#>   can extract tables from PDF documents. The 'tabulizerjars' package
#>   <https://github.com/ropensci/tabulizerjars> provides versioned 'Java' .jar
#>   files, including all dependencies, aligned to releases of 'Tabula'.
#>   https://github.com/ropensci/tabulizer
#> 
#> 7 jsonld @ 2.1                                        Jeroen Ooms, 2 months ago 
#> --------------
#>   # JSON for Linking Data
#>   JSON-LD is a light-weight syntax for expressing linked data. It is primarily
#>   intended for web-based programming environments, interoperable web services
#>   and for storing linked data in JSON-based databases. This package provides
#>   bindings to the JavaScript library for converting, expanding and compacting
#>   JSON-LD documents.
#>   https://github.com/ropensci/jsonld#readme
#>   (devel)
#>   https://www.w3.org/TR/json-ld
#>   (spec)
#> 
#> 8 RSelenium @ 1.7.5                                  Ju Yeong Kim, 3 months ago 
#> -------------------
#>   # R Bindings for 'Selenium WebDriver'
#>   Provides a set of R bindings for the 'Selenium 2.0 WebDriver' (see
#>   <https://seleniumhq.github.io/docs/wd.html> for more information) using the
#>   'JsonWireProtocol' (see
#>   <https://github.com/SeleniumHQ/selenium/wiki/JsonWireProtocol> for more
#>   information). 'Selenium 2.0 WebDriver' allows driving a web browser natively
#>   as a user would either locally or on a remote machine using the Selenium
#>   server it marks a leap forward in terms of web browser automation. Selenium
#>   automates web browsers (commonly referred to as browsers). Using RSelenium
#>   you can automate browsers locally or remotely.
#>   http://ropensci.github.io/RSelenium
#> 
#> 9 taxize @ 0.9.6                                Scott Chamberlain, 2 months ago 
#> ----------------
#>   # Taxonomic Information from Around the Web
#>   Interacts with a suite of web 'APIs' for taxonomic tasks, such as getting
#>   database specific taxonomic identifiers, verifying species names, getting
#>   taxonomic hierarchies, fetching downstream and upstream taxonomic names,
#>   getting taxonomic synonyms, converting scientific to common names and vice
#>   versa, and more.
#>   https://github.com/ropensci/taxize
#>   (devel)
#>   https://ropensci.github.io/taxize-book/
#>   (user
#>   manual)
#> 
#> 10 spocc @ 0.9.0                                Scott Chamberlain, 5 months ago 
#> ----------------
#>   # Interface to Species Occurrence Data Sources
#>   A programmatic interface to many species occurrence data sources, including
#>   Global Biodiversity Information Facility ('GBIF'), 'USGSs' Biodiversity
#>   Information Serving Our Nation ('BISON'), 'iNaturalist', Berkeley
#>   'Ecoinformatics' Engine, 'eBird', Integrated Digitized 'Biocollections'
#>   ('iDigBio'), 'VertNet', Ocean 'Biogeographic' Information System ('OBIS'),
#>   and Atlas of Living Australia ('ALA'). Includes functionality for retrieving
#>   species occurrence data, and combining those data.
#>   https://github.com/ropensci/spocc