Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1, size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

# S3 method for pkg_search_result
summary(object, ...)

# S3 method for pkg_search_result
print(x, ...)

Arguments

query
Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

object
Object to summarize.

...
Additional arguments, ignored currently.

x
Object to print.

Value

A tibble with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")

#> - "survival" ---------------------------------- 818 packages in 0.008 seconds -
#>   #     package         version by                      @ title                
#>   1 100 survival        3.1.8   Terry M Therneau       2M Survival Analysis    
#>   2   8 flexsurv        1.1.1   Christopher Jackson   11M Flexible Parametri...
#>   3   7 relsurv         2.2.3   Maja Pohar Perme       1y Relative Survival    
#>   4   6 survminer       0.4.6   Alboukadel Kassambara  5M Drawing Survival C...
#>   5   6 rpart           4.1.15  Beth Atkinson         10M Recursive Partitio...
#>   6   6 timereg         1.9.4   Thomas Scheike         6M Flexible Regressio...
#>   7   6 randomForestSRC 2.9.3   Udaya B. Kogalur      15d Fast Unified Rando...
#>   8   6 survRM2         1.0.2   Hajime Uno             3y Comparing Restrict...
#>   9   5 rstpm2          1.5.1   Mark Clements          3M Smooth Survival Mo...
#>  10   5 muhaz           1.2.6.1 David Winsemius        1y Hazard Function Es...

# Pagination
ps("networks")

#> - "networks" ---------------------------------- 648 packages in 0.007 seconds -
#>   #     package    version  by                    @ title                      
#>   1 100 igraph     1.2.4.2  Gábor Csárdi         2M Network Analysis and Vis...
#>   2  48 network    1.16.0   Carter T. Butts      2M Classes for Relational Data
#>   3  46 RCurl      1.98.1.1 CRAN Team           17d General Network (HTTP/FT...
#>   4  41 nnet       7.3.12   Brian Ripley         4y Feed-Forward Neural Netw...
#>   5  27 sna        2.5      Carter T. Butts      2M Tools for Social Network...
#>   6  27 DiagrammeR 1.0.5    Richard Iannone     20d Graph/Network Visualization
#>   7  22 visNetwork 2.0.9    Benoit Thieurmel     2M Network Visualization us...
#>   8  21 snow       0.4.3    Luke Tierney         1y Simple Network of Workst...
#>   9  20 neuralnet  1.44.2   Marvin N. Wright     1y Training of Neural Networks
#>  10  19 ggraph     2.0.0    Thomas Lin Pedersen  5M An Implementation of Gra...

more()

#> - "networks" ---------------------------------- 648 packages in 0.008 seconds -
#>   #    package        version by                    @ title                    
#>  11 16 ergm           3.10.4  Pavel N. Krivitsky   8M Fit, Simulate and Diag...
#>  12 16 bnlearn        4.5     Marco Scutari        6M Bayesian Network Struc...
#>  13 16 networkD3      0.4     Christopher Gandrud  3y D3 JavaScript Network ...
#>  14 13 diagram        1.6.4   Karline Soetaert     2y Functions for Visualis...
#>  15 12 networkDynamic 0.10.1  Skye Bender-deMoll  15d Dynamic Extensions for...
#>  16 12 intergraph     2.0.2   Michal Bojanowski    5y Coercion Routines for ...
#>  17 12 RSNNS          0.4.12  Christoph Bergmeir   5M Neural Networks using ...
#>  18 11 WGCNA          1.68    Peter Langfelder     9M Weighted Correlation N...
#>  19  9 gRain          1.3.3   Søren Højsgaard     14d Graphical Independence...
#>  20  9 ggnetwork      0.5.1   Francois Briatte     4y Geometries to Plot Net...

# Details
ps("visualization")

#> - "visualization" ---------------------------- 1087 packages in 0.011 seconds -
#>   #     package    version  by                @ title                          
#>   1 100 ggplot2    3.2.1    Hadley Wickham   6M Create Elegant Data Visualis...
#>   2  67 igraph     1.2.4.2  Gábor Csárdi     2M Network Analysis and Visuali...
#>   3  57 scales     1.1.0    Hadley Wickham   3M Scale Functions for Visualiz...
#>   4  45 rgl        0.100.30 Duncan Murdoch   6M 3D Visualization Using OpenGL  
#>   5  21 ROCR       1.0.7    Tobias Sing      5y Visualizing the Performance ...
#>   6  21 vdiffr     0.3.1    Lionel Henry     8M Visual Regression Testing an...
#>   7  20 corrplot   0.84     Taiyun Wei       2y Visualization of a Correlati...
#>   8  19 DiagrammeR 1.0.5    Richard Iannone 20d Graph/Network Visualization    
#>   9  19 vcd        1.4.5    David Meyer     23d Visualizing Categorical Data   
#>  10  18 ggmap      3.0.0    ORPHANED         1y Spatial Visualization with g...

ps()

#> - "visualization" ---------------------------- 1087 packages in 0.011 seconds -
#> 
#> 1 ggplot2 @ 3.2.1                                  Hadley Wickham, 6 months ago
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar
#>   of Graphics". You provide the data, tell 'ggplot2' how to map
#>   variables to aesthetics, what graphical primitives to use, and it
#>   takes care of the details.
#>   http://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 2 igraph @ 1.2.4.2                                   Gábor Csárdi, 2 months ago
#> ------------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large
#>   graphs very well and provides functions for generating random and
#>   regular graphs, graph visualization, centrality methods and much
#>   more.
#>   http://igraph.org
#> 
#> 3 scales @ 1.1.0                                   Hadley Wickham, 3 months ago
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 4 rgl @ 0.100.30                                   Duncan Murdoch, 6 months ago
#> ----------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as
#>   well as functions for constructing representations of geometric
#>   objects (cube3d(), etc.).  Output may be on screen using OpenGL, or
#>   to various standard 3D file formats including WebGL, PLY, OBJ, STL as
#>   well as 2D image formats, including PNG, Postscript, SVG, PGF.
#>   https://r-forge.r-project.org/projects/rgl/
#> 
#> 5 ROCR @ 1.0.7                                         Tobias Sing, 5 years ago
#> --------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts,<U+000a>and
#>   precision/recall plots are popular examples of
#>   trade-off<U+000a>visualizations for specific pairs of performance
#>   measures. ROCR is a<U+000a>flexible tool for creating
#>   cutoff-parameterized 2D performance curves<U+000a>by freely combining
#>   two from over 25 performance measures (new<U+000a>performance
#>   measures can be added using a standard interface).<U+000a>Curves from
#>   different cross-validation or bootstrapping runs can
#>   be<U+000a>averaged by different methods, and standard deviations,
#>   standard<U+000a>errors or box plots can be used to visualize the
#>   variability across<U+000a>the runs. The parameterization can be
#>   visualized by printing cutoff<U+000a>values at the corresponding
#>   curve positions, or by coloring the<U+000a>curve according to cutoff.
#>   All components of a performance plot can<U+000a>be quickly adjusted
#>   using a flexible parameter dispatching<U+000a>mechanism. Despite its
#>   flexibility, ROCR is easy to use, with only<U+000a>three commands and
#>   reasonable default values for all optional<U+000a>parameters.
#>   http://rocr.bioinf.mpi-sb.mpg.de/
#> 
#> 6 vdiffr @ 0.3.1                                     Lionel Henry, 8 months ago
#> ----------------
#>   # Visual Regression Testing and Graphical Diffing
#>   An extension to the 'testthat' package that makes it easy to add
#>   graphical unit tests. It provides a Shiny application to manage the
#>   test cases.
#>   https://github.com/r-lib/vdiffr
#> 
#> 7 corrplot @ 0.84                                       Taiyun Wei, 2 years ago
#> -----------------
#>   # Visualization of a Correlation Matrix
#>   A graphical display of a correlation matrix or general matrix. It
#>   also contains some algorithms to do matrix reordering. In addition,
#>   corrplot is good at details, including choosing color, text labels,
#>   color labels, layout, etc.
#>   https://github.com/taiyun/corrplot
#> 
#> 8 DiagrammeR @ 1.0.5                               Richard Iannone, 20 days ago
#> --------------------
#>   # Graph/Network Visualization
#>   Build graph/network structures using functions for stepwise addition
#>   and deletion of nodes and edges. Work with data available in tables
#>   for bulk addition of nodes, edges, and associated metadata. Use graph
#>   selections and traversals to apply changes to specific nodes or
#>   edges. A wide selection of graph algorithms allow for the analysis of
#>   graphs. Visualize the graphs and take advantage of any aesthetic
#>   properties assigned to nodes and edges.
#>   https://github.com/rich-iannone/DiagrammeR
#> 
#> 9 vcd @ 1.4.5                                          David Meyer, 23 days ago
#> -------------
#>   # Visualizing Categorical Data
#>   Visualization techniques, data sets, summary and inference procedures
#>   aimed particularly at categorical data. Special emphasis is given to
#>   highly extensible grid graphics. The package was package was
#>   originally inspired by the book "Visualizing Categorical Data" by
#>   Michael Friendly and is now the main support package for a new book,
#>   "Discrete Data Analysis with R" by Michael Friendly and David Meyer
#>   (2015).
#> 
#> 10 ggmap @ 3.0.0                                     ORPHANED, about a year ago
#> ----------------
#>   # Spatial Visualization with ggplot2
#>   A collection of functions to visualize spatial data and models on top
#>   of static maps from various online sources (e.g Google Maps and
#>   Stamen Maps). It includes tools common to those tasks, including
#>   functions for geolocation and routing.
#>   https://github.com/dkahle/ggmap

# See the underlying tibble
ps("ropensci")

#> - "ropensci" ---------------------------------- 230 packages in 0.008 seconds -
#>   #     package    version by                       @ title                    
#>   1 100 rtweet     0.7.0   Michael W. Kearney     28d Collecting Twitter Data  
#>   2  77 tabulizer  0.2.2   Tom Paskhalis           2y Bindings for 'Tabula' ...
#>   3  67 webmockr   0.5.0   Scott Chamberlain       2M Stubbing and Setting E...
#>   4  54 ropenaq    0.2.7   Maëlle Salmon           1y Accesses Air Quality D...
#>   5  53 rfisheries 0.2     Karthik Ram             4y 'Programmatic Interfac...
#>   6  53 chromer    0.1     Matthew Pennell         5y Interface to Chromosom...
#>   7  49 visdat     0.5.3   Nicholas Tierney        1y Preliminary Visualisat...
#>   8  47 taxize     0.9.91  Scott Chamberlain       3M Taxonomic Information ...
#>   9  45 tracerer   2.0.2   Richèl J.C. Bilderbeek  2M Tracer from R            
#>  10  44 spocc      1.0.2   Scott Chamberlain       3M Interface to Species O...

ps()[]

#> # A tibble: 10 x 14
#>    score package version title description date                maintainer_name
#>    <dbl> <chr>   <pckg_> <chr> <chr>       <dttm>              <chr>          
#>  1  594. rtweet  0.7.0   Coll… "An implem… 2020-01-08 22:00:10 Michael W. Kea…
#>  2  456. tabuli… 0.2.2   Bind… "Bindings … 2018-06-07 18:16:06 Tom Paskhalis  
#>  3  396. webmoc… 0.5.0   Stub… "Stubbing … 2019-12-04 21:20:03 Scott Chamberl…
#>  4  322. ropenaq 0.2.7   Acce… "Allows ac… 2019-01-31 14:20:03 Maëlle Salmon  
#>  5  316. rfishe… 0.2     'Pro… "A program… 2016-02-19 08:50:03 Karthik Ram    
#>  6  312. chromer 0.1     Inte… "A program… 2015-01-13 10:27:09 Matthew Pennell
#>  7  292. visdat  0.5.3   Prel… "Create pr… 2019-02-15 14:30:03 Nicholas Tiern…
#>  8  277. taxize  0.9.91  Taxo… "Interacts… 2019-11-15 10:40:02 Scott Chamberl…
#>  9  264. tracer… 2.0.2   Trac… "'BEAST2' … 2019-12-02 11:10:02 Richèl J.C. Bi…
#> 10  264. spocc   1.0.2   Inte… "A program… 2019-11-02 08:30:02 Scott Chamberl…
#> # … with 7 more variables: maintainer_email <chr>, revdeps <int>,
#> #   downloads_last_month <int>, license <chr>, url <chr>, bugreports <chr>,
#> #   package_data <I<list>>