Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1,
  size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

Arguments

query
Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

Value

A tibble with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")

#> - "survival" ---------------------------------- 733 packages in 0.009 seconds - 
#>   #     package         version  by                      @ title               
#>   1 100 survival        2.44.1.1 Terry M Therneau       1M Survival Analysis   
#>   2   8 flexsurv        1.1.1    Christopher Jackson    2M Flexible Parametr...
#>   3   6 rpart           4.1.15   Beth Atkinson          1M Recursive Partiti...
#>   4   6 timereg         1.9.3    Thomas Scheike         4M Flexible Regressi...
#>   5   6 relsurv         2.2.3    Maja Pohar Perme       5M Relative Survival   
#>   6   6 randomForestSRC 2.9.0    Udaya B. Kogalur      21d Fast Unified Rand...
#>   7   5 muhaz           1.2.6.1  David Winsemius        4M Hazard Function E...
#>   8   5 survminer       0.4.3    Alboukadel Kassambara  9M Drawing Survival ...
#>   9   5 survRM2         1.0.2    Hajime Uno             2y Comparing Restric...
#>  10   5 titanic         0.1.0    Paul Hendricks         4y Titanic Passenger...

# Pagination
ps("networks")

#> - "networks" ---------------------------------- 573 packages in 0.009 seconds - 
#>   #     package    version   by                   @ title                      
#>   1 100 igraph     1.2.4.1   Gábor Csárdi       21d Network Analysis and Vis...
#>   2  49 network    1.15      Carter T. Butts     1M Classes for Relational Data
#>   3  49 RCurl      1.95.4.12 Duncan Temple Lang  2M General Network (HTTP/FT...
#>   4  41 nnet       7.3.12    Brian Ripley        3y Feed-Forward Neural Netw...
#>   5  29 sna        2.4       Carter T. Butts     3y Tools for Social Network...
#>   6  27 DiagrammeR 1.0.1     Richard Iannone    21d Graph/Network Visualization
#>   7  22 visNetwork 2.0.6     Benoit Thieurmel    2M Network Visualization us...
#>   8  22 snow       0.4.3     Luke Tierney        8M Simple Network of Workst...
#>   9  17 ergm       3.9.4     Pavel N. Krivitsky  9M Fit, Simulate and Diagno...
#>  10  17 neuralnet  1.44.2    Marvin N. Wright    3M Training of Neural Networks

more()

#> - "networks" ----------------------------------- 573 packages in 0.01 seconds - 
#>   #    package        version by                    @ title                    
#>  11 16 ggraph         1.0.2   Thomas Lin Pedersen 10M An Implementation of G...
#>  12 16 networkD3      0.4     Christopher Gandrud  2y D3 JavaScript Network ...
#>  13 16 bnlearn        4.4.1   Marco Scutari        2M Bayesian Network Struc...
#>  14 13 diagram        1.6.4   Karline Soetaert     2y Functions for Visualis...
#>  15 13 networkDynamic 0.10.0  Skye Bender-deMoll   1M Dynamic Extensions for...
#>  16 12 intergraph     2.0.2   Michal Bojanowski    4y Coercion Routines for ...
#>  17 11 RSNNS          0.4.11  Christoph Bergmeir   9M Neural Networks using ...
#>  18 11 WGCNA          1.67    Peter Langfelder     1M Weighted Correlation N...
#>  19 10 gRain          1.3.0   Søren Højsgaard      3y Graphical Independence...
#>  20  9 ggnetwork      0.5.1   Francois Briatte     3y Geometries to Plot Net...

# Details
ps("visualization")

#> - "visualization" ----------------------------- 949 packages in 0.012 seconds - 
#>   #     package    version  by                @ title                          
#>   1 100 ggplot2    3.1.1    Hadley Wickham   1M Create Elegant Data Visualis...
#>   2  70 igraph     1.2.4.1  Gábor Csárdi    21d Network Analysis and Visuali...
#>   3  57 scales     1.0.0    Hadley Wickham   9M Scale Functions for Visualiz...
#>   4  48 rgl        0.100.19 Duncan Murdoch   2M 3D Visualization Using OpenGL  
#>   5  23 ROCR       1.0.7    Tobias Sing      4y Visualizing the Performance ...
#>   6  20 vcd        1.4.4    David Meyer      1y Visualizing Categorical Data   
#>   7  20 ggmap      3.0.0    ORPHANED         3M Spatial Visualization with g...
#>   8  19 DiagrammeR 1.0.1    Richard Iannone 21d Graph/Network Visualization    
#>   9  19 corrplot   0.84     Taiyun Wei       2y Visualization of a Correlati...
#>  10  17 vdiffr     0.3.0    Lionel Henry     4M Visual Regression Testing an...

ps()

#> - "visualization" ----------------------------- 949 packages in 0.012 seconds - 
#> 
#> 1 ggplot2 @ 3.1.1                             Hadley Wickham, about a month ago 
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar of
#>   Graphics". You provide the data, tell 'ggplot2' how to map variables to
#>   aesthetics, what graphical primitives to use, and it takes care of the
#>   details.
#>   http://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 2 igraph @ 1.2.4.1                                    Gábor Csárdi, 21 days ago 
#> ------------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large graphs
#>   very well and provides functions for generating random and regular graphs,
#>   graph visualization, centrality methods and much more.
#>   http://igraph.org
#> 
#> 3 scales @ 1.0.0                                   Hadley Wickham, 9 months ago 
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 4 rgl @ 0.100.19                                   Duncan Murdoch, 2 months ago 
#> ----------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as well as
#>   functions for constructing representations of geometric objects (cube3d(),
#>   etc.).  Output may be on screen using OpenGL, or to various standard 3D file
#>   formats including WebGL, PLY, OBJ, STL as well as 2D image formats,
#>   including PNG, Postscript, SVG, PGF.
#>   https://r-forge.r-project.org/projects/rgl/
#> 
#> 5 ROCR @ 1.0.7                                         Tobias Sing, 4 years ago 
#> --------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts,<U+000a>and
#>   precision/recall plots are popular examples of
#>   trade-off<U+000a>visualizations for specific pairs of performance measures.
#>   ROCR is a<U+000a>flexible tool for creating cutoff-parameterized 2D
#>   performance curves<U+000a>by freely combining two from over 25 performance
#>   measures (new<U+000a>performance measures can be added using a standard
#>   interface).<U+000a>Curves from different cross-validation or bootstrapping
#>   runs can be<U+000a>averaged by different methods, and standard deviations,
#>   standard<U+000a>errors or box plots can be used to visualize the variability
#>   across<U+000a>the runs. The parameterization can be visualized by printing
#>   cutoff<U+000a>values at the corresponding curve positions, or by coloring
#>   the<U+000a>curve according to cutoff. All components of a performance plot
#>   can<U+000a>be quickly adjusted using a flexible parameter
#>   dispatching<U+000a>mechanism. Despite its flexibility, ROCR is easy to use,
#>   with only<U+000a>three commands and reasonable default values for all
#>   optional<U+000a>parameters.
#>   http://rocr.bioinf.mpi-sb.mpg.de/
#> 
#> 6 vcd @ 1.4.4                                     David Meyer, about a year ago 
#> -------------
#>   # Visualizing Categorical Data
#>   Visualization techniques, data sets, summary and inference procedures aimed
#>   particularly at categorical data. Special emphasis is given to highly
#>   extensible grid graphics. The package was package was originally inspired by
#>   the book "Visualizing Categorical Data" by Michael Friendly and is now the
#>   main support package for a new book, "Discrete Data Analysis with R" by
#>   Michael Friendly and David Meyer (2015).
#> 
#> 7 ggmap @ 3.0.0                                          ORPHANED, 3 months ago 
#> ---------------
#>   # Spatial Visualization with ggplot2
#>   A collection of functions to visualize spatial data and models on top of
#>   static maps from various online sources (e.g Google Maps and Stamen Maps).
#>   It includes tools common to those tasks, including functions for geolocation
#>   and routing.
#>   https://github.com/dkahle/ggmap
#> 
#> 8 DiagrammeR @ 1.0.1                               Richard Iannone, 21 days ago 
#> --------------------
#>   # Graph/Network Visualization
#>   Build graph/network structures using functions for stepwise addition and
#>   deletion of nodes and edges. Work with data available in tables for bulk
#>   addition of nodes, edges, and associated metadata. Use graph selections and
#>   traversals to apply changes to specific nodes or edges. A wide selection of
#>   graph algorithms allow for the analysis of graphs. Visualize the graphs and
#>   take advantage of any aesthetic properties assigned to nodes and edges.
#>   https://github.com/rich-iannone/DiagrammeR
#> 
#> 9 corrplot @ 0.84                                       Taiyun Wei, 2 years ago 
#> -----------------
#>   # Visualization of a Correlation Matrix
#>   A graphical display of a correlation matrix or general matrix. It also
#>   contains some algorithms to do matrix reordering. In addition, corrplot is
#>   good at details, including choosing color, text labels, color labels,
#>   layout, etc.
#>   https://github.com/taiyun/corrplot
#> 
#> 10 vdiffr @ 0.3.0                                    Lionel Henry, 4 months ago 
#> -----------------
#>   # Visual Regression Testing and Graphical Diffing
#>   An extension to the 'testthat' package that makes it easy to add graphical
#>   unit tests. It provides a Shiny application to manage the test cases.
#>   https://github.com/r-lib/vdiffr

# See the underlying tibble
ps("ropensci")

#> - "ropensci" ---------------------------------- 199 packages in 0.008 seconds - 
#>   #     package    version by                  @ title                         
#>   1 100 webmockr   0.3.4   Scott Chamberlain  3M Stubbing and Setting Expect...
#>   2  87 chromer    0.1     Matthew Pennell    4y Interface to Chromosome Cou...
#>   3  86 wdman      0.2.4   Ju Yeong Kim      10M 'Webdriver'/'Selenium' Bina...
#>   4  86 rfisheries 0.2     Karthik Ram        3y 'Programmatic Interface to ...
#>   5  86 ropenaq    0.2.7   Maëlle Salmon      3M Accesses Air Quality Data f...
#>   6  85 tabulizer  0.2.2   Tom Paskhalis      1y Bindings for 'Tabula' PDF T...
#>   7  83 RSelenium  1.7.5   Ju Yeong Kim       4M R Bindings for 'Selenium We...
#>   8  76 jsonld     2.1     Jeroen Ooms        3M JSON for Linking Data         
#>   9  70 spocc      0.9.0   Scott Chamberlain  6M Interface to Species Occurr...
#>  10  69 taxize     0.9.7   Scott Chamberlain  4d Taxonomic Information from ...

ps()[]

#> # A tibble: 10 x 14
#>    score package version title description date                maintainer_name
#>    <dbl> <chr>   <pckg_> <chr> <chr>       <dttm>              <chr>          
#>  1  377. webmoc… 0.3.4   Stub… "Stubbing … 2019-02-09 04:40:03 Scott Chamberl…
#>  2  327. chromer 0.1     Inte… "A program… 2015-01-13 10:27:09 Matthew Pennell
#>  3  324. wdman   0.2.4   'Web… "There are… 2018-07-21 20:50:03 Ju Yeong Kim   
#>  4  324. rfishe… 0.2     'Pro… "A program… 2016-02-19 08:50:03 Karthik Ram    
#>  5  323. ropenaq 0.2.7   Acce… "Allows ac… 2019-01-31 14:20:03 Maëlle Salmon  
#>  6  321. tabuli… 0.2.2   Bind… "Bindings … 2018-06-07 18:16:06 Tom Paskhalis  
#>  7  314. RSelen… 1.7.5   R Bi… "Provides … 2019-01-03 08:20:11 Ju Yeong Kim   
#>  8  288. jsonld  2.1     JSON… "JSON-LD i… 2019-02-05 12:43:19 Jeroen Ooms    
#>  9  265. spocc   0.9.0   Inte… "A program… 2018-11-15 05:40:03 Scott Chamberl…
#> 10  259. taxize  0.9.7   Taxo… "Interacts… 2019-05-09 16:30:02 Scott Chamberl…
#> # … with 7 more variables: maintainer_email <chr>, revdeps <int>,
#> #   downloads_last_month <int>, license <chr>, url <chr>, bugreports <chr>,
#> #   package_data <I<list>>