Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1, size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

# S3 method for pkg_search_result
summary(object, ...)

# S3 method for pkg_search_result
print(x, ...)

Arguments

query
Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

object
Object to summarize.

...
Additional arguments, ignored currently.

x
Object to print.

Value

A tibble with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")

#> - "survival" ---------------------------------- 941 packages in 0.009 seconds -
#>   #     package         version    by                      @ title             
#>   1 100 survival        3.2.10     Terry M Therneau      25d Survival Analysis 
#>   2   9 flexsurv        2.0        Christopher Jackson    2M Flexible Parame...
#>   3   7 survminer       0.4.9      Alboukadel Kassambara  1M Drawing Surviva...
#>   4   7 rpart           4.1.15     Beth Atkinson          2y Recursive Parti...
#>   5   6 randomForestSRC 2.11.0     Udaya B. Kogalur      10d Fast Unified Ra...
#>   6   6 rstpm2          1.5.2      Mark Clements          1M Smooth Survival...
#>   7   6 timereg         1.9.8      Thomas Scheike         6M Flexible Regres...
#>   8   5 relsurv         2.2.3      Maja Pohar Perme       2y Relative Survival 
#>   9   5 survRM2         1.0.3      Hajime Uno            10M Comparing Restr...
#>  10   5 pec             2020.11.17 Thomas A. Gerds        5M Prediction Erro...

# Pagination
ps("networks")

#> - "networks" ---------------------------------- 776 packages in 0.008 seconds -
#>   #     package    version  by                    @ title                      
#>   1 100 igraph     1.2.6    Gábor Csárdi         6M Network Analysis and Vis...
#>   2  48 network    1.16.1   Carter T. Butts      6M Classes for Relational Data
#>   3  42 RCurl      1.98.1.3 CRAN Team           25d General Network (HTTP/FT...
#>   4  40 nnet       7.3.15   Brian Ripley         3M Feed-Forward Neural Netw...
#>   5  29 DiagrammeR 1.0.6.1  Richard Iannone      1y Graph/Network Visualization
#>   6  27 sna        2.6      Carter T. Butts      6M Tools for Social Network...
#>   7  26 visNetwork 2.0.9    Benoit Thieurmel     1y Network Visualization us...
#>   8  22 snow       0.4.3    Luke Tierney         3y Simple Network of Workst...
#>   9  22 ggraph     2.0.5    Thomas Lin Pedersen  2M An Implementation of Gra...
#>  10  19 neuralnet  1.44.2   Marvin N. Wright     2y Training of Neural Networks

more()

#> - "networks" ---------------------------------- 776 packages in 0.012 seconds -
#>   #    package         version by                    @ title                   
#>  11 16 spatstat.linnet 2.1.1   Adrian Baddeley     13d Linear Networks Funct...
#>  12 16 bnlearn         4.6.1   Marco Scutari        7M Bayesian Network Stru...
#>  13 16 ergm            3.11.0  Pavel N. Krivitsky   6M Fit, Simulate and Dia...
#>  14 15 networkD3       0.4     Christopher Gandrud  4y D3 JavaScript Network...
#>  15 13 diagram         1.6.5   Karline Soetaert     6M Functions for Visuali...
#>  16 12 WGCNA           1.70.3  Peter Langfelder     1M Weighted Correlation ...
#>  17 12 RSNNS           0.4.12  Christoph Bergmeir   2y Neural Networks using...
#>  18 12 networkDynamic  0.10.1  Skye Bender-deMoll   1y Dynamic Extensions fo...
#>  19 12 intergraph      2.0.2   Michal Bojanowski    6y Coercion Routines for...
#>  20 10 ggnetwork       0.5.8   François Briatte     1y Geometries to Plot Ne...

# Details
ps("visualization")

#> - "visualization" ---------------------------- 1349 packages in 0.009 seconds -
#>   #     package    version  by                    @ title                      
#>   1 100 ggplot2    3.3.3    Thomas Lin Pedersen  3M Create Elegant Data Visu...
#>   2  63 igraph     1.2.6    Gábor Csárdi         6M Network Analysis and Vis...
#>   3  58 scales     1.1.1    Hadley Wickham      11M Scale Functions for Visu...
#>   4  39 rgl        0.105.22 Duncan Murdoch       1M 3D Visualization Using O...
#>   5  22 vdiffr     0.3.3    Lionel Henry         6M Visual Regression Testin...
#>   6  19 DiagrammeR 1.0.6.1  Richard Iannone      1y Graph/Network Visualization
#>   7  19 ROCR       1.0.11   Felix G.M. Ernst     1y Visualizing the Performa...
#>   8  18 corrplot   0.84     Taiyun Wei           3y Visualization of a Corre...
#>   9  17 ggmap      3.0.0    ORPHANED             2y Spatial Visualization wi...
#>  10  16 visNetwork 2.0.9    Benoit Thieurmel     1y Network Visualization us...

ps()

#> - "visualization" ---------------------------- 1349 packages in 0.009 seconds -
#> 
#> 1 ggplot2 @ 3.3.3                             Thomas Lin Pedersen, 3 months ago
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar
#>   of Graphics". You provide the data, tell 'ggplot2' how to map
#>   variables to aesthetics, what graphical primitives to use, and it
#>   takes care of the details.
#>   https://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 2 igraph @ 1.2.6                                     Gábor Csárdi, 6 months ago
#> ----------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large
#>   graphs very well and provides functions for generating random and
#>   regular graphs, graph visualization, centrality methods and much
#>   more.
#>   https://igraph.org
#> 
#> 3 scales @ 1.1.1                                  Hadley Wickham, 11 months ago
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 4 rgl @ 0.105.22                              Duncan Murdoch, about a month ago
#> ----------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as
#>   well as functions for constructing representations of geometric
#>   objects (cube3d(), etc.).  Output may be on screen using OpenGL, or
#>   to various standard 3D file formats including WebGL, PLY, OBJ, STL as
#>   well as 2D image formats, including PNG, Postscript, SVG, PGF.
#>   https://github.com/dmurdoch/rgl
#> 
#> 5 vdiffr @ 0.3.3                                     Lionel Henry, 6 months ago
#> ----------------
#>   # Visual Regression Testing and Graphical Diffing
#>   An extension to the 'testthat' package that makes it easy to add
#>   graphical unit tests. It provides a Shiny application to manage the
#>   test cases.
#>   https://github.com/r-lib/vdiffr
#> 
#> 6 DiagrammeR @ 1.0.6.1                        Richard Iannone, about a year ago
#> ----------------------
#>   # Graph/Network Visualization
#>   Build graph/network structures using functions for stepwise addition
#>   and deletion of nodes and edges. Work with data available in tables
#>   for bulk addition of nodes, edges, and associated metadata. Use graph
#>   selections and traversals to apply changes to specific nodes or
#>   edges. A wide selection of graph algorithms allow for the analysis of
#>   graphs. Visualize the graphs and take advantage of any aesthetic
#>   properties assigned to nodes and edges.
#>   https://github.com/rich-iannone/DiagrammeR
#> 
#> 7 ROCR @ 1.0.11                              Felix G.M. Ernst, about a year ago
#> ---------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts, and
#>   precision/recall plots are popular examples of trade-off
#>   visualizations for specific pairs of performance measures. ROCR is a
#>   flexible tool for creating cutoff-parameterized 2D performance curves
#>   by freely combining two from over 25 performance measures (new
#>   performance measures can be added using a standard interface). Curves
#>   from different cross-validation or bootstrapping runs can be averaged
#>   by different methods, and standard deviations, standard errors or box
#>   plots can be used to visualize the variability across the runs. The
#>   parameterization can be visualized by printing cutoff values at the
#>   corresponding curve positions, or by coloring the curve according to
#>   cutoff. All components of a performance plot can be quickly adjusted
#>   using a flexible parameter dispatching mechanism. Despite its
#>   flexibility, ROCR is easy to use, with only three commands and
#>   reasonable default values for all optional parameters.
#>   http://ipa-tys.github.io/ROCR/
#> 
#> 8 corrplot @ 0.84                                       Taiyun Wei, 3 years ago
#> -----------------
#>   # Visualization of a Correlation Matrix
#>   A graphical display of a correlation matrix or general matrix. It
#>   also contains some algorithms to do matrix reordering. In addition,
#>   corrplot is good at details, including choosing color, text labels,
#>   color labels, layout, etc.
#>   https://github.com/taiyun/corrplot
#> 
#> 9 ggmap @ 3.0.0                                           ORPHANED, 2 years ago
#> ---------------
#>   # Spatial Visualization with ggplot2
#>   A collection of functions to visualize spatial data and models on top
#>   of static maps from various online sources (e.g Google Maps and
#>   Stamen Maps). It includes tools common to those tasks, including
#>   functions for geolocation and routing.
#>   https://github.com/dkahle/ggmap
#> 
#> 10 visNetwork @ 2.0.9                        Benoit Thieurmel, about a year ago
#> ---------------------
#>   # Network Visualization using 'vis.js' Library
#>   Provides an R interface to the 'vis.js' JavaScript charting library.
#>   It allows an interactive visualization of networks.
#>   http://datastorm-open.github.io/visNetwork/

# See the underlying tibble
ps("ropensci")

#> - "ropensci" ---------------------------------- 263 packages in 0.008 seconds -
#>   #     package    version by                       @ title                    
#>   1 100 rtweet     0.7.0   Michael W. Kearney      1y Collecting Twitter Data  
#>   2  77 vcr        0.6.0   Scott Chamberlain       4M Record 'HTTP' Calls to...
#>   3  63 RSelenium  1.7.7   Ju Yeong Kim            1y R Bindings for 'Seleni...
#>   4  61 tabulizer  0.2.2   Tom Paskhalis           3y Bindings for 'Tabula' ...
#>   5  57 webmockr   0.8.0   Scott Chamberlain      27d Stubbing and Setting E...
#>   6  46 visdat     0.5.3   Nicholas Tierney        2y Preliminary Visualisat...
#>   7  45 wdman      0.2.5   Ju Yeong Kim            1y 'Webdriver'/'Selenium'...
#>   8  43 rfisheries 0.2     Karthik Ram             5y 'Programmatic Interfac...
#>   9  41 mcbette    1.13    Richèl J.C. Bilderbeek  4M Model Comparison Using...
#>  10  40 chromer    0.1     Matthew Pennell         6y Interface to Chromosom...

ps()[]

#> # A tibble: 10 x 14
#>    score package  version title description  date                maintainer_name
#>    <dbl> <chr>    <pckg_> <chr> <chr>        <dttm>              <chr>          
#>  1  796. rtweet   0.7.0   Coll… "An impleme… 2020-01-08 22:00:10 Michael W. Kea…
#>  2  610. vcr      0.6.0   Reco… "Record tes… 2020-12-12 16:30:02 Scott Chamberl…
#>  3  498. RSeleni… 1.7.7   R Bi… "Provides a… 2020-02-03 19:00:03 Ju Yeong Kim   
#>  4  484. tabuliz… 0.2.2   Bind… "Bindings f… 2018-06-07 18:16:06 Tom Paskhalis  
#>  5  453. webmockr 0.8.0   Stub… "Stubbing a… 2021-03-14 14:50:02 Scott Chamberl…
#>  6  364. visdat   0.5.3   Prel… "Create pre… 2019-02-15 14:30:03 Nicholas Tiern…
#>  7  361. wdman    0.2.5   'Web… "There are … 2020-01-31 21:30:02 Ju Yeong Kim   
#>  8  341. rfisher… 0.2     'Pro… "A programm… 2016-02-19 08:50:03 Karthik Ram    
#>  9  323. mcbette  1.13    Mode… "'BEAST2' (… 2020-12-05 12:20:02 Richèl J.C. Bi…
#> 10  322. chromer  0.1     Inte… "A programm… 2015-01-13 10:27:09 Matthew Pennell
#> # … with 7 more variables: maintainer_email <chr>, revdeps <int>,
#> #   downloads_last_month <int>, license <chr>, url <chr>, bugreports <chr>,
#> #   package_data <I<list>>