Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1, size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

# S3 method for class 'pkg_search_result'
summary(object, ...)

# S3 method for class 'pkg_search_result'
print(x, ...)

Arguments

query

Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

object

Object to summarize.

...

Additional arguments, ignored currently.

x

Object to print.

Value

A data frame with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")
#> - "survival" --------------------------------- 1248 packages in 0.096 seconds -
#>   #     package         version   by                      @ title              
#>   1 100 survival        3.8.3     Terry M Therneau       4M Survival Analysis  
#>   2  11 survminer       0.5.0     Alboukadel Kassambara  5M Drawing Survival...
#>   3  10 flexsurv        2.3.2     Christopher Jackson    8M Flexible Paramet...
#>   4   7 rstpm2          1.6.6.1   Mark Clements          4M Smooth Survival ...
#>   5   6 rpart           4.1.24    Beth Atkinson          3M Recursive Partit...
#>   6   5 pec             2023.4.12 Thomas A. Gerds        2y Prediction Error...
#>   7   5 randomForestSRC 3.3.3     Udaya B. Kogalur       3M Fast Unified Ran...
#>   8   5 timereg         2.0.6     Thomas Scheike         7M Flexible Regress...
#>   9   5 relsurv         2.3.2     Damjan Manevski        2M Relative Survival  
#>  10   4 muhaz           1.2.6.4   David Winsemius        4y Hazard Function ...

# Pagination
ps("networks")
#> - "networks" --------------------------------- 1035 packages in 0.014 seconds -
#>   #     package    version   by                    @ title                     
#>   1 100 igraph     2.1.4     Kirill Müller        3M Network Analysis and Vi...
#>   2  54 network    1.19.0    Carter T. Butts      4M Classes for Relational ...
#>   3  40 nnet       7.3.20    Brian Ripley         3M Feed-Forward Neural Net...
#>   4  37 DiagrammeR 1.0.11    Richard Iannone      1y Graph/Network Visualiza...
#>   5  33 RCurl      1.98.1.17 CRAN Team           21d General Network (HTTP/F...
#>   6  29 ggraph     2.2.1     Thomas Lin Pedersen  1y An Implementation of Gr...
#>   7  26 visNetwork 2.1.2     Benoit Thieurmel     3y Network Visualization u...
#>   8  25 sna        2.8       Carter T. Butts      7M Tools for Social Networ...
#>   9  21 snow       0.4.4     Luke Tierney         3y Simple Network of Works...
#>  10  19 neuralnet  1.44.2    Marvin N. Wright     6y Training of Neural Netw...
more()
#> - "networks" --------------------------------- 1035 packages in 0.082 seconds -
#>   #    package         version by                    @ title                   
#>  11 17 torch           0.14.2  Daniel Falbel        2M Tensors and Neural Ne...
#>  12 15 networkD3       0.4     Christopher Gandrud  8y D3 JavaScript Network...
#>  13 15 ergm            4.8.1   Pavel N. Krivitsky   3M Fit, Simulate and Dia...
#>  14 14 bnlearn         5.0.2   Marco Scutari        3M Bayesian Network Stru...
#>  15 13 spatstat.linnet 3.2.5   Adrian Baddeley      3M Linear Networks Funct...
#>  16 13 intergraph      2.0.4   Michał Bojanowski    1y Coercion Routines for...
#>  17 12 WGCNA           1.73    Peter Langfelder     7M Weighted Correlation ...
#>  18 12 diagram         1.6.5   Karline Soetaert     5y Functions for Visuali...
#>  19 11 RSNNS           0.4.17  Christoph Bergmeir   1y Neural Networks using...
#>  20 11 networkDynamic  0.11.5  Skye Bender-deMoll   5M Dynamic Extensions fo...

# Details
ps("visualization")
#> - "visualization" ---------------------------- 2005 packages in 0.044 seconds -
#>   #     package    version by                    @ title                       
#>   1 100 ggplot2    3.5.2   Thomas Lin Pedersen  3d Create Elegant Data Visua...
#>   2  60 scales     1.3.0   Thomas Lin Pedersen  1y Scale Functions for Visua...
#>   3  54 igraph     2.1.4   Kirill Müller        3M Network Analysis and Visu...
#>   4  29 rgl        1.3.18  Duncan Murdoch      15d 3D Visualization Using Op...
#>   5  28 vdiffr     1.0.8   Lionel Henry         5M Visual Regression Testing...
#>   6  20 DiagrammeR 1.0.11  Richard Iannone      1y Graph/Network Visualization 
#>   7  17 corrplot   0.95    Taiyun Wei           6M Visualization of a Correl...
#>   8  16 circlize   0.4.16  Zuguang Gu           1y Circular Visualization      
#>   9  14 visNetwork 2.1.2   Benoit Thieurmel     3y Network Visualization usi...
#>  10  14 ROCR       1.0.11  Felix G.M. Ernst     5y Visualizing the Performan...
ps()
#> - "visualization" ---------------------------- 2005 packages in 0.044 seconds -
#> 
#> 1 ggplot2 @ 3.5.2                               Thomas Lin Pedersen, 3 days ago
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar
#>   of Graphics". You provide the data, tell 'ggplot2' how to map
#>   variables to aesthetics, what graphical primitives to use, and it
#>   takes care of the details.
#>   https://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 2 scales @ 1.3.0                          Thomas Lin Pedersen, about a year ago
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 3 igraph @ 2.1.4                                    Kirill Müller, 3 months ago
#> ----------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large
#>   graphs very well and provides functions for generating random and
#>   regular graphs, graph visualization, centrality methods and much
#>   more.
#>   https://r.igraph.org/
#>   https://igraph.org/
#>   https://igraph.discourse.group/
#> 
#> 4 rgl @ 1.3.18                                      Duncan Murdoch, 15 days ago
#> --------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as
#>   well as functions for constructing representations of geometric
#>   objects (cube3d(), etc.).  Output may be on screen using OpenGL, or
#>   to various standard 3D file formats including WebGL, PLY, OBJ, STL as
#>   well as 2D image formats, including PNG, Postscript, SVG, PGF.
#>   https://github.com/dmurdoch/rgl
#>   https://dmurdoch.github.io/rgl/
#> 
#> 5 vdiffr @ 1.0.8                                     Lionel Henry, 5 months ago
#> ----------------
#>   # Visual Regression Testing and Graphical Diffing
#>   An extension to the 'testthat' package that makes it easy to add
#>   graphical unit tests. It provides a Shiny application to manage the
#>   test cases.
#>   https://vdiffr.r-lib.org/
#>   https://github.com/r-lib/vdiffr
#> 
#> 6 DiagrammeR @ 1.0.11                         Richard Iannone, about a year ago
#> ---------------------
#>   # Graph/Network Visualization
#>   Build graph/network structures using functions for stepwise addition
#>   and deletion of nodes and edges. Work with data available in tables
#>   for bulk addition of nodes, edges, and associated metadata. Use graph
#>   selections and traversals to apply changes to specific nodes or
#>   edges. A wide selection of graph algorithms allow for the analysis of
#>   graphs. Visualize the graphs and take advantage of any aesthetic
#>   properties assigned to nodes and edges.
#>   https://rich-iannone.github.io/DiagrammeR/
#>   https://github.com/rich-iannone/DiagrammeR
#> 
#> 7 corrplot @ 0.95                                      Taiyun Wei, 6 months ago
#> -----------------
#>   # Visualization of a Correlation Matrix
#>   Provides a visual exploratory tool on correlation matrix that
#>   supports automatic variable reordering to help detect hidden patterns
#>   among variables.
#>   https://github.com/taiyun/corrplot
#> 
#> 8 circlize @ 0.4.16                                Zuguang Gu, about a year ago
#> -------------------
#>   # Circular Visualization
#>   Circular layout is an efficient way for the visualization of huge
#>   amounts of information. Here this package provides an implementation
#>   of circular layout generation in R as well as an enhancement of
#>   available software. The flexibility of the package is based on the
#>   usage of low-level graphics functions such that self-defined
#>   high-level graphics can be easily implemented by users for specific
#>   purposes. Together with the seamless connection between the powerful
#>   computational and visual environment in R, it gives users more
#>   convenience and freedom to design figures for better understanding
#>   complex patterns behind multiple dimensional data. The package is
#>   described in Gu et al. 2014 <doi:10.1093/bioinformatics/btu393>.
#>   https://github.com/jokergoo/circlize
#>   https://jokergoo.github.io/circlize_book/book/
#> 
#> 9 visNetwork @ 2.1.2                              Benoit Thieurmel, 3 years ago
#> --------------------
#>   # Network Visualization using 'vis.js' Library
#>   Provides an R interface to the 'vis.js' JavaScript charting library.
#>   It allows an interactive visualization of networks.
#>   https://datastorm-open.github.io/visNetwork/
#> 
#> 10 ROCR @ 1.0.11                                  Felix G.M. Ernst, 5 years ago
#> ----------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts, and
#>   precision/recall plots are popular examples of trade-off
#>   visualizations for specific pairs of performance measures. ROCR is a
#>   flexible tool for creating cutoff-parameterized 2D performance curves
#>   by freely combining two from over 25 performance measures (new
#>   performance measures can be added using a standard interface). Curves
#>   from different cross-validation or bootstrapping runs can be averaged
#>   by different methods, and standard deviations, standard errors or box
#>   plots can be used to visualize the variability across the runs. The
#>   parameterization can be visualized by printing cutoff values at the
#>   corresponding curve positions, or by coloring the curve according to
#>   cutoff. All components of a performance plot can be quickly adjusted
#>   using a flexible parameter dispatching mechanism. Despite its
#>   flexibility, ROCR is easy to use, with only three commands and
#>   reasonable default values for all optional parameters.
#>   http://ipa-tys.github.io/ROCR/

# See the underlying data frame
ps("ropensci")
#> - "ropensci" ---------------------------------- 286 packages in 0.071 seconds -
#>   #     package         version by                       @ title               
#>   1 100 vcr             1.7.0   Scott Chamberlain       1M Record 'HTTP' Cal...
#>   2  77 webmockr        2.0.0   Scott Chamberlain       2M Stubbing and Sett...
#>   3  71 RSelenium       1.7.9   Ju Yeong Kim            3y R Bindings for 'S...
#>   4  63 tracerer        2.2.3   Richèl J.C. Bilderbeek  2y Tracer from R       
#>   5  52 fastMatMR       1.2.5   Rohit Goswami           1y High-Performance ...
#>   6  51 rfisheries      0.2     Karthik Ram             9y 'Programmatic Int...
#>   7  51 googleLanguageR 0.3.0   Mark Edmondson          5y Call Google's 'Na...
#>   8  51 mcbette         1.15.3  Richèl J.C. Bilderbeek  8M Model Comparison ...
#>   9  50 beastier        2.5.2   Richèl J.C. Bilderbeek  6M Call 'BEAST2'       
#>  10  46 charlatan       0.6.1   Roel M. Hogervorst      6M Make Fake Data      
ps()[]
#> # A data frame: 10 × 14
#>    score package   version title description date                maintainer_name
#>    <dbl> <chr>     <pckg_> <chr> <chr>       <dttm>              <chr>          
#>  1  743. vcr       1.7.0   "Rec… "Record te… 2025-03-10 09:10:02 Scott Chamberl…
#>  2  572. webmockr  2.0.0   "Stu… "Stubbing … 2025-02-11 19:30:02 Scott Chamberl…
#>  3  528. RSelenium 1.7.9   "R B… "Provides … 2022-09-02 07:10:11 Ju Yeong Kim   
#>  4  468. tracerer  2.2.3   "Tra… "'BEAST2' … 2023-09-27 10:30:02 Richèl J.C. Bi…
#>  5  389. fastMatMR 1.2.5   "Hig… "An interf… 2023-11-03 21:00:06 Rohit Goswami  
#>  6  381. rfisheri… 0.2     "'Pr… "A program… 2016-02-19 08:50:03 Karthik Ram    
#>  7  378. googleLa… 0.3.0   "Cal… "Call 'Goo… 2020-04-19 12:40:02 Mark Edmondson 
#>  8  375. mcbette   1.15.3  "Mod… "'BEAST2' … 2024-08-19 07:40:06 Richèl J.C. Bi…
#>  9  373. beastier  2.5.2   "Cal… "'BEAST2' … 2024-10-06 14:00:02 Richèl J.C. Bi…
#> 10  342. charlatan 0.6.1   "Mak… "Make fake… 2024-10-17 04:20:02 Roel M. Hogerv…
#> # ℹ 7 more variables: maintainer_email <chr>, revdeps <int>,
#> #   downloads_last_month <int>, license <chr>, url <chr>, bugreports <chr>,
#> #   package_data <I<list>>