Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1, size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

# S3 method for class 'pkg_search_result'
summary(object, ...)

# S3 method for class 'pkg_search_result'
print(x, ...)

Arguments

query

Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

object

Object to summarize.

...

Additional arguments, ignored currently.

x

Object to print.

Value

A data frame with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")
#> - "survival" --------------------------------- 1186 packages in 0.072 seconds -
#>   #     package        version    by                       @ title             
#>   1 100 survival       3.7.0      Terry M Therneau        4M Survival Analysis 
#>   2  19 survminer      0.4.9      Alboukadel Kassambara   4y Drawing Surviva...
#>   3  10 rpart          4.1.23     Beth Atkinson          10M Recursive Parti...
#>   4   9 pec            2023.4.12  Thomas A. Gerds         1y Prediction Erro...
#>   5   9 relsurv        2.2.9      Damjan Manevski         2y Relative Survival 
#>   6   8 muhaz          1.2.6.4    David Winsemius         3y Hazard Function...
#>   7   8 simsurv        1.0.0      Sam Brilleman           4y Simulate Surviv...
#>   8   7 riskRegression 2023.12.21 Thomas Alexander Gerds  9M Risk Regression...
#>   9   7 survRM2        1.0.4      Hajime Uno              2y Comparing Restr...
#>  10   7 titanic        0.1.0      Paul Hendricks          9y Titanic Passeng...

# Pagination
ps("networks")
#> - "networks" ----------------------------------- 996 packages in 0.02 seconds -
#>   #     package    version   by                    @ title                     
#>   1 100 network    1.18.2    Carter T. Butts     10M Classes for Relational ...
#>   2  78 igraph     2.0.3     Kirill Müller        7M Network Analysis and Vi...
#>   3  70 nnet       7.3.19    Brian Ripley         1y Feed-Forward Neural Net...
#>   4  63 DiagrammeR 1.0.11    Richard Iannone      8M Graph/Network Visualiza...
#>   5  56 RCurl      1.98.1.16 CRAN Team            2M General Network (HTTP/F...
#>   6  52 ggraph     2.2.1     Thomas Lin Pedersen  7M An Implementation of Gr...
#>   7  46 visNetwork 2.1.2     Benoit Thieurmel     2y Network Visualization u...
#>   8  38 snow       0.4.4     Luke Tierney         3y Simple Network of Works...
#>   9  34 neuralnet  1.44.2    Marvin N. Wright     6y Training of Neural Netw...
#>  10  28 sna        2.8       Carter T. Butts     16d Tools for Social Networ...
more()
#> - "networks" ---------------------------------- 996 packages in 0.014 seconds -
#>   #    package        version by                    @ title                    
#>  11 26 networkD3      0.4     Christopher Gandrud  8y D3 JavaScript Network ...
#>  12 26 ergm           4.6.0   Pavel N. Krivitsky   9M Fit, Simulate and Diag...
#>  13 23 intergraph     2.0.4   Michał Bojanowski    8M Coercion Routines for ...
#>  14 21 diagram        1.6.5   Karline Soetaert     4y Functions for Visualis...
#>  15 20 RSNNS          0.4.17  Christoph Bergmeir  10M Neural Networks using ...
#>  16 19 networkDynamic 0.11.4  Skye Bender-deMoll  10M Dynamic Extensions for...
#>  17 19 ggnetwork      0.5.13  François Briatte     7M Geometries to Plot Net...
#>  18 18 netmeta        2.9.0   Guido Schwarzer      9M Network Meta-Analysis ...
#>  19 17 igraphdata     1.0.1   Gabor Csardi         9y A Collection of Networ...
#>  20 16 nnfor          0.9.9   Nikolaos Kourentzes 10M Time Series Forecastin...

# Details
ps("visualization")
#> - "visualization" ---------------------------- 1830 packages in 0.007 seconds -
#>   #     package    version by                    @ title                       
#>   1 100 scales     1.3.0   Thomas Lin Pedersen 10M Scale Functions for Visua...
#>   2  47 vdiffr     1.0.7   Lionel Henry         1y Visual Regression Testing...
#>   3  47 ggplot2    3.5.1   Thomas Lin Pedersen  5M Create Elegant Data Visua...
#>   4  40 igraph     2.0.3   Kirill Müller        7M Network Analysis and Visu...
#>   5  38 rgl        1.3.1   Duncan Murdoch       7M 3D Visualization Using Op...
#>   6  32 DiagrammeR 1.0.11  Richard Iannone      8M Graph/Network Visualization 
#>   7  26 circlize   0.4.16  Zuguang Gu           7M Circular Visualization      
#>   8  25 ROCR       1.0.11  Felix G.M. Ernst     4y Visualizing the Performan...
#>   9  24 visNetwork 2.1.2   Benoit Thieurmel     2y Network Visualization usi...
#>  10  20 vcd        1.4.13  David Meyer          8d Visualizing Categorical Data
ps()
#> - "visualization" ---------------------------- 1830 packages in 0.007 seconds -
#> 
#> 1 scales @ 1.3.0                             Thomas Lin Pedersen, 10 months ago
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 2 vdiffr @ 1.0.7                                 Lionel Henry, about a year ago
#> ----------------
#>   # Visual Regression Testing and Graphical Diffing
#>   An extension to the 'testthat' package that makes it easy to add
#>   graphical unit tests. It provides a Shiny application to manage the
#>   test cases.
#>   https://vdiffr.r-lib.org/
#>   https://github.com/r-lib/vdiffr
#> 
#> 3 ggplot2 @ 3.5.1                             Thomas Lin Pedersen, 5 months ago
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar
#>   of Graphics". You provide the data, tell 'ggplot2' how to map
#>   variables to aesthetics, what graphical primitives to use, and it
#>   takes care of the details.
#>   https://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 4 igraph @ 2.0.3                                    Kirill Müller, 7 months ago
#> ----------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large
#>   graphs very well and provides functions for generating random and
#>   regular graphs, graph visualization, centrality methods and much
#>   more.
#>   https://r.igraph.org/
#>   https://igraph.org/
#>   https://igraph.discourse.group/
#> 
#> 5 rgl @ 1.3.1                                      Duncan Murdoch, 7 months ago
#> -------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as
#>   well as functions for constructing representations of geometric
#>   objects (cube3d(), etc.).  Output may be on screen using OpenGL, or
#>   to various standard 3D file formats including WebGL, PLY, OBJ, STL as
#>   well as 2D image formats, including PNG, Postscript, SVG, PGF.
#>   https://github.com/dmurdoch/rgl
#>   https://dmurdoch.github.io/rgl/
#> 
#> 6 DiagrammeR @ 1.0.11                             Richard Iannone, 8 months ago
#> ---------------------
#>   # Graph/Network Visualization
#>   Build graph/network structures using functions for stepwise addition
#>   and deletion of nodes and edges. Work with data available in tables
#>   for bulk addition of nodes, edges, and associated metadata. Use graph
#>   selections and traversals to apply changes to specific nodes or
#>   edges. A wide selection of graph algorithms allow for the analysis of
#>   graphs. Visualize the graphs and take advantage of any aesthetic
#>   properties assigned to nodes and edges.
#>   https://rich-iannone.github.io/DiagrammeR/
#>   https://github.com/rich-iannone/DiagrammeR
#> 
#> 7 circlize @ 0.4.16                                    Zuguang Gu, 7 months ago
#> -------------------
#>   # Circular Visualization
#>   Circular layout is an efficient way for the visualization of huge
#>   amounts of information. Here this package provides an implementation
#>   of circular layout generation in R as well as an enhancement of
#>   available software. The flexibility of the package is based on the
#>   usage of low-level graphics functions such that self-defined
#>   high-level graphics can be easily implemented by users for specific
#>   purposes. Together with the seamless connection between the powerful
#>   computational and visual environment in R, it gives users more
#>   convenience and freedom to design figures for better understanding
#>   complex patterns behind multiple dimensional data. The package is
#>   described in Gu et al. 2014 <doi:10.1093/bioinformatics/btu393>.
#>   https://github.com/jokergoo/circlize
#>   https://jokergoo.github.io/circlize_book/book/
#> 
#> 8 ROCR @ 1.0.11                                   Felix G.M. Ernst, 4 years ago
#> ---------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts, and
#>   precision/recall plots are popular examples of trade-off
#>   visualizations for specific pairs of performance measures. ROCR is a
#>   flexible tool for creating cutoff-parameterized 2D performance curves
#>   by freely combining two from over 25 performance measures (new
#>   performance measures can be added using a standard interface). Curves
#>   from different cross-validation or bootstrapping runs can be averaged
#>   by different methods, and standard deviations, standard errors or box
#>   plots can be used to visualize the variability across the runs. The
#>   parameterization can be visualized by printing cutoff values at the
#>   corresponding curve positions, or by coloring the curve according to
#>   cutoff. All components of a performance plot can be quickly adjusted
#>   using a flexible parameter dispatching mechanism. Despite its
#>   flexibility, ROCR is easy to use, with only three commands and
#>   reasonable default values for all optional parameters.
#>   http://ipa-tys.github.io/ROCR/
#> 
#> 9 visNetwork @ 2.1.2                              Benoit Thieurmel, 2 years ago
#> --------------------
#>   # Network Visualization using 'vis.js' Library
#>   Provides an R interface to the 'vis.js' JavaScript charting library.
#>   It allows an interactive visualization of networks.
#>   https://datastorm-open.github.io/visNetwork/
#> 
#> 10 vcd @ 1.4.13                                         David Meyer, 8 days ago
#> ---------------
#>   # Visualizing Categorical Data
#>   Visualization techniques, data sets, summary and inference procedures
#>   aimed particularly at categorical data. Special emphasis is given to
#>   highly extensible grid graphics. The package was package was
#>   originally inspired by the book "Visualizing Categorical Data" by
#>   Michael Friendly and is now the main support package for a new book,
#>   "Discrete Data Analysis with R" by Michael Friendly and David Meyer
#>   (2015).

# See the underlying data frame
ps("ropensci")
#> - "ropensci" ---------------------------------- 274 packages in 0.023 seconds -
#>   #     package         version by                       @ title               
#>   1 100 RSelenium       1.7.9   Ju Yeong Kim            2y R Bindings for 'S...
#>   2  84 tracerer        2.2.3   Richèl J.C. Bilderbeek  1y Tracer from R       
#>   3  72 jsonld          2.2     Jeroen Ooms             4y JSON for Linking ...
#>   4  71 charlatan       0.5.1   Roel M. Hogervorst      1y Make Fake Data      
#>   5  68 fastMatMR       1.2.5   Rohit Goswami          11M High-Performance ...
#>   6  68 googleLanguageR 0.3.0   Mark Edmondson          4y Call Google's 'Na...
#>   7  67 mcbette         1.15.3  Richèl J.C. Bilderbeek  1M Model Comparison ...
#>   8  67 visdat          0.6.0   Nicholas Tierney        2y Preliminary Visua...
#>   9  66 rfisheries      0.2     Karthik Ram             9y 'Programmatic Int...
#>  10  59 ritis           1.0.0   Scott Chamberlain       4y Integrated Taxono...
ps()[]
#> # A data frame: 10 × 14
#>    score package   version title description date                maintainer_name
#>    <dbl> <chr>     <pckg_> <chr> <chr>       <dttm>              <chr>          
#>  1  560. RSelenium 1.7.9   "R B… "Provides … 2022-09-02 07:10:11 Ju Yeong Kim   
#>  2  472. tracerer  2.2.3   "Tra… "'BEAST2' … 2023-09-27 10:30:02 Richèl J.C. Bi…
#>  3  405. jsonld    2.2     "JSO… "JSON-LD i… 2020-05-27 06:20:03 Jeroen Ooms    
#>  4  399. charlatan 0.5.1   "Mak… "Make fake… 2023-09-13 12:40:02 Roel M. Hogerv…
#>  5  380. fastMatMR 1.2.5   "Hig… "An interf… 2023-11-03 21:00:06 Rohit Goswami  
#>  6  380. googleLa… 0.3.0   "Cal… "Call 'Goo… 2020-04-19 12:40:02 Mark Edmondson 
#>  7  373. mcbette   1.15.3  "Mod… "'BEAST2' … 2024-08-19 07:40:06 Richèl J.C. Bi…
#>  8  373. visdat    0.6.0   "Pre… "Create pr… 2023-02-02 02:10:02 Nicholas Tiern…
#>  9  372. rfisheri… 0.2     "'Pr… "A program… 2016-02-19 08:50:03 Karthik Ram    
#> 10  332. ritis     1.0.0   "Int… "An interf… 2021-02-02 06:30:02 Scott Chamberl…
#> # ℹ 7 more variables: maintainer_email <chr>, revdeps <int>,
#> #   downloads_last_month <int>, license <chr>, url <chr>, bugreports <chr>,
#> #   package_data <I<list>>