Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1, size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

# S3 method for pkg_search_result
summary(object, ...)

# S3 method for pkg_search_result
print(x, ...)

Arguments

query
Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format
Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from
Where to start listing the results, for pagination.

size
The number of results to list.

object
Object to summarize.

...
Additional arguments, ignored currently.

x
Object to print.

Value

A tibble with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")

#> - "survival" ----------------------------------- 876 packages in 0.01 seconds -
#>   #     package         version   by                      @ title              
#>   1 100 survival        3.2.7     Terry M Therneau      11d Survival Analysis  
#>   2   8 flexsurv        1.1.1     Christopher Jackson    2y Flexible Paramet...
#>   3   7 survminer       0.4.8     Alboukadel Kassambara  3M Drawing Survival...
#>   4   6 rpart           4.1.15    Beth Atkinson          1y Recursive Partit...
#>   5   6 randomForestSRC 2.9.3     Udaya B. Kogalur       9M Fast Unified Ran...
#>   6   6 rstpm2          1.5.1     Mark Clements          1y Smooth Survival ...
#>   7   6 relsurv         2.2.3     Maja Pohar Perme       2y Relative Survival  
#>   8   5 survRM2         1.0.3     Hajime Uno             4M Comparing Restri...
#>   9   5 muhaz           1.2.6.1   David Winsemius        2y Hazard Function ...
#>  10   5 pec             2019.11.3 Thomas A. Gerds        1y Prediction Error...

# Pagination
ps("networks")

#> - "networks" ---------------------------------- 708 packages in 0.008 seconds -
#>   #     package    version  by                    @ title                      
#>   1 100 igraph     1.2.6    Gábor Csárdi         3d Network Analysis and Vis...
#>   2  46 network    1.16.1   Carter T. Butts      2d Classes for Relational Data
#>   3  44 RCurl      1.98.1.2 CRAN Team            6M General Network (HTTP/FT...
#>   4  41 nnet       7.3.14   Brian Ripley         6M Feed-Forward Neural Netw...
#>   5  30 DiagrammeR 1.0.6.1  Richard Iannone      5M Graph/Network Visualization
#>   6  26 sna        2.6      Carter T. Butts      3d Tools for Social Network...
#>   7  25 visNetwork 2.0.9    Benoit Thieurmel    10M Network Visualization us...
#>   8  22 snow       0.4.3    Luke Tierney         2y Simple Network of Workst...
#>   9  21 ggraph     2.0.3    Thomas Lin Pedersen  5M An Implementation of Gra...
#>  10  19 neuralnet  1.44.2   Marvin N. Wright     2y Training of Neural Networks

more()

#> - "networks" ---------------------------------- 708 packages in 0.007 seconds -
#>   #    package        version by                    @ title                    
#>  11 17 bnlearn        4.6.1   Marco Scutari       18d Bayesian Network Struc...
#>  12 16 ergm           3.10.4  Pavel N. Krivitsky   1y Fit, Simulate and Diag...
#>  13 15 networkD3      0.4     Christopher Gandrud  4y D3 JavaScript Network ...
#>  14 12 networkDynamic 0.10.1  Skye Bender-deMoll   9M Dynamic Extensions for...
#>  15 12 WGCNA          1.69    Peter Langfelder     7M Weighted Correlation N...
#>  16 12 diagram        1.6.5   Karline Soetaert     9d Functions for Visualis...
#>  17 12 RSNNS          0.4.12  Christoph Bergmeir   1y Neural Networks using ...
#>  18 11 intergraph     2.0.2   Michal Bojanowski    5y Coercion Routines for ...
#>  19  9 gRain          1.3.6   Søren Højsgaard      2M Graphical Independence...
#>  20  9 ggnetwork      0.5.8   François Briatte     8M Geometries to Plot Net...

# Details
ps("visualization")

#> - "visualization" ---------------------------- 1232 packages in 0.008 seconds -
#>   #     package    version  by                    @ title                      
#>   1 100 ggplot2    3.3.2    Thomas Lin Pedersen  4M Create Elegant Data Visu...
#>   2  64 igraph     1.2.6    Gábor Csárdi         3d Network Analysis and Vis...
#>   3  57 scales     1.1.1    Hadley Wickham       5M Scale Functions for Visu...
#>   4  42 rgl        0.100.54 Duncan Murdoch       6M 3D Visualization Using O...
#>   5  23 vdiffr     0.3.3    Lionel Henry         2d Visual Regression Testin...
#>   6  20 ROCR       1.0.11   Felix G.M. Ernst     5M Visualizing the Performa...
#>   7  20 DiagrammeR 1.0.6.1  Richard Iannone      5M Graph/Network Visualization
#>   8  19 corrplot   0.84     Taiyun Wei           3y Visualization of a Corre...
#>   9  17 vcd        1.4.8    David Meyer         18d Visualizing Categorical ...
#>  10  17 ggmap      3.0.0    ORPHANED             2y Spatial Visualization wi...

ps()

#> - "visualization" ---------------------------- 1232 packages in 0.008 seconds -
#> 
#> 1 ggplot2 @ 3.3.2                             Thomas Lin Pedersen, 4 months ago
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar
#>   of Graphics". You provide the data, tell 'ggplot2' how to map
#>   variables to aesthetics, what graphical primitives to use, and it
#>   takes care of the details.
#>   http://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 2 igraph @ 1.2.6                                       Gábor Csárdi, 3 days ago
#> ----------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large
#>   graphs very well and provides functions for generating random and
#>   regular graphs, graph visualization, centrality methods and much
#>   more.
#>   https://igraph.org
#> 
#> 3 scales @ 1.1.1                                   Hadley Wickham, 5 months ago
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 4 rgl @ 0.100.54                                   Duncan Murdoch, 6 months ago
#> ----------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as
#>   well as functions for constructing representations of geometric
#>   objects (cube3d(), etc.).  Output may be on screen using OpenGL, or
#>   to various standard 3D file formats including WebGL, PLY, OBJ, STL as
#>   well as 2D image formats, including PNG, Postscript, SVG, PGF.
#>   https://r-forge.r-project.org/projects/rgl/
#> 
#> 5 vdiffr @ 0.3.3                                       Lionel Henry, 2 days ago
#> ----------------
#>   # Visual Regression Testing and Graphical Diffing
#>   An extension to the 'testthat' package that makes it easy to add
#>   graphical unit tests. It provides a Shiny application to manage the
#>   test cases.
#>   https://github.com/r-lib/vdiffr
#> 
#> 6 ROCR @ 1.0.11                                  Felix G.M. Ernst, 5 months ago
#> ---------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts, and
#>   precision/recall plots are popular examples of trade-off
#>   visualizations for specific pairs of performance measures. ROCR is a
#>   flexible tool for creating cutoff-parameterized 2D performance curves
#>   by freely combining two from over 25 performance measures (new
#>   performance measures can be added using a standard interface). Curves
#>   from different cross-validation or bootstrapping runs can be averaged
#>   by different methods, and standard deviations, standard errors or box
#>   plots can be used to visualize the variability across the runs. The
#>   parameterization can be visualized by printing cutoff values at the
#>   corresponding curve positions, or by coloring the curve according to
#>   cutoff. All components of a performance plot can be quickly adjusted
#>   using a flexible parameter dispatching mechanism. Despite its
#>   flexibility, ROCR is easy to use, with only three commands and
#>   reasonable default values for all optional parameters.
#>   http://ipa-tys.github.io/ROCR/
#> 
#> 7 DiagrammeR @ 1.0.6.1                            Richard Iannone, 5 months ago
#> ----------------------
#>   # Graph/Network Visualization
#>   Build graph/network structures using functions for stepwise addition
#>   and deletion of nodes and edges. Work with data available in tables
#>   for bulk addition of nodes, edges, and associated metadata. Use graph
#>   selections and traversals to apply changes to specific nodes or
#>   edges. A wide selection of graph algorithms allow for the analysis of
#>   graphs. Visualize the graphs and take advantage of any aesthetic
#>   properties assigned to nodes and edges.
#>   https://github.com/rich-iannone/DiagrammeR
#> 
#> 8 corrplot @ 0.84                                       Taiyun Wei, 3 years ago
#> -----------------
#>   # Visualization of a Correlation Matrix
#>   A graphical display of a correlation matrix or general matrix. It
#>   also contains some algorithms to do matrix reordering. In addition,
#>   corrplot is good at details, including choosing color, text labels,
#>   color labels, layout, etc.
#>   https://github.com/taiyun/corrplot
#> 
#> 9 vcd @ 1.4.8                                          David Meyer, 18 days ago
#> -------------
#>   # Visualizing Categorical Data
#>   Visualization techniques, data sets, summary and inference procedures
#>   aimed particularly at categorical data. Special emphasis is given to
#>   highly extensible grid graphics. The package was package was
#>   originally inspired by the book "Visualizing Categorical Data" by
#>   Michael Friendly and is now the main support package for a new book,
#>   "Discrete Data Analysis with R" by Michael Friendly and David Meyer
#>   (2015).
#> 
#> 10 ggmap @ 3.0.0                                          ORPHANED, 2 years ago
#> ----------------
#>   # Spatial Visualization with ggplot2
#>   A collection of functions to visualize spatial data and models on top
#>   of static maps from various online sources (e.g Google Maps and
#>   Stamen Maps). It includes tools common to those tasks, including
#>   functions for geolocation and routing.
#>   https://github.com/dkahle/ggmap

# See the underlying tibble
ps("ropensci")

#> - "ropensci" ---------------------------------- 249 packages in 0.008 seconds -
#>   #     package         version by                   @ title                   
#>   1 100 rtweet          0.7.0   Michael W. Kearney  9M Collecting Twitter Data 
#>   2  64 RSelenium       1.7.7   Ju Yeong Kim        8M R Bindings for 'Selen...
#>   3  60 webmockr        0.7.0   Scott Chamberlain   9d Stubbing and Setting ...
#>   4  50 visdat          0.5.3   Nicholas Tierney    2y Preliminary Visualisa...
#>   5  50 wdman           0.2.5   Ju Yeong Kim        8M 'Webdriver'/'Selenium...
#>   6  49 tabulizer       0.2.2   Tom Paskhalis       2y Bindings for 'Tabula'...
#>   7  48 rfisheries      0.2     Karthik Ram         5y 'Programmatic Interfa...
#>   8  45 chromer         0.1     Matthew Pennell     6y Interface to Chromoso...
#>   9  45 googleLanguageR 0.3.0   Mark Edmondson      6M Call Google's 'Natura...
#>  10  43 jsonld          2.2     Jeroen Ooms         5M JSON for Linking Data   

ps()[]

#> # A tibble: 10 x 14
#>    score package version title description date                maintainer_name
#>    <dbl> <chr>   <pckg_> <chr> <chr>       <dttm>              <chr>          
#>  1  700. rtweet  0.7.0   "Col… "An implem… 2020-01-08 22:00:10 Michael W. Kea…
#>  2  451. RSelen… 1.7.7   "R B… "Provides … 2020-02-03 19:00:03 Ju Yeong Kim   
#>  3  418. webmoc… 0.7.0   "Stu… "Stubbing … 2020-09-30 18:40:02 Scott Chamberl…
#>  4  351. visdat  0.5.3   "Pre… "Create pr… 2019-02-15 14:30:03 Nicholas Tiern…
#>  5  350. wdman   0.2.5   "'We… "There are… 2020-01-31 21:30:02 Ju Yeong Kim   
#>  6  345. tabuli… 0.2.2   "Bin… "Bindings … 2018-06-07 18:16:06 Tom Paskhalis  
#>  7  337. rfishe… 0.2     "'Pr… "A program… 2016-02-19 08:50:03 Karthik Ram    
#>  8  317. chromer 0.1     "Int… "A program… 2015-01-13 10:27:09 Matthew Pennell
#>  9  313. google… 0.3.0   "Cal… "Call 'Goo… 2020-04-19 12:40:02 Mark Edmondson 
#> 10  297. jsonld  2.2     "JSO… "JSON-LD i… 2020-05-27 06:20:03 Jeroen Ooms    
#> # … with 7 more variables: maintainer_email <chr>, revdeps <int>,
#> #   downloads_last_month <int>, license <chr>, url <chr>, bugreports <chr>,
#> #   package_data <I<list>>