Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1, size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

# S3 method for pkg_search_result
summary(object, ...)

# S3 method for pkg_search_result
print(x, ...)

Arguments

query

Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

object

Object to summarize.

...

Additional arguments, ignored currently.

x

Object to print.

Value

A data frame with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10001 milliseconds

# Pagination
ps("networks")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10000 milliseconds
more()
#> Error in "more()": ! No query, start with 'pkg_search()'

# Details
ps("visualization")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10000 milliseconds
ps()
#> Error in "pkg_search_again()": ! No query given, and no previous query

# See the underlying data frame
ps("ropensci")
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [search.r-pkg.org:80] Connection timed out after 10000 milliseconds
ps()[]
#> Error in "pkg_search_again()": ! No query given, and no previous query