Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1, size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

# S3 method for pkg_search_result
summary(object, ...)

# S3 method for pkg_search_result
print(x, ...)

Arguments

query

Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

object

Object to summarize.

...

Additional arguments, ignored currently.

x

Object to print.

Value

A data frame with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")
#> - "survival" --------------------------------- 1131 packages in 0.021 seconds -
#>   #     package         version   by                      @ title              
#>   1 100 survival        3.5.8     Terry M Therneau       2M Survival Analysis  
#>   2  10 survminer       0.4.9     Alboukadel Kassambara  3y Drawing Survival...
#>   3   9 flexsurv        2.2.2     Christopher Jackson    1y Flexible Paramet...
#>   4   6 rstpm2          1.6.3     Mark Clements          5M Smooth Survival ...
#>   5   6 rpart           4.1.23    Beth Atkinson          5M Recursive Partit...
#>   6   5 randomForestSRC 3.2.3     Udaya B. Kogalur       5M Fast Unified Ran...
#>   7   5 relsurv         2.2.9     Damjan Manevski        1y Relative Survival  
#>   8   5 muhaz           1.2.6.4   David Winsemius        3y Hazard Function ...
#>   9   5 pec             2023.4.12 Thomas A. Gerds        1y Prediction Error...
#>  10   4 timereg         2.0.5     Thomas Scheike         1y Flexible Regress...

# Pagination
ps("networks")
#> - "networks" ---------------------------------- 959 packages in 0.008 seconds -
#>   #     package    version   by                    @ title                     
#>   1 100 igraph     2.0.3     Kirill Müller        1M Network Analysis and Vi...
#>   2  54 network    1.18.2    Carter T. Butts      5M Classes for Relational ...
#>   3  39 nnet       7.3.19    Brian Ripley         1y Feed-Forward Neural Net...
#>   4  34 DiagrammeR 1.0.11    Richard Iannone      3M Graph/Network Visualiza...
#>   5  34 RCurl      1.98.1.14 CRAN Team            3M General Network (HTTP/F...
#>   6  28 ggraph     2.2.1     Thomas Lin Pedersen  1M An Implementation of Gr...
#>   7  26 sna        2.7.2     Carter T. Butts      5M Tools for Social Networ...
#>   8  25 visNetwork 2.1.2     Benoit Thieurmel     2y Network Visualization u...
#>   9  21 snow       0.4.4     Luke Tierney         2y Simple Network of Works...
#>  10  19 neuralnet  1.44.2    Marvin N. Wright     5y Training of Neural Netw...
more()
#> - "networks" ---------------------------------- 959 packages in 0.012 seconds -
#>   #    package         version by                    @ title                   
#>  11 16 torch           0.12.0  Daniel Falbel        4M Tensors and Neural Ne...
#>  12 15 ergm            4.6.0   Pavel N. Krivitsky   4M Fit, Simulate and Dia...
#>  13 15 networkD3       0.4     Christopher Gandrud  7y D3 JavaScript Network...
#>  14 13 bnlearn         4.9.3   Marco Scutari        1M Bayesian Network Stru...
#>  15 13 WGCNA           1.72.5  Peter Langfelder     5M Weighted Correlation ...
#>  16 13 intergraph      2.0.4   Michał Bojanowski    3M Coercion Routines for...
#>  17 12 spatstat.linnet 3.1.5   Adrian Baddeley     27d Linear Networks Funct...
#>  18 12 diagram         1.6.5   Karline Soetaert     4y Functions for Visuali...
#>  19 12 RSNNS           0.4.17  Christoph Bergmeir   5M Neural Networks using...
#>  20 11 networkDynamic  0.11.4  Skye Bender-deMoll   4M Dynamic Extensions fo...

# Details
ps("visualization")
#> - "visualization" ---------------------------- 1733 packages in 0.043 seconds -
#>   #     package    version by                    @ title                       
#>   1 100 ggplot2    3.5.0   Thomas Lin Pedersen  2M Create Elegant Data Visua...
#>   2  60 scales     1.3.0   Thomas Lin Pedersen  5M Scale Functions for Visua...
#>   3  56 igraph     2.0.3   Kirill Müller        1M Network Analysis and Visu...
#>   4  30 rgl        1.3.1   Duncan Murdoch       2M 3D Visualization Using Op...
#>   5  27 vdiffr     1.0.7   Lionel Henry         7M Visual Regression Testing...
#>   6  19 DiagrammeR 1.0.11  Richard Iannone      3M Graph/Network Visualization 
#>   7  17 corrplot   0.92    Taiyun Wei           2y Visualization of a Correl...
#>   8  16 circlize   0.4.16  Zuguang Gu           2M Circular Visualization      
#>   9  15 ROCR       1.0.11  Felix G.M. Ernst     4y Visualizing the Performan...
#>  10  14 visNetwork 2.1.2   Benoit Thieurmel     2y Network Visualization usi...
ps()
#> - "visualization" ---------------------------- 1733 packages in 0.043 seconds -
#> 
#> 1 ggplot2 @ 3.5.0                             Thomas Lin Pedersen, 2 months ago
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar
#>   of Graphics". You provide the data, tell 'ggplot2' how to map
#>   variables to aesthetics, what graphical primitives to use, and it
#>   takes care of the details.
#>   https://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 2 scales @ 1.3.0                              Thomas Lin Pedersen, 5 months ago
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 3 igraph @ 2.0.3                               Kirill Müller, about a month ago
#> ----------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large
#>   graphs very well and provides functions for generating random and
#>   regular graphs, graph visualization, centrality methods and much
#>   more.
#>   https://r.igraph.org/
#>   https://igraph.org/
#>   https://igraph.discourse.group/
#> 
#> 4 rgl @ 1.3.1                                      Duncan Murdoch, 2 months ago
#> -------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as
#>   well as functions for constructing representations of geometric
#>   objects (cube3d(), etc.).  Output may be on screen using OpenGL, or
#>   to various standard 3D file formats including WebGL, PLY, OBJ, STL as
#>   well as 2D image formats, including PNG, Postscript, SVG, PGF.
#>   https://github.com/dmurdoch/rgl
#>   https://dmurdoch.github.io/rgl/
#> 
#> 5 vdiffr @ 1.0.7                                     Lionel Henry, 7 months ago
#> ----------------
#>   # Visual Regression Testing and Graphical Diffing
#>   An extension to the 'testthat' package that makes it easy to add
#>   graphical unit tests. It provides a Shiny application to manage the
#>   test cases.
#>   https://vdiffr.r-lib.org/
#>   https://github.com/r-lib/vdiffr
#> 
#> 6 DiagrammeR @ 1.0.11                             Richard Iannone, 3 months ago
#> ---------------------
#>   # Graph/Network Visualization
#>   Build graph/network structures using functions for stepwise addition
#>   and deletion of nodes and edges. Work with data available in tables
#>   for bulk addition of nodes, edges, and associated metadata. Use graph
#>   selections and traversals to apply changes to specific nodes or
#>   edges. A wide selection of graph algorithms allow for the analysis of
#>   graphs. Visualize the graphs and take advantage of any aesthetic
#>   properties assigned to nodes and edges.
#>   https://rich-iannone.github.io/DiagrammeR/
#>   https://github.com/rich-iannone/DiagrammeR
#> 
#> 7 corrplot @ 0.92                                       Taiyun Wei, 2 years ago
#> -----------------
#>   # Visualization of a Correlation Matrix
#>   Provides a visual exploratory tool on correlation matrix that
#>   supports automatic variable reordering to help detect hidden patterns
#>   among variables.
#>   https://github.com/taiyun/corrplot
#> 
#> 8 circlize @ 0.4.16                                    Zuguang Gu, 2 months ago
#> -------------------
#>   # Circular Visualization
#>   Circular layout is an efficient way for the visualization of huge
#>   amounts of information. Here this package provides an implementation
#>   of circular layout generation in R as well as an enhancement of
#>   available software. The flexibility of the package is based on the
#>   usage of low-level graphics functions such that self-defined
#>   high-level graphics can be easily implemented by users for specific
#>   purposes. Together with the seamless connection between the powerful
#>   computational and visual environment in R, it gives users more
#>   convenience and freedom to design figures for better understanding
#>   complex patterns behind multiple dimensional data. The package is
#>   described in Gu et al. 2014 <doi:10.1093/bioinformatics/btu393>.
#>   https://github.com/jokergoo/circlize
#>   https://jokergoo.github.io/circlize_book/book/
#> 
#> 9 ROCR @ 1.0.11                                   Felix G.M. Ernst, 4 years ago
#> ---------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts, and
#>   precision/recall plots are popular examples of trade-off
#>   visualizations for specific pairs of performance measures. ROCR is a
#>   flexible tool for creating cutoff-parameterized 2D performance curves
#>   by freely combining two from over 25 performance measures (new
#>   performance measures can be added using a standard interface). Curves
#>   from different cross-validation or bootstrapping runs can be averaged
#>   by different methods, and standard deviations, standard errors or box
#>   plots can be used to visualize the variability across the runs. The
#>   parameterization can be visualized by printing cutoff values at the
#>   corresponding curve positions, or by coloring the curve according to
#>   cutoff. All components of a performance plot can be quickly adjusted
#>   using a flexible parameter dispatching mechanism. Despite its
#>   flexibility, ROCR is easy to use, with only three commands and
#>   reasonable default values for all optional parameters.
#>   http://ipa-tys.github.io/ROCR/
#> 
#> 10 visNetwork @ 2.1.2                             Benoit Thieurmel, 2 years ago
#> ---------------------
#>   # Network Visualization using 'vis.js' Library
#>   Provides an R interface to the 'vis.js' JavaScript charting library.
#>   It allows an interactive visualization of networks.
#>   https://datastorm-open.github.io/visNetwork/

# See the underlying data frame
ps("ropensci")
#> - "ropensci" ---------------------------------- 266 packages in 0.023 seconds -
#>   #     package         version by                       @ title               
#>   1 100 vcr             1.2.2   Scott Chamberlain      10M Record 'HTTP' Cal...
#>   2  79 RSelenium       1.7.9   Ju Yeong Kim            2y R Bindings for 'S...
#>   3  72 webmockr        0.9.0   Scott Chamberlain       1y Stubbing and Sett...
#>   4  57 jsonld          2.2     Jeroen Ooms             4y JSON for Linking ...
#>   5  56 fastMatMR       1.2.5   Rohit Goswami           6M High-Performance ...
#>   6  56 charlatan       0.5.1   Roel M. Hogervorst      7M Make Fake Data      
#>   7  55 rfisheries      0.2     Karthik Ram             8y 'Programmatic Int...
#>   8  54 googleLanguageR 0.3.0   Mark Edmondson          4y Call Google's 'Na...
#>   9  53 mcbette         1.15.2  Richèl J.C. Bilderbeek  7M Model Comparison ...
#>  10  47 taxize          0.9.100 Zachary Foster          2y Taxonomic Informa...
ps()[]
#> # A data frame: 10 × 14
#>    score package   version title description date                maintainer_name
#>    <dbl> <chr>     <pckg_> <chr> <chr>       <dttm>              <chr>          
#>  1  699. vcr       1.2.2   "Rec… "Record te… 2023-06-25 11:30:02 Scott Chamberl…
#>  2  551. RSelenium 1.7.9   "R B… "Provides … 2022-09-02 07:10:11 Ju Yeong Kim   
#>  3  501. webmockr  0.9.0   "Stu… "Stubbing … 2023-02-28 14:20:02 Scott Chamberl…
#>  4  396. jsonld    2.2     "JSO… "JSON-LD i… 2020-05-27 06:20:03 Jeroen Ooms    
#>  5  393. fastMatMR 1.2.5   "Hig… "An interf… 2023-11-03 21:00:06 Rohit Goswami  
#>  6  389. charlatan 0.5.1   "Mak… "Make fake… 2023-09-13 12:40:02 Roel M. Hogerv…
#>  7  386. rfisheri… 0.2     "'Pr… "A program… 2016-02-19 08:50:03 Karthik Ram    
#>  8  374. googleLa… 0.3.0   "Cal… "Call 'Goo… 2020-04-19 12:40:02 Mark Edmondson 
#>  9  374. mcbette   1.15.2  "Mod… "'BEAST2' … 2023-09-27 08:00:02 Richèl J.C. Bi…
#> 10  330. taxize    0.9.100 "Tax… "Interacts… 2022-04-22 07:30:02 Zachary Foster 
#> # ℹ 7 more variables: maintainer_email <chr>, revdeps <int>,
#> #   downloads_last_month <int>, license <chr>, url <chr>, bugreports <chr>,
#> #   package_data <I<list>>