Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1, size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

# S3 method for pkg_search_result
summary(object, ...)

# S3 method for pkg_search_result
print(x, ...)

Arguments

query

Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

object

Object to summarize.

...

Additional arguments, ignored currently.

x

Object to print.

Value

A data frame with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")
#> - "survival" --------------------------------- 1080 packages in 0.012 seconds -
#>   #     package         version   by                      @ title              
#>   1 100 survival        3.5.7     Terry M Therneau       3M Survival Analysis  
#>   2  11 survminer       0.4.9     Alboukadel Kassambara  3y Drawing Survival...
#>   3  10 flexsurv        2.2.2     Christopher Jackson   10M Flexible Paramet...
#>   4   7 rstpm2          1.6.2     Mark Clements          9M Smooth Survival ...
#>   5   6 randomForestSRC 3.2.2     Udaya B. Kogalur       6M Fast Unified Ran...
#>   6   5 pec             2023.4.12 Thomas A. Gerds        7M Prediction Error...
#>   7   5 muhaz           1.2.6.4   David Winsemius        3y Hazard Function ...
#>   8   5 relsurv         2.2.9     Damjan Manevski       11M Relative Survival  
#>   9   5 timereg         2.0.5     Thomas Scheike        10M Flexible Regress...
#>  10   4 survRM2         1.0.4     Hajime Uno             1y Comparing Restri...

# Pagination
ps("networks")
#> - "networks" ---------------------------------- 920 packages in 0.011 seconds -
#>   #     package    version   by                    @ title                     
#>   1 100 igraph     1.5.1     Kirill Müller        3M Network Analysis and Vi...
#>   2  48 network    1.18.1    Carter T. Butts     10M Classes for Relational ...
#>   3  38 nnet       7.3.19    Brian Ripley         7M Feed-Forward Neural Net...
#>   4  33 DiagrammeR 1.0.10    Richard Iannone      6M Graph/Network Visualiza...
#>   5  31 RCurl      1.98.1.13 CRAN Team           18d General Network (HTTP/F...
#>   6  29 ggraph     2.1.0     Thomas Lin Pedersen  1y An Implementation of Gr...
#>   7  27 visNetwork 2.1.2     Benoit Thieurmel     1y Network Visualization u...
#>   8  25 sna        2.7.1     Carter T. Butts     10M Tools for Social Networ...
#>   9  21 snow       0.4.4     Luke Tierney         2y Simple Network of Works...
#>  10  20 neuralnet  1.44.2    Marvin N. Wright     5y Training of Neural Netw...
more()
#> - "networks" ---------------------------------- 920 packages in 0.012 seconds -
#>   #    package        version by                    @ title                    
#>  11 16 networkD3      0.4     Christopher Gandrud  7y D3 JavaScript Network ...
#>  12 16 ergm           4.5.0   Pavel N. Krivitsky   6M Fit, Simulate and Diag...
#>  13 15 bnlearn        4.9     Marco Scutari        2M Bayesian Network Struc...
#>  14 15 torch          0.11.0  Daniel Falbel        6M Tensors and Neural Net...
#>  15 13 WGCNA          1.72.1  Peter Langfelder    10M Weighted Correlation N...
#>  16 13 intergraph     2.0.3   Michał Bojanowski    3M Coercion Routines for ...
#>  17 12 RSNNS          0.4.16  Christoph Bergmeir   7M Neural Networks using ...
#>  18 12 diagram        1.6.5   Karline Soetaert     3y Functions for Visualis...
#>  19 11 networkDynamic 0.11.3  Skye Bender-deMoll   9M Dynamic Extensions for...
#>  20 11 bipartite      2.18    Carsten F. Dormann   1y Visualising Bipartite ...

# Details
ps("visualization")
#> - "visualization" ---------------------------- 1631 packages in 0.012 seconds -
#>   #     package    version by                    @ title                       
#>   1 100 igraph     1.5.1   Kirill Müller        3M Network Analysis and Visu...
#>   2 100 scales     1.2.1   Hadley Wickham       1y Scale Functions for Visua...
#>   3  55 rgl        1.2.1   Duncan Murdoch       5M 3D Visualization Using Op...
#>   4  49 ggplot2    3.4.4   Thomas Lin Pedersen  1M Create Elegant Data Visua...
#>   5  32 DiagrammeR 1.0.10  Richard Iannone      6M Graph/Network Visualization 
#>   6  29 circlize   0.4.15  Zuguang Gu           2y Circular Visualization      
#>   7  29 corrplot   0.92    Taiyun Wei           2y Visualization of a Correl...
#>   8  26 visNetwork 2.1.2   Benoit Thieurmel     1y Network Visualization usi...
#>   9  26 ROCR       1.0.11  Felix G.M. Ernst     4y Visualizing the Performan...
#>  10  24 vcd        1.4.11  David Meyer         10M Visualizing Categorical Data
ps()
#> - "visualization" ---------------------------- 1631 packages in 0.012 seconds -
#> 
#> 1 igraph @ 1.5.1                                    Kirill Müller, 3 months ago
#> ----------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large
#>   graphs very well and provides functions for generating random and
#>   regular graphs, graph visualization, centrality methods and much
#>   more.
#>   https://r.igraph.org/
#>   https://igraph.org/
#>   https://igraph.discourse.group/
#> 
#> 2 scales @ 1.2.1                               Hadley Wickham, about a year ago
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 3 rgl @ 1.2.1                                      Duncan Murdoch, 5 months ago
#> -------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as
#>   well as functions for constructing representations of geometric
#>   objects (cube3d(), etc.).  Output may be on screen using OpenGL, or
#>   to various standard 3D file formats including WebGL, PLY, OBJ, STL as
#>   well as 2D image formats, including PNG, Postscript, SVG, PGF.
#>   https://github.com/dmurdoch/rgl
#>   https://dmurdoch.github.io/rgl/
#> 
#> 4 ggplot2 @ 3.4.4                        Thomas Lin Pedersen, about a month ago
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar
#>   of Graphics". You provide the data, tell 'ggplot2' how to map
#>   variables to aesthetics, what graphical primitives to use, and it
#>   takes care of the details.
#>   https://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 5 DiagrammeR @ 1.0.10                             Richard Iannone, 6 months ago
#> ---------------------
#>   # Graph/Network Visualization
#>   Build graph/network structures using functions for stepwise addition
#>   and deletion of nodes and edges. Work with data available in tables
#>   for bulk addition of nodes, edges, and associated metadata. Use graph
#>   selections and traversals to apply changes to specific nodes or
#>   edges. A wide selection of graph algorithms allow for the analysis of
#>   graphs. Visualize the graphs and take advantage of any aesthetic
#>   properties assigned to nodes and edges.
#>   https://github.com/rich-iannone/DiagrammeR
#> 
#> 6 circlize @ 0.4.15                                     Zuguang Gu, 2 years ago
#> -------------------
#>   # Circular Visualization
#>   Circular layout is an efficient way for the visualization of huge
#>   amounts of information. Here this package provides an implementation
#>   of circular layout generation in R as well as an enhancement of
#>   available software. The flexibility of the package is based on the
#>   usage of low-level graphics functions such that self-defined
#>   high-level graphics can be easily implemented by users for specific
#>   purposes. Together with the seamless connection between the powerful
#>   computational and visual environment in R, it gives users more
#>   convenience and freedom to design figures for better understanding
#>   complex patterns behind multiple dimensional data. The package is
#>   described in Gu et al. 2014 <doi:10.1093/bioinformatics/btu393>.
#>   https://github.com/jokergoo/circlize
#>   https://jokergoo.github.io/circlize_book/book/
#> 
#> 7 corrplot @ 0.92                                       Taiyun Wei, 2 years ago
#> -----------------
#>   # Visualization of a Correlation Matrix
#>   Provides a visual exploratory tool on correlation matrix that
#>   supports automatic variable reordering to help detect hidden patterns
#>   among variables.
#>   https://github.com/taiyun/corrplot
#> 
#> 8 visNetwork @ 2.1.2                         Benoit Thieurmel, about a year ago
#> --------------------
#>   # Network Visualization using 'vis.js' Library
#>   Provides an R interface to the 'vis.js' JavaScript charting library.
#>   It allows an interactive visualization of networks.
#>   https://datastorm-open.github.io/visNetwork/
#> 
#> 9 ROCR @ 1.0.11                                   Felix G.M. Ernst, 4 years ago
#> ---------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts, and
#>   precision/recall plots are popular examples of trade-off
#>   visualizations for specific pairs of performance measures. ROCR is a
#>   flexible tool for creating cutoff-parameterized 2D performance curves
#>   by freely combining two from over 25 performance measures (new
#>   performance measures can be added using a standard interface). Curves
#>   from different cross-validation or bootstrapping runs can be averaged
#>   by different methods, and standard deviations, standard errors or box
#>   plots can be used to visualize the variability across the runs. The
#>   parameterization can be visualized by printing cutoff values at the
#>   corresponding curve positions, or by coloring the curve according to
#>   cutoff. All components of a performance plot can be quickly adjusted
#>   using a flexible parameter dispatching mechanism. Despite its
#>   flexibility, ROCR is easy to use, with only three commands and
#>   reasonable default values for all optional parameters.
#>   http://ipa-tys.github.io/ROCR/
#> 
#> 10 vcd @ 1.4.11                                      David Meyer, 10 months ago
#> ---------------
#>   # Visualizing Categorical Data
#>   Visualization techniques, data sets, summary and inference procedures
#>   aimed particularly at categorical data. Special emphasis is given to
#>   highly extensible grid graphics. The package was package was
#>   originally inspired by the book "Visualizing Categorical Data" by
#>   Michael Friendly and is now the main support package for a new book,
#>   "Discrete Data Analysis with R" by Michael Friendly and David Meyer
#>   (2015).

# See the underlying data frame
ps("ropensci")
#> - "ropensci" ---------------------------------- 260 packages in 0.012 seconds -
#>   #     package         version by                       @ title               
#>   1 100 vcr             1.2.2   Scott Chamberlain       5M Record 'HTTP' Cal...
#>   2  81 RSelenium       1.7.9   Ju Yeong Kim            1y R Bindings for 'S...
#>   3  77 webmockr        0.9.0   Scott Chamberlain       9M Stubbing and Sett...
#>   4  56 rfisheries      0.2     Karthik Ram             8y 'Programmatic Int...
#>   5  52 jsonld          2.2     Jeroen Ooms             3y JSON for Linking ...
#>   6  51 charlatan       0.5.1   Roel M. Hogervorst      2M Make Fake Data      
#>   7  51 mcbette         1.15.2  Richèl J.C. Bilderbeek  2M Model Comparison ...
#>   8  49 googleLanguageR 0.3.0   Mark Edmondson          4y Call Google's 'Na...
#>   9  49 taxize          0.9.100 Zachary Foster          2y Taxonomic Informa...
#>  10  48 fastMatMR       1.2.5   Rohit Goswami          17d High-Performance ...
ps()[]
#> # A data frame: 10 × 14
#>    score package   version title description date                maintainer_name
#>    <dbl> <chr>     <pckg_> <chr> <chr>       <dttm>              <chr>          
#>  1  701. vcr       1.2.2   "Rec… "Record te… 2023-06-25 11:30:02 Scott Chamberl…
#>  2  570. RSelenium 1.7.9   "R B… "Provides … 2022-09-02 07:10:11 Ju Yeong Kim   
#>  3  537. webmockr  0.9.0   "Stu… "Stubbing … 2023-02-28 14:20:02 Scott Chamberl…
#>  4  391. rfisheri… 0.2     "'Pr… "A program… 2016-02-19 08:50:03 Karthik Ram    
#>  5  365. jsonld    2.2     "JSO… "JSON-LD i… 2020-05-27 06:20:03 Jeroen Ooms    
#>  6  359. charlatan 0.5.1   "Mak… "Make fake… 2023-09-13 12:40:02 Roel M. Hogerv…
#>  7  358. mcbette   1.15.2  "Mod… "'BEAST2' … 2023-09-27 08:00:02 Richèl J.C. Bi…
#>  8  346. googleLa… 0.3.0   "Cal… "Call 'Goo… 2020-04-19 12:40:02 Mark Edmondson 
#>  9  343. taxize    0.9.100 "Tax… "Interacts… 2022-04-22 07:30:02 Zachary Foster 
#> 10  335. fastMatMR 1.2.5   "Hig… "An interf… 2023-11-03 21:00:06 Rohit Goswami  
#> # ℹ 7 more variables: maintainer_email <chr>, revdeps <int>,
#> #   downloads_last_month <int>, license <chr>, url <chr>, bugreports <chr>,
#> #   package_data <I<list>>