Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1, size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

# S3 method for pkg_search_result
summary(object, ...)

# S3 method for pkg_search_result
print(x, ...)

Arguments

query

Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

object

Object to summarize.

...

Additional arguments, ignored currently.

x

Object to print.

Value

A data frame with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")
#> - "survival" ---------------------------------- 962 packages in 0.006 seconds -
#>   #     package         version 
#>   1 100 survival        3.3.1   
#>   2   7 KMsurv          0.1.5   
#>   3   6 rpart           4.1.16  
#>   4   6 timereg         2.0.2   
#>   5   6 randomForestSRC 3.1.0   
#>   6   5 pec             2022.5.4
#>   7   4 muhaz           1.2.6.4 
#>   8   4 survivalROC     1.0.3   
#>   9   4 survC1          1.0.3   
#>  10   4 flexsurv        2.1     
#>  by                                                                       @
#>  Terry M Therneau                                                        4M
#>  Jun Yan                                                                10y
#>  Beth Atkinson                                                           5M
#>  Thomas Scheike                                                          2M
#>  Udaya B. Kogalur                                                        2M
#>  Thomas A. Gerds                                                         1M
#>  David Winsemius                                                         1y
#>  Paramita Saha-Chaudhuri<U+000a><paramita.sahachaudhuri.work@gmail.com>  9y
#>  Hajime Uno                                                              1y
#>  Christopher Jackson                                                     9M

# Pagination
ps("networks")
#> - "networks" ---------------------------------- 812 packages in 0.006 seconds -
#>   #     package        version  by                   @ title                   
#>   1 100 igraph         1.3.2    Tamás Nepusz        4d Network Analysis and ...
#>   2  67 RCurl          1.98.1.7 CRAN Team           8d General Network (HTTP...
#>   3  56 network        1.17.2   Carter T. Butts    27d Classes for Relationa...
#>   4  49 nnet           7.3.17   Brian Ripley        5M Feed-Forward Neural N...
#>   5  34 sna            2.7      Carter T. Butts    16d Tools for Social Netw...
#>   6  28 snow           0.4.4    Luke Tierney        8M Simple Network of Wor...
#>   7  23 ergm           4.2.2    Pavel N. Krivitsky 16d Fit, Simulate and Dia...
#>   8  18 networkDynamic 0.11.2   Skye Bender-deMoll  1M Dynamic Extensions fo...
#>   9  16 diagram        1.6.5    Karline Soetaert    2y Functions for Visuali...
#>  10  15 RSNNS          0.4.14   Christoph Bergmeir 10M Neural Networks using...
more()
#> - "networks" ---------------------------------- 812 packages in 0.006 seconds -
#>   #    package    version by                   @ title                         
#>  11 14 neuralnet  1.44.2  Marvin N. Wright    3y Training of Neural Networks   
#>  12 12 latentnet  2.10.6  Pavel N. Krivitsky  1M Latent Position and Cluster...
#>  13 11 snowFT     1.6.0   Hana Sevcikova      5y Fault Tolerant Simple Netwo...
#>  14 11 gRain      1.3.10  Søren Højsgaard     1M Graphical Independence Netw...
#>  15 10 randnet    0.5     Tianxi Li           5M Random Network Model Estima...
#>  16 10 GeneNet    1.2.16  Korbinian Strimmer  7M Modeling and Inferring Gene...
#>  17 10 ndtv       0.13.2  Skye Bender-deMoll  8M Network Dynamic Temporal Vi...
#>  18 10 qrnn       2.0.5   Alex J. Cannon      3y Quantile Regression Neural ...
#>  19 10 WGCNA      1.71    Peter Langfelder    2M Weighted Correlation Networ...
#>  20  9 igraphdata 1.0.1   Gabor Csardi        7y A Collection of Network Dat...

# Details
ps("visualization")
#> - "visualization" ---------------------------- 1436 packages in 0.006 seconds -
#>   #     package  version by                    @ title                         
#>   1 100 ggplot2  3.3.6   Thomas Lin Pedersen  2M Create Elegant Data Visuali...
#>   2  93 igraph   1.3.2   Tamás Nepusz         4d Network Analysis and Visual...
#>   3  87 rgl      0.109.2 Duncan Murdoch       7d 3D Visualization Using OpenGL 
#>   4  55 scales   1.2.0   Hadley Wickham       2M Scale Functions for Visuali...
#>   5  47 vcd      1.4.10  David Meyer          8d Visualizing Categorical Data  
#>   6  37 ROCR     1.0.11  Felix G.M. Ernst     2y Visualizing the Performance...
#>   7  25 klaR     1.7.0   Uwe Ligges           3M Classification and Visualiz...
#>   8  21 ggmap    3.0.0   ORPHANED             3y Spatial Visualization with ...
#>   9  20 lattice  0.20.45 Deepayan Sarkar      9M Trellis Graphics for R        
#>  10  19 corrplot 0.92    Taiyun Wei           7M Visualization of a Correlat...
ps()
#> - "visualization" ---------------------------- 1436 packages in 0.006 seconds -
#> 
#> 1 ggplot2 @ 3.3.6                             Thomas Lin Pedersen, 2 months ago
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar
#>   of Graphics". You provide the data, tell 'ggplot2' how to map
#>   variables to aesthetics, what graphical primitives to use, and it
#>   takes care of the details.
#>   https://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 2 igraph @ 1.3.2                                       Tamás Nepusz, 4 days ago
#> ----------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large
#>   graphs very well and provides functions for generating random and
#>   regular graphs, graph visualization, centrality methods and much
#>   more.
#>   https://igraph.org
#>   https://igraph.discourse.group/
#> 
#> 3 rgl @ 0.109.2                                      Duncan Murdoch, 7 days ago
#> ---------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as
#>   well as functions for constructing representations of geometric
#>   objects (cube3d(), etc.).  Output may be on screen using OpenGL, or
#>   to various standard 3D file formats including WebGL, PLY, OBJ, STL as
#>   well as 2D image formats, including PNG, Postscript, SVG, PGF.
#>   https://github.com/dmurdoch/rgl
#>   https://dmurdoch.github.io/rgl/
#> 
#> 4 scales @ 1.2.0                                   Hadley Wickham, 2 months ago
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 5 vcd @ 1.4.10                                          David Meyer, 8 days ago
#> --------------
#>   # Visualizing Categorical Data
#>   Visualization techniques, data sets, summary and inference procedures
#>   aimed particularly at categorical data. Special emphasis is given to
#>   highly extensible grid graphics. The package was package was
#>   originally inspired by the book "Visualizing Categorical Data" by
#>   Michael Friendly and is now the main support package for a new book,
#>   "Discrete Data Analysis with R" by Michael Friendly and David Meyer
#>   (2015).
#> 
#> 6 ROCR @ 1.0.11                                   Felix G.M. Ernst, 2 years ago
#> ---------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts, and
#>   precision/recall plots are popular examples of trade-off
#>   visualizations for specific pairs of performance measures. ROCR is a
#>   flexible tool for creating cutoff-parameterized 2D performance curves
#>   by freely combining two from over 25 performance measures (new
#>   performance measures can be added using a standard interface). Curves
#>   from different cross-validation or bootstrapping runs can be averaged
#>   by different methods, and standard deviations, standard errors or box
#>   plots can be used to visualize the variability across the runs. The
#>   parameterization can be visualized by printing cutoff values at the
#>   corresponding curve positions, or by coloring the curve according to
#>   cutoff. All components of a performance plot can be quickly adjusted
#>   using a flexible parameter dispatching mechanism. Despite its
#>   flexibility, ROCR is easy to use, with only three commands and
#>   reasonable default values for all optional parameters.
#>   http://ipa-tys.github.io/ROCR/
#> 
#> 7 klaR @ 1.7.0                                         Uwe Ligges, 3 months ago
#> --------------
#>   # Classification and Visualization
#>   Miscellaneous functions for classification and visualization, e.g.
#>   regularized discriminant analysis, sknn() kernel-density naive Bayes,
#>   an interface to 'svmlight' and stepclass() wrapper variable selection
#>   for supervised classification, partimat() visualization of
#>   classification rules and shardsplot() of cluster results as well as
#>   kmodes() clustering for categorical data, corclust() variable
#>   clustering, variable extraction from different variable clustering
#>   models and weight of evidence preprocessing.
#>   http://www.statistik.tu-dortmund.de
#> 
#> 8 ggmap @ 3.0.0                                           ORPHANED, 3 years ago
#> ---------------
#>   # Spatial Visualization with ggplot2
#>   A collection of functions to visualize spatial data and models on top
#>   of static maps from various online sources (e.g Google Maps and
#>   Stamen Maps). It includes tools common to those tasks, including
#>   functions for geolocation and routing.
#>   https://github.com/dkahle/ggmap
#> 
#> 9 lattice @ 0.20.45                               Deepayan Sarkar, 9 months ago
#> -------------------
#>   # Trellis Graphics for R
#>   A powerful and elegant high-level data visualization system inspired
#>   by Trellis graphics, with an emphasis on multivariate data. Lattice
#>   is sufficient for typical graphics needs, and is also flexible enough
#>   to handle most nonstandard requirements. See ?Lattice for an
#>   introduction.
#>   http://lattice.r-forge.r-project.org/
#> 
#> 10 corrplot @ 0.92                                     Taiyun Wei, 7 months ago
#> ------------------
#>   # Visualization of a Correlation Matrix
#>   Provides a visual exploratory tool on correlation matrix that
#>   supports automatic variable reordering to help detect hidden patterns
#>   among variables.
#>   https://github.com/taiyun/corrplot

# See the underlying data frame
ps("ropensci")
#> - "ropensci" ---------------------------------- 242 packages in 0.005 seconds -
#>   #     package         version by                    @ title                  
#>   1 100 rfisheries      0.2     Karthik Ram          6y 'Programmatic Interf...
#>   2  96 chromer         0.1     Matthew Pennell      7y Interface to Chromos...
#>   3  81 workloopR       1.1.4   Vikram B. Baliga     1y Analysis of Work Loo...
#>   4  76 rtweet          0.7.0   Michael W. Kearney   2y Collecting Twitter Data
#>   5  74 opentripplanner 0.4.0   Malcolm Morgan       5M Setup and connect to...
#>   6  66 outsider.base   0.1.4   Dom Bennett          1y Base Package for 'Ou...
#>   7  59 hydroscoper     1.4.1   Konstantinos Vantas  1y Interface to the Gre...
#>   8  59 unifir          0.2.1   Michael Mahoney      1M A Unifying API for C...
#>   9  57 terrainr        0.7.0   Michael Mahoney      1M Landscape Visualizat...
#>  10  56 dittodb         0.1.4   Jonathan Keane       5h A Test Environment f...
ps()[]
#> # A data frame: 10 × 14
#>    score package   version title description date                maintainer_name
#>    <dbl> <chr>     <pckg_> <chr> <chr>       <dttm>              <chr>          
#>  1  351. rfisheri… 0.2     "'Pr… "A program… 2016-02-19 08:50:03 Karthik Ram    
#>  2  338. chromer   0.1     "Int… "A program… 2015-01-13 10:27:09 Matthew Pennell
#>  3  283. workloopR 1.1.4   "Ana… "Functions… 2021-05-06 07:10:02 Vikram B. Bali…
#>  4  267. rtweet    0.7.0   "Col… "An implem… 2020-01-08 22:00:10 Michael W. Kea…
#>  5  260. opentrip… 0.4.0   "Set… "Setup and… 2022-01-20 16:02:42 Malcolm Morgan 
#>  6  231. outsider… 0.1.4   "Bas… "Base pack… 2021-04-18 20:40:40 Dom Bennett    
#>  7  209. hydrosco… 1.4.1   "Int… "R interfa… 2021-05-14 15:30:03 Konstantinos V…
#>  8  207. unifir    0.2.1   "A U… "Functions… 2022-05-13 13:10:09 Michael Mahoney
#>  9  201. terrainr  0.7.0   "Lan… "Functions… 2022-05-05 22:20:06 Michael Mahoney
#> 10  197. dittodb   0.1.4   "A T… "Testing a… 2022-06-17 06:10:05 Jonathan Keane 
#> # … with 7 more variables: maintainer_email <chr>, revdeps <int>,
#> #   downloads_last_month <int>, license <chr>, url <chr>, bugreports <chr>,
#> #   package_data <I<list>>