Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1, size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

# S3 method for pkg_search_result
summary(object, ...)

# S3 method for pkg_search_result
print(x, ...)

Arguments

query

Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

object

Object to summarize.

...

Additional arguments, ignored currently.

x

Object to print.

Value

A data frame with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")
#> - "survival" ---------------------------------- 955 packages in 0.007 seconds -
#>   #     package         version  by                      @ title               
#>   1 100 survival        3.4.0    Terry M Therneau       3M Survival Analysis   
#>   2  10 survminer       0.4.9    Alboukadel Kassambara  2y Drawing Survival ...
#>   3  10 flexsurv        2.2      Christopher Jackson    4M Flexible Parametr...
#>   4   6 rpart           4.1.19   Beth Atkinson          5d Recursive Partiti...
#>   5   6 rstpm2          1.5.8    Mark Clements          9d Smooth Survival M...
#>   6   6 pec             2022.5.4 Thomas A. Gerds        6M Prediction Error ...
#>   7   6 muhaz           1.2.6.4  David Winsemius        2y Hazard Function E...
#>   8   6 randomForestSRC 3.1.1    Udaya B. Kogalur       4M Fast Unified Rand...
#>   9   5 relsurv         2.2.8    Damjan Manevski        2M Relative Survival   
#>  10   5 timereg         2.0.3    Thomas Scheike         2d Flexible Regressi...

# Pagination
ps("networks")
#> - "networks" ---------------------------------- 794 packages in 0.008 seconds -
#>   #     package    version  by                    @ title                      
#>   1 100 igraph     1.3.5    Tamás Nepusz         1M Network Analysis and Vis...
#>   2  46 network    1.18.0   Carter T. Butts     20d Classes for Relational Data
#>   3  37 nnet       7.3.18   Brian Ripley        28d Feed-Forward Neural Netw...
#>   4  34 RCurl      1.98.1.9 CRAN Team           23d General Network (HTTP/FT...
#>   5  31 DiagrammeR 1.0.9    Richard Iannone      8M Graph/Network Visualization
#>   6  27 ggraph     2.1.0    Thomas Lin Pedersen 16d An Implementation of Gra...
#>   7  26 visNetwork 2.1.2    Benoit Thieurmel    27d Network Visualization us...
#>   8  25 sna        2.7      Carter T. Butts      5M Tools for Social Network...
#>   9  23 snow       0.4.4    Luke Tierney         1y Simple Network of Workst...
#>  10  18 neuralnet  1.44.2   Marvin N. Wright     4y Training of Neural Networks
more()
#> - "networks" ---------------------------------- 794 packages in 0.007 seconds -
#>   #    package         version by                    @ title                   
#>  11 15 bnlearn         4.8.1   Marco Scutari        1M Bayesian Network Stru...
#>  12 14 ergm            4.2.3   Pavel N. Krivitsky  24d Fit, Simulate and Dia...
#>  13 14 networkD3       0.4     Christopher Gandrud  6y D3 JavaScript Network...
#>  14 14 spatstat.linnet 2.3.2   Adrian Baddeley      8M Linear Networks Funct...
#>  15 13 torch           0.9.0   Daniel Falbel        2d Tensors and Neural Ne...
#>  16 13 WGCNA           1.71    Peter Langfelder     6M Weighted Correlation ...
#>  17 12 intergraph      2.0.2   Michal Bojanowski    7y Coercion Routines for...
#>  18 12 diagram         1.6.5   Karline Soetaert     2y Functions for Visuali...
#>  19 11 RSNNS           0.4.14  Christoph Bergmeir   1y Neural Networks using...
#>  20 11 networkDynamic  0.11.2  Skye Bender-deMoll   6M Dynamic Extensions fo...

# Details
ps("visualization")
#> - "visualization" ---------------------------- 1417 packages in 0.011 seconds -
#>   #     package    version by                    @ title                       
#>   1 100 ggplot2    3.3.6   Thomas Lin Pedersen  6M Create Elegant Data Visua...
#>   2  61 igraph     1.3.5   Tamás Nepusz         1M Network Analysis and Visu...
#>   3  58 scales     1.2.1   Hadley Wickham       2M Scale Functions for Visua...
#>   4  35 rgl        0.110.2 Duncan Murdoch      30d 3D Visualization Using Op...
#>   5  27 vdiffr     1.0.4   Lionel Henry         7M Visual Regression Testing...
#>   6  20 DiagrammeR 1.0.9   Richard Iannone      8M Graph/Network Visualization 
#>   7  18 corrplot   0.92    Taiyun Wei           1y Visualization of a Correl...
#>   8  17 ROCR       1.0.11  Felix G.M. Ernst     2y Visualizing the Performan...
#>   9  17 circlize   0.4.15  Zuguang Gu           6M Circular Visualization      
#>  10  16 visNetwork 2.1.2   Benoit Thieurmel    27d Network Visualization usi...
ps()
#> - "visualization" ---------------------------- 1417 packages in 0.011 seconds -
#> 
#> 1 ggplot2 @ 3.3.6                             Thomas Lin Pedersen, 6 months ago
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar
#>   of Graphics". You provide the data, tell 'ggplot2' how to map
#>   variables to aesthetics, what graphical primitives to use, and it
#>   takes care of the details.
#>   https://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 2 igraph @ 1.3.5                                Tamás Nepusz, about a month ago
#> ----------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large
#>   graphs very well and provides functions for generating random and
#>   regular graphs, graph visualization, centrality methods and much
#>   more.
#>   https://igraph.org
#>   https://igraph.discourse.group/
#> 
#> 3 scales @ 1.2.1                                   Hadley Wickham, 2 months ago
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 4 rgl @ 0.110.2                                     Duncan Murdoch, 30 days ago
#> ---------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as
#>   well as functions for constructing representations of geometric
#>   objects (cube3d(), etc.).  Output may be on screen using OpenGL, or
#>   to various standard 3D file formats including WebGL, PLY, OBJ, STL as
#>   well as 2D image formats, including PNG, Postscript, SVG, PGF.
#>   https://github.com/dmurdoch/rgl
#>   https://dmurdoch.github.io/rgl/
#> 
#> 5 vdiffr @ 1.0.4                                     Lionel Henry, 7 months ago
#> ----------------
#>   # Visual Regression Testing and Graphical Diffing
#>   An extension to the 'testthat' package that makes it easy to add
#>   graphical unit tests. It provides a Shiny application to manage the
#>   test cases.
#>   https://vdiffr.r-lib.org/
#>   https://github.com/r-lib/vdiffr
#> 
#> 6 DiagrammeR @ 1.0.9                              Richard Iannone, 8 months ago
#> --------------------
#>   # Graph/Network Visualization
#>   Build graph/network structures using functions for stepwise addition
#>   and deletion of nodes and edges. Work with data available in tables
#>   for bulk addition of nodes, edges, and associated metadata. Use graph
#>   selections and traversals to apply changes to specific nodes or
#>   edges. A wide selection of graph algorithms allow for the analysis of
#>   graphs. Visualize the graphs and take advantage of any aesthetic
#>   properties assigned to nodes and edges.
#>   https://github.com/rich-iannone/DiagrammeR
#> 
#> 7 corrplot @ 0.92                                  Taiyun Wei, about a year ago
#> -----------------
#>   # Visualization of a Correlation Matrix
#>   Provides a visual exploratory tool on correlation matrix that
#>   supports automatic variable reordering to help detect hidden patterns
#>   among variables.
#>   https://github.com/taiyun/corrplot
#> 
#> 8 ROCR @ 1.0.11                                   Felix G.M. Ernst, 2 years ago
#> ---------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts, and
#>   precision/recall plots are popular examples of trade-off
#>   visualizations for specific pairs of performance measures. ROCR is a
#>   flexible tool for creating cutoff-parameterized 2D performance curves
#>   by freely combining two from over 25 performance measures (new
#>   performance measures can be added using a standard interface). Curves
#>   from different cross-validation or bootstrapping runs can be averaged
#>   by different methods, and standard deviations, standard errors or box
#>   plots can be used to visualize the variability across the runs. The
#>   parameterization can be visualized by printing cutoff values at the
#>   corresponding curve positions, or by coloring the curve according to
#>   cutoff. All components of a performance plot can be quickly adjusted
#>   using a flexible parameter dispatching mechanism. Despite its
#>   flexibility, ROCR is easy to use, with only three commands and
#>   reasonable default values for all optional parameters.
#>   http://ipa-tys.github.io/ROCR/
#> 
#> 9 circlize @ 0.4.15                                    Zuguang Gu, 6 months ago
#> -------------------
#>   # Circular Visualization
#>   Circular layout is an efficient way for the visualization of huge
#>   amounts of information. Here this package provides an implementation
#>   of circular layout generation in R as well as an enhancement of
#>   available software. The flexibility of the package is based on the
#>   usage of low-level graphics functions such that self-defined
#>   high-level graphics can be easily implemented by users for specific
#>   purposes. Together with the seamless connection between the powerful
#>   computational and visual environment in R, it gives users more
#>   convenience and freedom to design figures for better understanding
#>   complex patterns behind multiple dimensional data. The package is
#>   described in Gu et al. 2014 <doi:10.1093/bioinformatics/btu393>.
#>   https://github.com/jokergoo/circlize
#>   https://jokergoo.github.io/circlize_book/book/
#> 
#> 10 visNetwork @ 2.1.2                             Benoit Thieurmel, 27 days ago
#> ---------------------
#>   # Network Visualization using 'vis.js' Library
#>   Provides an R interface to the 'vis.js' JavaScript charting library.
#>   It allows an interactive visualization of networks.
#>   https://datastorm-open.github.io/visNetwork/

# See the underlying data frame
ps("ropensci")
#> - "ropensci" ---------------------------------- 242 packages in 0.008 seconds -
#>   #     package         version by                       @ title               
#>   1 100 vcr             1.0.2   Scott Chamberlain       1y Record 'HTTP' Cal...
#>   2  70 webmockr        0.8.2   Scott Chamberlain       2M Stubbing and Sett...
#>   3  67 RSelenium       1.7.9   Ju Yeong Kim            2M R Bindings for 'S...
#>   4  56 tracerer        2.2.2   Richèl J.C. Bilderbeek  1y Tracer from R       
#>   5  49 rfisheries      0.2     Karthik Ram             7y 'Programmatic Int...
#>   6  49 beastier        2.4.11  Richèl J.C. Bilderbeek  3M Call 'BEAST2'       
#>   7  49 spocc           1.2.0   Scott Chamberlain       2y Interface to Spec...
#>   8  48 mcbette         1.15    Richèl J.C. Bilderbeek  2M Model Comparison ...
#>   9  47 taxize          0.9.100 Zachary Foster          6M Taxonomic Informa...
#>  10  44 googleLanguageR 0.3.0   Mark Edmondson          3y Call Google's 'Na...
ps()[]
#> # A data frame: 10 × 14
#>    score package       version title descr…¹ date                maint…² maint…³
#>    <dbl> <chr>         <pckg_> <chr> <chr>   <dttm>              <chr>   <chr>  
#>  1  762. vcr           1.0.2   "Rec… "Recor… 2021-05-31 17:00:02 Scott … sckott…
#>  2  531. webmockr      0.8.2   "Stu… "Stubb… 2022-08-28 19:20:02 Scott … myrmec…
#>  3  514. RSelenium     1.7.9   "R B… "Provi… 2022-09-02 07:10:11 Ju Yeo… jkim23…
#>  4  427. tracerer      2.2.2   "Tra… "'BEAS… 2021-05-30 08:40:03 Richèl… richel…
#>  5  376. rfisheries    0.2     "'Pr… "A pro… 2016-02-19 08:50:03 Karthi… karthi…
#>  6  375. beastier      2.4.11  "Cal… "'BEAS… 2022-08-11 13:40:04 Richèl… richel…
#>  7  370. spocc         1.2.0   "Int… "A pro… 2021-01-05 19:50:03 Scott … myrmec…
#>  8  362. mcbette       1.15    "Mod… "'BEAS… 2022-08-27 12:30:02 Richèl… richel…
#>  9  354. taxize        0.9.100 "Tax… "Inter… 2022-04-22 07:30:02 Zachar… zachar…
#> 10  339. googleLangua… 0.3.0   "Cal… "Call … 2020-04-19 12:40:02 Mark E… r@sunh…
#> # … with 6 more variables: revdeps <int>, downloads_last_month <int>,
#> #   license <chr>, url <chr>, bugreports <chr>, package_data <I<list>>, and
#> #   abbreviated variable names ¹​description, ²​maintainer_name,
#> #   ³​maintainer_email