Search CRAN packages

pkg_search() starts a new search query, or shows the details of the previous query, if called without arguments.

ps() is an alias to pkg_search().

more() retrieves that next page of results for the previous query.

pkg_search(query = NULL, format = c("short", "long"), from = 1, size = 10)

ps(query = NULL, format = c("short", "long"), from = 1, size = 10)

more(format = NULL, size = NULL)

# S3 method for pkg_search_result
summary(object, ...)

# S3 method for pkg_search_result
print(x, ...)

Arguments

query

Search query string. If this argument is missing or NULL, then the results of the last query are printed, in short and long formats, in turns for successive pkg_search() calls. If this argument is missing, then all other arguments are ignored.

format

Default formatting of the results. short only outputs the name and title of the packages, long also prints the author, last version, full description and URLs. Note that this only affects the default printing, and you can still inspect the full results, even if you specify short here.

from

Where to start listing the results, for pagination.

size

The number of results to list.

object

Object to summarize.

...

Additional arguments, ignored currently.

x

Object to print.

Value

A data frame with columns:

score: Score of the hit. See Section Scoring for some details.

package: Package name.

version: Latest package version.

title: Package title.

description: Short package description.

date: Time stamp of the last release.

maintainer_name: Name of the package maintainer.

maintainer_email: Email address of the package maintainer.

revdeps: Number of (strong and weak) reverse dependencies of the package.

downloads_last_month: Raw number of package downloads last month, from the RStudio CRAN mirror.

license: Package license.

url: Package URL(s).

bugreports: URL of issue tracker, or email address for bug reports.

Details

Note that the search needs a working Internet connection.

Examples

# Example
ps("survival")
#> - "survival" ---------------------------------- 934 packages in 0.006 seconds -
#>   #     package         version    by                      @ title             
#>   1 100 survival        3.2.13     Terry M Therneau       6M Survival Analysis 
#>   2  10 survminer       0.4.9      Alboukadel Kassambara  1y Drawing Surviva...
#>   3   9 flexsurv        2.1        Christopher Jackson    5M Flexible Parame...
#>   4   6 rpart           4.1.16     Beth Atkinson         20d Recursive Parti...
#>   5   6 randomForestSRC 3.0.0      Udaya B. Kogalur       1M Fast Unified Ra...
#>   6   6 rstpm2          1.5.2      Mark Clements          1y Smooth Survival...
#>   7   5 muhaz           1.2.6.4    David Winsemius       10M Hazard Function...
#>   8   5 relsurv         2.2.6      Damjan Manevski        2M Relative Survival 
#>   9   5 pec             2021.10.11 Thomas A. Gerds        4M Prediction Erro...
#>  10   5 timereg         2.0.1      Thomas Scheike         4M Flexible Regres...

# Pagination
ps("networks")
#> - "networks" ---------------------------------- 771 packages in 0.006 seconds -
#>   #     package    version  by                    @ title                      
#>   1 100 igraph     1.2.11   Tamás Nepusz         1M Network Analysis and Vis...
#>   2  49 network    1.17.1   Carter T. Butts      8M Classes for Relational Data
#>   3  39 RCurl      1.98.1.6 CRAN Team            5d General Network (HTTP/FT...
#>   4  38 nnet       7.3.17   Brian Ripley         1M Feed-Forward Neural Netw...
#>   5  31 DiagrammeR 1.0.8    Richard Iannone     20d Graph/Network Visualization
#>   6  26 visNetwork 2.1.0    Benoit Thieurmel     5M Network Visualization us...
#>   7  26 sna        2.6      Carter T. Butts      1y Tools for Social Network...
#>   8  25 ggraph     2.0.5    Thomas Lin Pedersen  1y An Implementation of Gra...
#>   9  23 snow       0.4.4    Luke Tierney         4M Simple Network of Workst...
#>  10  18 neuralnet  1.44.2   Marvin N. Wright     3y Training of Neural Networks
more()
#> - "networks" ---------------------------------- 771 packages in 0.006 seconds -
#>   #    package         version by                    @ title                   
#>  11 16 bnlearn         4.7     Marco Scutari        5M Bayesian Network Stru...
#>  12 15 spatstat.linnet 2.3.1   Adrian Baddeley      2M Linear Networks Funct...
#>  13 15 ergm            4.1.2   Pavel N. Krivitsky   7M Fit, Simulate and Dia...
#>  14 15 networkD3       0.4     Christopher Gandrud  5y D3 JavaScript Network...
#>  15 12 intergraph      2.0.2   Michal Bojanowski    7y Coercion Routines for...
#>  16 12 diagram         1.6.5   Karline Soetaert     1y Functions for Visuali...
#>  17 12 networkDynamic  0.11.0  Skye Bender-deMoll   8M Dynamic Extensions fo...
#>  18 12 WGCNA           1.70.3  Peter Langfelder     1y Weighted Correlation ...
#>  19 11 RSNNS           0.4.14  Christoph Bergmeir   6M Neural Networks using...
#>  20 11 torch           0.6.0   Daniel Falbel        4M Tensors and Neural Ne...

# Details
ps("visualization")
#> - "visualization" ---------------------------- 1371 packages in 0.005 seconds -
#>   #     package    version by                    @ title                       
#>   1 100 ggplot2    3.3.5   Thomas Lin Pedersen  8M Create Elegant Data Visua...
#>   2  62 igraph     1.2.11  Tamás Nepusz         1M Network Analysis and Visu...
#>   3  58 scales     1.1.1   Hadley Wickham       2y Scale Functions for Visua...
#>   4  37 rgl        0.108.3 Duncan Murdoch       3M 3D Visualization Using Op...
#>   5  26 vdiffr     1.0.2   Lionel Henry         8M Visual Regression Testing...
#>   6  19 DiagrammeR 1.0.8   Richard Iannone     20d Graph/Network Visualization 
#>   7  19 ROCR       1.0.11  Felix G.M. Ernst     2y Visualizing the Performan...
#>   8  18 corrplot   0.92    Taiyun Wei           3M Visualization of a Correl...
#>   9  16 circlize   0.4.14  Zuguang Gu           2d Circular Visualization      
#>  10  16 visNetwork 2.1.0   Benoit Thieurmel     5M Network Visualization usi...
ps()
#> - "visualization" ---------------------------- 1371 packages in 0.005 seconds -
#> 
#> 1 ggplot2 @ 3.3.5                             Thomas Lin Pedersen, 8 months ago
#> -----------------
#>   # Create Elegant Data Visualisations Using the Grammar of Graphics
#>   A system for 'declaratively' creating graphics, based on "The Grammar
#>   of Graphics". You provide the data, tell 'ggplot2' how to map
#>   variables to aesthetics, what graphical primitives to use, and it
#>   takes care of the details.
#>   https://ggplot2.tidyverse.org
#>   https://github.com/tidyverse/ggplot2
#> 
#> 2 igraph @ 1.2.11                               Tamás Nepusz, about a month ago
#> -----------------
#>   # Network Analysis and Visualization
#>   Routines for simple graphs and network analysis. It can handle large
#>   graphs very well and provides functions for generating random and
#>   regular graphs, graph visualization, centrality methods and much
#>   more.
#>   https://igraph.org
#> 
#> 3 scales @ 1.1.1                                    Hadley Wickham, 2 years ago
#> ----------------
#>   # Scale Functions for Visualization
#>   Graphical scales map data to aesthetics, and provide methods for
#>   automatically determining breaks and labels for axes and legends.
#>   https://scales.r-lib.org
#>   https://github.com/r-lib/scales
#> 
#> 4 rgl @ 0.108.3                                    Duncan Murdoch, 3 months ago
#> ---------------
#>   # 3D Visualization Using OpenGL
#>   Provides medium to high level functions for 3D interactive graphics,
#>   including functions modelled on base graphics (plot3d(), etc.) as
#>   well as functions for constructing representations of geometric
#>   objects (cube3d(), etc.).  Output may be on screen using OpenGL, or
#>   to various standard 3D file formats including WebGL, PLY, OBJ, STL as
#>   well as 2D image formats, including PNG, Postscript, SVG, PGF.
#>   https://github.com/dmurdoch/rgl
#>   https://dmurdoch.github.io/rgl/
#> 
#> 5 vdiffr @ 1.0.2                                     Lionel Henry, 8 months ago
#> ----------------
#>   # Visual Regression Testing and Graphical Diffing
#>   An extension to the 'testthat' package that makes it easy to add
#>   graphical unit tests. It provides a Shiny application to manage the
#>   test cases.
#>   https://vdiffr.r-lib.org/
#>   https://github.com/r-lib/vdiffr
#> 
#> 6 DiagrammeR @ 1.0.8                               Richard Iannone, 20 days ago
#> --------------------
#>   # Graph/Network Visualization
#>   Build graph/network structures using functions for stepwise addition
#>   and deletion of nodes and edges. Work with data available in tables
#>   for bulk addition of nodes, edges, and associated metadata. Use graph
#>   selections and traversals to apply changes to specific nodes or
#>   edges. A wide selection of graph algorithms allow for the analysis of
#>   graphs. Visualize the graphs and take advantage of any aesthetic
#>   properties assigned to nodes and edges.
#>   https://github.com/rich-iannone/DiagrammeR
#> 
#> 7 ROCR @ 1.0.11                                   Felix G.M. Ernst, 2 years ago
#> ---------------
#>   # Visualizing the Performance of Scoring Classifiers
#>   ROC graphs, sensitivity/specificity curves, lift charts, and
#>   precision/recall plots are popular examples of trade-off
#>   visualizations for specific pairs of performance measures. ROCR is a
#>   flexible tool for creating cutoff-parameterized 2D performance curves
#>   by freely combining two from over 25 performance measures (new
#>   performance measures can be added using a standard interface). Curves
#>   from different cross-validation or bootstrapping runs can be averaged
#>   by different methods, and standard deviations, standard errors or box
#>   plots can be used to visualize the variability across the runs. The
#>   parameterization can be visualized by printing cutoff values at the
#>   corresponding curve positions, or by coloring the curve according to
#>   cutoff. All components of a performance plot can be quickly adjusted
#>   using a flexible parameter dispatching mechanism. Despite its
#>   flexibility, ROCR is easy to use, with only three commands and
#>   reasonable default values for all optional parameters.
#>   http://ipa-tys.github.io/ROCR/
#> 
#> 8 corrplot @ 0.92                                      Taiyun Wei, 3 months ago
#> -----------------
#>   # Visualization of a Correlation Matrix
#>   Provides a visual exploratory tool on correlation matrix that
#>   supports automatic variable reordering to help detect hidden patterns
#>   among variables.
#>   https://github.com/taiyun/corrplot
#> 
#> 9 circlize @ 0.4.14                                      Zuguang Gu, 2 days ago
#> -------------------
#>   # Circular Visualization
#>   Circular layout is an efficient way for the visualization of huge
#>   amounts of information. Here this package provides an implementation
#>   of circular layout generation in R as well as an enhancement of
#>   available software. The flexibility of the package is based on the
#>   usage of low-level graphics functions such that self-defined
#>   high-level graphics can be easily implemented by users for specific
#>   purposes. Together with the seamless connection between the powerful
#>   computational and visual environment in R, it gives users more
#>   convenience and freedom to design figures for better understanding
#>   complex patterns behind multiple dimensional data. The package is
#>   described in Gu et al. 2014 <doi:10.1093/bioinformatics/btu393>.
#>   https://github.com/jokergoo/circlize
#>   https://jokergoo.github.io/circlize_book/book/
#> 
#> 10 visNetwork @ 2.1.0                            Benoit Thieurmel, 5 months ago
#> ---------------------
#>   # Network Visualization using 'vis.js' Library
#>   Provides an R interface to the 'vis.js' JavaScript charting library.
#>   It allows an interactive visualization of networks.
#>   http://datastorm-open.github.io/visNetwork/

# See the underlying data frame
ps("ropensci")
#> - "ropensci" ---------------------------------- 235 packages in 0.006 seconds -
#>   #     package         version by                   @ title                   
#>   1 100 rtweet          0.7.0   Michael W. Kearney  2y Collecting Twitter Data 
#>   2  77 vcr             1.0.2   Scott Chamberlain   9M Record 'HTTP' Calls t...
#>   3  59 RSelenium       1.7.7   Ju Yeong Kim        2y R Bindings for 'Selen...
#>   4  59 webmockr        0.8.0   Scott Chamberlain   1y Stubbing and Setting ...
#>   5  39 visdat          0.5.3   Nicholas Tierney    3y Preliminary Visualisa...
#>   6  39 wdman           0.2.5   Ju Yeong Kim        2y 'Webdriver'/'Selenium...
#>   7  37 rfisheries      0.2     Karthik Ram         6y 'Programmatic Interfa...
#>   8  36 taxize          0.9.99  Scott Chamberlain   1y Taxonomic Information...
#>   9  36 googleLanguageR 0.3.0   Mark Edmondson      2y Call Google's 'Natura...
#>  10  35 chromer         0.1     Matthew Pennell     7y Interface to Chromoso...
ps()[]
#> # A data frame: 10 × 14
#>    score package   version title description date                maintainer_name
#>    <dbl> <chr>     <pckg_> <chr> <chr>       <dttm>              <chr>          
#>  1  920. rtweet    0.7.0   "Col… "An implem… 2020-01-08 22:00:10 Michael W. Kea…
#>  2  711. vcr       1.0.2   "Rec… "Record te… 2021-05-31 17:00:02 Scott Chamberl…
#>  3  545. RSelenium 1.7.7   "R B… "Provides … 2020-02-03 19:00:03 Ju Yeong Kim   
#>  4  541. webmockr  0.8.0   "Stu… "Stubbing … 2021-03-14 14:50:02 Scott Chamberl…
#>  5  359. visdat    0.5.3   "Pre… "Create pr… 2019-02-15 14:30:03 Nicholas Tiern…
#>  6  355. wdman     0.2.5   "'We… "There are… 2020-01-31 21:30:02 Ju Yeong Kim   
#>  7  344. rfisheri… 0.2     "'Pr… "A program… 2016-02-19 08:50:03 Karthik Ram    
#>  8  332. taxize    0.9.99  "Tax… "Interacts… 2020-10-30 04:50:02 Scott Chamberl…
#>  9  331. googleLa… 0.3.0   "Cal… "Call 'Goo… 2020-04-19 12:40:02 Mark Edmondson 
#> 10  324. chromer   0.1     "Int… "A program… 2015-01-13 10:27:09 Matthew Pennell
#> # … with 7 more variables: maintainer_email <chr>, revdeps <int>,
#> #   downloads_last_month <int>, license <chr>, url <chr>, bugreports <chr>,
#> #   package_data <I<list>>